% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{Data of traffic accidents in the USA from 2013 to 2015}
\format{
A data frame with columns, cmpressin in csv.bz2 form
}
\source{
US National Highway Traffic Safety Administration's Fatality
Analysis Reporting System, which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered in
motor vehicle traffic crashes.
\url{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
accident_2013
}
\description{
Data of traffic accidents in the USA from 2013 to 2015
}
\examples{
\dontrun{
 accidents
}
}
\keyword{datasets}
