% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\docType{data}
\name{accident_2015}
\alias{accident_2015}
\title{This is data to be included in my package accident 2015}
\description{
This is data to be included in my package accident 2015
}
\author{
Jesus Obregon Dominguez, \email{gerencia@dataengineeringperu.com}
}
\keyword{data}
