% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_identifications.R
\name{apply_identifications}
\alias{apply_identifications}
\title{Apply Identification Constraints to Factors and Loadings}
\usage{
apply_identifications(factors, loadings)
}
\arguments{
\item{factors}{A numeric matrix with the current estimate of all factors (T × total factors).}

\item{loadings}{A numeric matrix with the current estimate of all factor loadings (total factors x N).}
}
\value{
A list containing:
\describe{
\item{factors}{A matrix containing identified factors.}
\item{loadings}{A matrix of identified factor loadings.}
}
}
\description{
Applies identification constraints to factors and loadings.
}
\keyword{internal}
