% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_factors_hat.R
\name{compute_factors_hat}
\alias{compute_factors_hat}
\title{Compute Factors_hat from Observed Data and Loadings}
\usage{
compute_factors_hat(Yorig, ranges, Final_list, Loadings_list)
}
\description{
Projects observed data onto the estimated loading space to reconstruct \code{Factors_hat}
for each combination of blocks in the multilevel structure.
}
\keyword{internal}
