% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fpr_gamma.R
\name{compute_fpr_gamma}
\alias{compute_fpr_gamma}
\title{Compute Adaptive Threshold Cross-Sectional Robust Gamma (FPR Gamma)}
\usage{
compute_fpr_gamma(residuals, loadings)
}
\arguments{
\item{residuals}{A \code{T x N} matrix of residualsl.}

\item{loadings}{A \code{N x K} matrix of factor loadings.}
}
\value{
A \code{K x K} matrix representing the FPR-adjusted covariance of the latent factors.
}
\description{
Computes the cross-sectional robust covariance estimator (gamma) based on the Adaptive Thresholding method proposed by Fresoli, Poncela, and Ruiz (2024).
}
\keyword{internal}
