% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gamma_atcsr.R
\name{compute_gamma_atcsr}
\alias{compute_gamma_atcsr}
\title{Compute Adaptive Threshold Cross-Sectional Robust Gamma (AT-CSR)}
\usage{
compute_gamma_atcsr(Residuals, Loadings)
}
\description{
Compute Adaptive Threshold Cross-Sectional Robust Gamma (AT-CSR)
}
\keyword{internal}
