% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gamma_FPR.R
\name{compute_gamma_FPR}
\alias{compute_gamma_FPR}
\title{Compute Adaptive Threshold Cross-Sectional Robust Gamma (AT-CSR)
Fresoli, Poncela and Ruiz (2024)}
\usage{
compute_gamma_FPR(Residuals, Loadings)
}
\description{
Compute Adaptive Threshold Cross-Sectional Robust Gamma (AT-CSR)
Fresoli, Poncela and Ruiz (2024)
}
\keyword{internal}
