% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lambda.R
\name{compute_lambda}
\alias{compute_lambda}
\title{Compute Loadings (Lambda Matrix)}
\usage{
compute_lambda(Yorig, num_blocks, ranges, r_list, Factor_list, Loadings_list)
}
\description{
Computes the full loadings matrix (Lambda) for the multilevel dynamic factor model,
including global and lower-level block-specific loadings, by regressing observed data
on previously extracted factors.
}
\keyword{internal}
