% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_optimal_delta.R
\name{compute_optimal_delta}
\alias{compute_optimal_delta}
\title{Compute Optimal Delta for FPR Gamma Computation}
\usage{
compute_optimal_delta(Sigma_eps, Theta, T)
}
\value{
A single numeric value representing the optimal threshold level \code{delta}.
}
\description{
Computes the optimal threshold level delta.
}
\keyword{internal}
