% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stressed_factors.R
\name{compute_stressed_factors}
\alias{compute_stressed_factors}
\title{Compute Stressed Factors}
\usage{
compute_stressed_factors(dep_variable, factors, scenario, h, QTAU, min)
}
\description{
Compute Stressed Factors
}
\keyword{internal}
