% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data2013}
\alias{data2013}
\title{Number of Accidents for the Year 2013}
\format{A data frame with columns:
\describe{
\item{Year}{2013}
\item{Region}{The States of the USA}
}}
\source{
Coursera
}
\usage{
data2013
}
\description{
Number of Accidents for the Year 2013
}
\keyword{datasets}
