% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_accident_2013}
\alias{demo_accident_2013}
\title{Postal Abbreviations for Mexico}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30202 rows and 50 columns.}
\usage{
demo_accident_2013
}
\description{
Postal Abbreviations for Mexico
}
\examples{
\dontrun{
 demo_accident_2013
}
}
\keyword{datasets}

