% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dep_variable}
\alias{dep_variable}
\title{US GDP Growth Series}
\format{
A time series object with 59 quarterly observations.
}
\source{
Replication materials of González-Rivera et al. (2024).
}
\usage{
data(dep_variable)
}
\description{
Quarterly US GDP growth series used as the dependent variable in the replication.
}
\details{
The original series contains quarterly GDP levels for 63 countries.
For replication, all series are converted to log-differenced annualized growth rates
(\code{diff(log(x)) * 400}). From this dataset, the U.S. series is extracted
and the first observation dropped to obtain 59 observations in total.
}
\keyword{datasets}
