% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_sorted.R
\name{eigen_sorted}
\alias{eigen_sorted}
\title{Eigen Decomposition (Descending Order)}
\usage{
eigen_sorted(matrix_y)
}
\description{
Computes the eigenvalues and eigenvectors of a symmetric matrix, returning them in descending order (default behavior of \code{eigen()}).
}
\keyword{internal}
