% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldfm_methods.R
\name{factors.mldfm}
\alias{factors.mldfm}
\title{Extract Estimated Factors from a \code{mldfm} Object}
\usage{
\method{factors}{mldfm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{...}{Further arguments (ignored).}
}
\value{
A matrix containing the estimated factors.
}
\description{
Extract Estimated Factors from a \code{mldfm} Object
}
\examples{
mldfm_result <- mldfm(data = matrix(rnorm(100 * 5), 100, 5), blocks = 1, global = 2)
factors(mldfm_result)

}
