\name{faRs-package}
\alias{faRs-package}
\alias{faRs}
\docType{package}
\title{
\packageTitle{faRs}
}
\description{
\packageDescription{faRs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{faRs}
\packageIndices{faRs}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{faRs}

Maintainer: \packageMaintainer{faRs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
