% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_2013-data.R
\docType{data}
\name{fars_2013}
\alias{fars_2013}
\title{Fatal injury data}
\format{A dataframe with columns:
\describe{
\item{STATE}{A number bewteen 1 and 50}
\item{ST_CASE}{A number indicating cases reported}
}}
\source{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\usage{
data(fars_2013)
}
\description{
Fatal injury data
}
\examples{
\dontrun{
data(fars_2013)
}
}
\keyword{datasets}
