% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_methods.R
\name{fitted.fars}
\alias{fitted.fars}
\title{Fitted Values for \code{fars} Object}
\usage{
\method{fitted}{fars}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fars}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric matrix with one column per quantile level and as many rows as observations used in the fit.
}
\description{
Returns a matrix of fitted values from all quantile regressions stored in a \code{fars} object.
}
\examples{
fars_result <- compute_fars(dep_variable = rnorm(100), 
                            factors = matrix(rnorm(100 * 3), ncol = 3))
fitted(fars_result)

}
