% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldfm_methods.R
\name{fitted.mldfm}
\alias{fitted.mldfm}
\title{Extract Fitted Values from a \code{mldfm} Object}
\usage{
\method{fitted}{mldfm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mldfm}.}

\item{...}{Further arguments (ignored).}
}
\value{
A matrix containing the fitted values.
}
\description{
Extract Fitted Values from a \code{mldfm} Object
}
\examples{
mldfm_result <- mldfm(data = matrix(rnorm(100 * 5), 100, 5), blocks = 1, global = 2)
fitted(mldfm_result)

}
