% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.mldfm.R
\name{generate_factor_plot}
\alias{generate_factor_plot}
\title{Generate Plot for a Single Factor}
\usage{
generate_factor_plot(factor_name, data_plot, ymin, ymax)
}
\description{
Generate Plot for a Single Factor
}
\keyword{internal}
