% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.mldfm.R
\name{generate_loadings_plot}
\alias{generate_loadings_plot}
\title{Generate Loadings for a Single Factor}
\usage{
generate_loadings_plot(factor_name, data_plot, y_min, y_max)
}
\description{
Generate Loadings for a Single Factor
}
\keyword{internal}
