% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution.R
\name{get_distribution.fars_density}
\alias{get_distribution.fars_density}
\title{Extract Distribution from a \code{fars_density} Object}
\usage{
\method{get_distribution}{fars_density}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fars_density}.}

\item{...}{Further arguments (ignored).}
}
\value{
A matrix containing the random draws from the fitted skew-t distribution if available, otherwise NULL.
}
\description{
Extracts the distribution from a \code{fars_density} object.
}
\examples{
\donttest{
fars_density_result <- compute_density(quantiles = matrix(rnorm(100 * 5), nrow = 100, ncol = 5))
get_distribution(fars_density_result)
}
}
