% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution.R
\name{get_distribution}
\alias{get_distribution}
\title{Generic Function to Extract Distribution}
\usage{
get_distribution(object, ...)
}
\arguments{
\item{object}{An object from which to extract the distribution.}

\item{...}{Additional arguments (ignored).}
}
\value{
A matrix containing the random draws from the fitted skew-t distribution.
}
\description{
Generic Function to Extract Distribution
}
\examples{
\donttest{
fars_density_result <- compute_density(
  quantiles = matrix(rnorm(100 * 5), nrow = 100, ncol = 5)
)
get_distribution(fars_density_result)
}

}
