% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_methods.R
\name{get_ellipsoids.fars_scenario}
\alias{get_ellipsoids.fars_scenario}
\title{Get Ellipsoids from a \code{fars_scenario} Object.}
\usage{
\method{get_ellipsoids}{fars_scenario}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fars_scenario}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list of matrices defining the ellipsoids at each time.
}
\description{
Returns the list of ellipsoids from a \code{fars_scenario} object.
}
