% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ellipsoids.R
\name{get_ellipsoids}
\alias{get_ellipsoids}
\title{Get Ellipsoids from \code{fars_scenario}}
\usage{
get_ellipsoids(x)
}
\arguments{
\item{x}{An object of class \code{fars_scenario}.}
}
\value{
A list of matrices defining the ellipsoids at each time.
}
\description{
Returns the list of ellipsoids from a \code{fars_scenario} object.
}
