% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factors.R
\name{get_factors.mldfm}
\alias{get_factors.mldfm}
\title{Extract Estimated Factors from a \code{mldfm} Object}
\usage{
\method{get_factors}{mldfm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{...}{Further arguments (ignored).}
}
\value{
A matrix containing the estimated factors.
}
\description{
Extract Estimated Factors from a \code{mldfm} Object
}
\examples{
mldfm_result <- mldfm(data = matrix(rnorm(100 * 5), 100, 5), blocks = 1, global = 2)
get_factors(mldfm_result)

}
