% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factors.R
\name{get_factors}
\alias{get_factors}
\title{Extract Factors at a Given Hierarchical Level}
\usage{
get_factors(Factor_list, combination, level)
}
\description{
Extract Factors at a Given Hierarchical Level
}
\keyword{internal}
