% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loadings.R
\name{get_loadings}
\alias{get_loadings}
\title{Generic Function to Extract Factor Loadings}
\usage{
get_loadings(x, ...)
}
\arguments{
\item{x}{An object from which to extract the factor loadings.}

\item{...}{Additional arguments.}
}
\value{
A matrix of factor loadings.
}
\description{
Generic Function to Extract Factor Loadings
}
