% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mldfm_list.R
\name{get_mldfm_list.mldfm_subsample}
\alias{get_mldfm_list.mldfm_subsample}
\title{Extract List of MLDFMs from a \code{mldfm_subsample} Object}
\usage{
\method{get_mldfm_list}{mldfm_subsample}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm_subsample}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list of \code{mldfm} objects.
}
\description{
Returns the list of all \code{mldfm} stored in a \code{mldfm_subsample} object.
}
