% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldfm_subsample_methods.R
\name{get_mldfm_model.mldfm_subsample}
\alias{get_mldfm_model.mldfm_subsample}
\title{Extract a Specific \code{mldfm} Object from a \code{mldfm_subsample} Object}
\usage{
\method{get_mldfm_model}{mldfm_subsample}(x, index, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm_subsample}.}

\item{index}{Integer. The position of the desired model (between 1 and \code{n_samples}).}

\item{...}{Additional arguments (ignored).}
}
\value{
A single \code{mldfm} object.
}
\description{
Returns the \code{mldfm} object at the specified position in a \code{mldfm_subsample} object.
}
