% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mldfm_model.R
\name{get_mldfm_model}
\alias{get_mldfm_model}
\title{Generic Function to Extract a Specific \code{mldfm} Object}
\usage{
get_mldfm_model(x, index, ...)
}
\arguments{
\item{x}{An object from which to extract a \code{mldfm} Object}

\item{index}{Integer. The position of the desired model}

\item{...}{Additional arguments.}
}
\value{
A single \code{mldfm} object.
}
\description{
Generic Function to Extract a Specific \code{mldfm} Object
}
