% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_methods.R
\name{get_quantile_levels}
\alias{get_quantile_levels}
\title{Generic Function to Extract Quantile Levels}
\usage{
get_quantile_levels(x, ...)
}
\arguments{
\item{x}{An object from which to extract the quantile levels}

\item{...}{Additional arguments.}
}
\value{
A vector of quantile levels.
}
\description{
Generic Function to Extract Quantile Levels
}
