% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Generic Function to Extract Quantiles}
\usage{
get_quantiles(x, ...)
}
\arguments{
\item{x}{An object from which to extract the quantiles.}

\item{...}{Additional arguments (ignored).}
}
\value{
A matrix of quantiles.
}
\description{
Generic Function to Extract Quantiles
}
