% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals.mldfm}
\alias{get_residuals.mldfm}
\title{Extract Residuals from a \code{mldfm} Object}
\usage{
\method{get_residuals}{mldfm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{...}{Further arguments (ignored).}
}
\value{
A matrix containing the residuals.
}
\description{
Extract Residuals from a \code{mldfm} Object
}
\examples{
mldfm_result <- mldfm(data = matrix(rnorm(100 * 5), 100, 5), blocks = 1, global = 2)
get_residuals(mldfm_result)

}
