% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Generic function to extract residuals}
\usage{
get_residuals(object, ...)
}
\arguments{
\item{object}{An object from which to extract the residuals.}

\item{...}{Additional arguments.}
}
\value{
A matrix of residuals.
}
\description{
Generic function to extract residuals
}
