% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_methods.R
\name{get_rq_model}
\alias{get_rq_model}
\title{Generic Function to Extract a Specific \code{rq} Object}
\usage{
get_rq_model(x, tau, ...)
}
\arguments{
\item{x}{An object from which to extract a \code{rq} Object}

\item{tau}{Numeric scalar in (0, 1) indicating the desired quantile.}

\item{...}{Additional arguments.}
}
\value{
A single \code{rq} object.
}
\description{
Generic Function to Extract a Specific \code{rq} Object
}
