% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stressed_factors.R
\name{get_stressed_factors}
\alias{get_stressed_factors}
\title{Generic Function to Extract Stressed Factors}
\usage{
get_stressed_factors(x, ...)
}
\arguments{
\item{x}{An object from which to extract the stressed factors.}

\item{...}{Additional arguments (ignored).}
}
\value{
A matrix of stressed factors.
}
\description{
Generic Function to Extract Stressed Factors
}
