% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldfm_methods.R
\name{loadings.mldfm}
\alias{loadings.mldfm}
\title{Extract Factor Loadings from a \code{mldfm} Object}
\usage{
\method{loadings}{mldfm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{...}{Further arguments (ignored).}
}
\value{
A matrix containing the estimated factor loadings.
}
\description{
Extract Factor Loadings from a \code{mldfm} Object
}
\examples{
mldfm_result <- mldfm(data = matrix(rnorm(100 * 5), 100, 5), blocks = 1, global = 2)
loadings(mldfm_result)

}
