% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mf_data}
\alias{mf_data}
\title{Macro-Financial Database}
\format{
A numeric matrix with 59 rows and 519 columns.
}
\source{
Replication materials of González-Rivera et al. (2024).
}
\usage{
data(mf_data)
}
\description{
Macro-financial variables used in the replication exercise.
}
\details{
The original dataset contains 624 variables. For replication, the first 519 variables
are selected, converted to a numeric matrix, and outliers are corrected using
the function \code{correct_outliers(..., threshold = 5)} provided in FARS.
}
\keyword{datasets}
