% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_density.R
\name{nl_density}
\alias{nl_density}
\title{Compute Skew-t Densities from Forecasted Quantiles (Nonlinear Optimization)}
\usage{
nl_density(
  quantiles,
  levels = c(0.05, 0.25, 0.5, 0.75, 0.95),
  est_points = 512,
  random_samples = 5000,
  support = c(-10, 10),
  seed = NULL
)
}
\description{
Compute Skew-t Densities from Forecasted Quantiles (Nonlinear Optimization)
}
\keyword{internal}
