% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fars_scenario}
\alias{plot.fars_scenario}
\title{Plot Method for \code{fars_scenario} Object}
\usage{
\method{plot}{fars_scenario}(x, obs = 1, ...)
}
\arguments{
\item{x}{An object of class \code{fars_scenario}.}

\item{obs}{Integer. Time index to plot (default = 1).}

\item{...}{Additional arguments (ignored).}
}
\description{
Plots the hyperellipsoid for a given time observation (only for 1D or 2D cases).
}
