% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mldfm_subsample}
\alias{plot.mldfm_subsample}
\title{Plot Method for \code{mldfm_subsample} Object}
\usage{
\method{plot}{mldfm_subsample}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm_subsample}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A ggplot object (invisibly).
}
\description{
Plots a histogram of the number of iterations used in each subsample estimation.
}
