% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors.mldfm}
\alias{plot_factors.mldfm}
\title{Plot Extracted Factors from MLDFM}
\usage{
plot_factors.mldfm(x, dates = NULL)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{dates}{Optional vector of dates to label the x-axis.}
}
\value{
A list of ggplot objects, one for each factor.
}
\description{
Plot Extracted Factors from MLDFM
}
