% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fars.R
\name{print.fars}
\alias{print.fars}
\title{Print Method for \code{fars} Object}
\usage{
\method{print}{fars}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fars}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{object}, returned invisibly.
}
\description{
Prints a short summary of the fars object.
}
