% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fars_density.R
\name{print.fars_density}
\alias{print.fars_density}
\title{Print method for fars_density objects}
\usage{
\method{print}{fars_density}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fars_density}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Displays a brief summary of the density estimation object produced by the `density()` or `nl_density()` function.
}
