% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fars_scenario}
\alias{print.fars_scenario}
\title{Print Method for \code{fars_scenario} Object}
\usage{
\method{print}{fars_scenario}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fars_scenario}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{fars_scenario} object, invisibly.
}
\description{
Prints a short summary of the FARS scenario object.
}
