% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mldfm}
\alias{print.mldfm}
\title{Print Method for \code{mldfm} Object}
\usage{
\method{print}{mldfm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{mldfm} object, invisibly.
}
\description{
Prints a short summary of the MLDFM object.
}
