% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mldfm_subsample}
\alias{print.mldfm_subsample}
\title{Print Method for \code{mldfm_subsample} Object}
\usage{
\method{print}{mldfm_subsample}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm_subsample}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{object}, invisibly.
}
\description{
Prints a brief summary of the \code{mldfm_subsample} object.
}
