% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_methods.R
\name{residuals.fars}
\alias{residuals.fars}
\title{Residuals for \code{fars} Object}
\usage{
\method{residuals}{fars}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fars}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric matrix with one column per quantile level and as many rows as observations used in the fit.
}
\description{
Returns a matrix of residuals from all quantile regressions stored in a \code{fars} object.
}
\examples{
fars_result <- compute_fars(dep_variable = rnorm(100), 
                            factors = matrix(rnorm(100 * 3), ncol = 3))
residuals(fars_result)

}
