% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldfm_methods.R
\name{residuals.mldfm}
\alias{residuals.mldfm}
\title{Extract Residuals from a \code{mldfm} Object}
\usage{
\method{residuals}{mldfm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mldfm}.}

\item{...}{Further arguments (ignored).}
}
\value{
A matrix containing the residuals.
}
\description{
Extract Residuals from a \code{mldfm} Object
}
\examples{
mldfm_result <- mldfm(data = matrix(rnorm(100 * 5), 100, 5), blocks = 1, global = 2)
residuals(mldfm_result)

}
