% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stressed_factors.R, R/q_reg.R
\name{shift}
\alias{shift}
\title{Shift a time series vector}
\usage{
shift(x, lag)

shift(x, lag)
}
\description{
Shift a time series vector

Shift a time series vector
}
\keyword{internal}
