% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_block.R
\name{single_block}
\alias{single_block}
\title{Multi-Level Dynamic Factor Model - Single block}
\usage{
single_block(data, r)
}
\description{
Multi-Level Dynamic Factor Model - Single block
}
\keyword{internal}
