% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fars}
\alias{summary.fars}
\title{Summary Method for \code{fars} Object}
\usage{
\method{summary}{fars}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fars}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{fars} object, returned invisibly.
}
\description{
Displays summaries of all quantile regressions stored in a \code{fars} object.
}
