% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fars_density}
\alias{summary.fars_density}
\title{Summary Method for \code{fars_density} Object}
\usage{
\method{summary}{fars_density}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fars_density}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{fars_density} object, invisibly.
}
\description{
Displays a complete summary of the \code{fars_density} object.
}
