% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fars_scenario}
\alias{summary.fars_scenario}
\title{Summary Method for \code{fars_scenario} Object}
\usage{
\method{summary}{fars_scenario}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fars_scenario}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{fars_scenario} object, invisibly.
}
\description{
Provides a detailed summary of the FARS scenario object.
}
