% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mldfm_subsample}
\alias{summary.mldfm_subsample}
\title{Summary Method for \code{mldfm_subsample} Object}
\usage{
\method{summary}{mldfm_subsample}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mldfm_subsample}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{object}, invisibly.
}
\description{
Provides a structured summary of a \code{mldfm_subsample} object.
}
