% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{Nationwide census providing data regarding fatal injuries suffered in motor vehicle traffic crashes
for the year 2013 in the USA.}
\format{
A data frame with columns:
\describe{
\item{STATE}{A valube between 1 and 56.}
\time{ST_CASE}{Identifier for each row.}
\item{VE_TOTAL}{Number of vehicles involved.}
\item{VE_FORMS}{Forms of vehicles involved.}
\item{PVH_INVL}{Number of PVH involved.}
\item{PEDS}{Number of peds.}
\item{PERMNOTVIT}{???}
\item{PERMVIT}{???}
\item{PERSONS}{Number of persons.}
\item{CITY}{County number.}
\item{CITY}{City identifier.}
\item{DAY}{A value between 1 and 31.}
\item{MONTH}{A value between 1 and 12.}
\item{YEAR}{2013}
\item{DAY_WEEK}{Day of the week}
\item{HOUR}{Hour.}
\item{MINUTE}{Minute.}
\item{NHS}{Either 0, 1 or 9.}
\item{and many others}{This dataframe has so many columns...}
}
}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\usage{
accident_2013
}
\description{
Nationwide census providing data regarding fatal injuries suffered in motor vehicle traffic crashes
for the year 2013 in the USA.
}
\examples{
\dontrun{
 accident_2013
}
}
\keyword{datasets}
