% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_identifications.R
\name{apply_identifications}
\alias{apply_identifications}
\title{Apply Identification Constraints to Factors and Loadings}
\usage{
apply_identifications(
  Yorig,
  num_blocks,
  ranges,
  num_factors,
  r,
  currentFactors,
  Factor_list,
  Loadings_list
)
}
\description{
Apply Identification Constraints to Factors and Loadings
}
\keyword{internal}
