% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_factor_structure.R
\name{build_factor_structure}
\alias{build_factor_structure}
\title{Create the key value pair system to track the factors}
\usage{
build_factor_structure(
  global = 1,
  local = NULL,
  middle_layer = NULL,
  num_blocks = NULL
)
}
\description{
Create the key value pair system to track the factors
}
\keyword{internal}
