% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_correlation_analysis.R
\name{canonical_correlation_analysis}
\alias{canonical_correlation_analysis}
\title{Canonical Correlation Analysis (for MLDFM)}
\usage{
canonical_correlation_analysis(y, Nregio, r_glob, r_reg)
}
\description{
Computes factors using CCA across blocks, based on local PCA-extracted factors.
}
\keyword{(final}
\keyword{Nregio:}
\keyword{all}
\keyword{be}
\keyword{containing}
\keyword{each}
\keyword{extracted}
\keyword{factor}
\keyword{factors}
\keyword{global}
\keyword{in}
\keyword{internal}
\keyword{matrix}
\keyword{node}
\keyword{number}
\keyword{of}
\keyword{output)}
\keyword{r_global:}
\keyword{r_reg:}
\keyword{regional}
\keyword{specify}
\keyword{the}
\keyword{to}
\keyword{variables}
\keyword{vector}
\keyword{which}
\keyword{y:}
