% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_initial_factors.R
\name{compute_initial_factors}
\alias{compute_initial_factors}
\title{Compute Initial Factors (Multilevel Dynamic Factor Model)}
\usage{
compute_initial_factors(
  Yorig,
  num_vars,
  num_obs,
  num_blocks,
  ranges,
  r_list,
  method
)
}
\description{
Compute Initial Factors (Multilevel Dynamic Factor Model)
}
\keyword{internal}
