% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_subsample.R
\name{compute_subsample}
\alias{compute_subsample}
\title{Internal helper to compute a column-wise subsample of the data by block}
\usage{
compute_subsample(data, block_ind, n, sample_size = 1, seed = NULL)
}
\description{
Internal helper to compute a column-wise subsample of the data by block
}
\keyword{internal}
