% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS_package.R
\docType{package}
\name{FARS-package}
\alias{FARS-package}
\alias{FARS}
\title{Analysis of accidents in the USA per year}
\description{
FARS is an R package for analysis of multiple accident-per-year data sets.

\tabular{ll}{ Package: \tab FARS\cr Type: \tab Package\cr
Version: \tab 0.1.0\cr Date: \tab 2017-11-06\cr License: \tab GPL-3\cr }
}
\author{
Dunja Kurtoic \email{dunja_kurtoic@psych.mpg.de}

Maintainer:Dunja Kurtoic \email{dunja_kurtoic@psych.mpg.de}
}
\keyword{package}
