% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars-package.R
\docType{package}
\name{fars-package}
\alias{fars}
\alias{fars-package}
\title{fars: Summarize and Map Traffic Fatality Data From FARS}
\description{
Read Fatality Analysis Reporting System (FARS) data by entering the year(s) corresponding to the files in the working directory. The 'fars' functions can summarize and map traffic fatalities for a year and state.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RonECox/fars}
  \item Report bugs at \url{https://github.com/RonECox/fars/issues}
}

}
\author{
\strong{Maintainer}: Ron Cox \email{RonECox.Social@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
