% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots longitude and latitude of state number given present in given years of accident}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character with state ID}

\item{year}{A vector with years}
}
\value{
This function returns plot with longitude and latitudes with accidents
}
\description{
This function reads all files with years, filters on state ID and plots longitude and latitudes.
It only plots longitude greater than 900 and latitude greater than 90
}
\details{
if there is no file with year prefixed after accident_ and suffixed with .csv.bz2, it will give error
}
\examples{
\dontrun{
fars_map_state("100","2021")
fars_map_state("100",c(2021,2022))
}

}
