% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS data for given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{This is the number of a state. Should be valid}

\item{year}{This is the year being considered in YYYY format}
}
\value{
This function returns NULL
}
\description{
This function reads in the FARs data and plots the given state with each accident
    represented by a black dot on the plot.
}
\examples{
fars_map_state(1, 2013)
fars_map_state(4, 2013)

}
\author{
John J Smith, \email{fakeemail123@duvman.ie}
}
\keyword{FARS}
