% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character or integer vector of the state number.}

\item{year}{A character or integer vector of the years a user wants to read in.}
}
\value{
Plot of accidents for a given year and state.
}
\description{
This function creates plots of all the accidents in a given state, year combination.
This function imports filter from the dplyr package, map from the maps package and points from the graphics package.
An error will be thrown if the state.num input is not in the STATE column of the dataset. It will alsow throw an error
if there are no accidents recorded in the data frame for the given year and state.num combination.
}
\examples{
fars_map_state_years(49, c(2000,2001,2002))

}
