% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state: Returns a map of the accidents in a state for a given year
from an inputted state number and an inputted year.
If state number does not exist it will return INVALID state number message.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number corresponding to the alphabetical order of US state}

\item{year}{A number, string or integer, for which FARS data is desired}
}
\description{
fars_map_state: Returns a map of the accidents in a state for a given year
from an inputted state number and an inputted year.
If state number does not exist it will return INVALID state number message.
}
\examples{
fars_map_state(1,2013)
fars_map_state(36,2013)
}
