% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw Maps with accidents and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer or character string that represent the state to be analyzed.}

\item{year}{An integer or character string that that represent the year.}
}
\value{
This function return a map where the dots represent the accidents
}
\description{
This is function for plotting a map of the count of accidents in a specific state in a specific year.
}
\details{
If state does not exist a error message apprear. Also there is not accidents in particular state
within specified year you will be notified.
}
\examples{
fars_map_state(1,2014)
}
