% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_w2.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS Map state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number.}

\item{year}{Year in which the accidents to be plotted occurred.}
}
\value{
Map with accidents plotted
}
\description{
Plot accidents occurred on a specific state for a given year
If there is no data for a state, the function will stop and throw an error.
}
\note{
This function depends on map and points functions from the maps and graphics package respectively.
}
\examples{
\dontrun{fars_map_state("1", "2013")}

}
