% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Detaild summarization of FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the state numer that should be considered. If the number of the state
would be wrong the arror would should up.}

\item{year}{a vector containing the list of the years}
}
\value{
Returns a graphical representation of the data.
}
\description{
The function makes a graphical representation of the FARS data on the
state map.
}
\note{
You must include tydyr and dplyr packages.
The data files should be named "accident_X.csv.bz2" and be in the same directory.
}
\examples{
\dontrun{fars_map_state(13 , c(2013) )}

}
