% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Print "Fars Map State"}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character string signifying the code for a state. For example,
Washington state equals '53'}

\item{year}{A character string representing 4 character length year value like '2015'.}
}
\value{
This function returns a map with accidents plotted for a given state and year.
Any records with invalid latitude and longitude coordinates will be excluded from the plot.
}
\description{
This function plots any accidents for a single state and year.
}
\note{
An invalid year and/or state code will result in a warning message the failed condition.
}
\examples{
 \dontrun{
fars_map_state('53', '2015')
fars_map_state(state.num = '53', year = '2015')
}

}
