% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots car fatalities in a specified US state and for a specific year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number corresponding to US state. Needs to be an integer.}

\item{year}{A numeric vector representing the years in a time period.}
}
\description{
It plots the car fatalities that happened in a specified US state and in a
specific year. The map corresponds to the shape of the state, and shows where the accidents happened, geographically.
}
\section{Warning}{

   If the \code{state.num} parameter is not an existing state number, or the
   \code{year} parameter does not exist, the function will show an error.
   If there are no fatalities in the given year and the given state,
   the function will stop and state that there are no fatalities to show.
}

\examples{
fars_map_state(12, 2015)
}
