% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draws the accidents on the US map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric or integer value of US state as defined in the FARS data.}

\item{year}{A numeric or integer value for the year of accident record.}
}
\value{
This function returns a plot with the accidents on US map. The accidents are selected
based on the input argument \code{state.num} and \code{year}.
}
\description{
This function creates a plot with all incidents in specific US state (defined by \code{state.num})
per indicaed \code{year}. The \code{year} should correspond to the recorded analysis from US
National Highway Traffic Safety Administration's Fatality Analysis Reporting System.
}
\note{
This function will give error if the file with indicated argument \code{year} is not
located in the working directory. It also gives error if the \code{state.num} is non-numeric
or non-integer value or does not correspond to any US state as defined in the FARS data.
}
\examples{
fars_map_state(1, 2013)

}

