% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Make a map of accident data per state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric value of length 1}

\item{year}{A four-digit number indicating the year in which the data collected.}
}
\value{
A list with \code{x} and \code{y} range and \code{names} components.
}
\description{
\code{fars_map_state} function creates a map of accident data corresponding to
the value of \code{year} argument. It then adds some points on the map based on
the coordinates provided by the dataset.
}
\details{
This function takes two numeric values as its arguments. It first calls
 \code{make_filename} and \code{fars_read} to create a character vector of file name
 corresponding to the argument \code{year} and then reads it into a tibble.
 If the date for that year does not exist in the working directory the function
 throws an error. It then coerced the numeric value for argument \code{state.num}
 into integer. After that it checks whether the required \code{state.num} exists
 in the column \code{STATE} in the dataset. If it does not exists the function
 throws an error mentioning that \code{STATE} number provided is invalid.
 Subsequently it calls to \code{\link[dplyr:filter]} to create a subset of the
 dataset based on \code{state.num} value. It then counts the number of rows of
 the subset and if it equals to zero a message appears on the console that no
 accidents to plot. The function turns the values which are greater than 900 on
 \code{LONGITUD} column of the dataset and also values which are greater than 90 on
 \code{LATITUDE} column into NAs. It then call to \code{\link[maps:map]} to make
 a map of subsetted dataset by means of \code{\link[base:with]}. After that the
 function uses \code{\link[graphics:points]} to add some points on the map.
}
\note{
Since the values of \code{STATE} in datasets are numeric, the function use
 \code{\link[base]{as.integer}} to coerce the value of the argument \code{state.num}
 into integer.
}
\examples{
 fars_map_state(17, 2015)
}
