% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Make a map of accidents in a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer reperesenting state of interest}
}
\value{
This function returns a graphics object that is a map of accidents within the
given state for the given year.  If there is no data for that year or the state number is invalid, the function
will return a NULL.
}
\description{
This function will produce a map of accidents within a state for a given year
}
\note{
Error returned if invalid state.num, NULL returned if no accidents occured in a given state/year
}
\examples{
fars_map_state(1,2014)

}
