% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map of a givne state with distribution of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of the state. If an invalid state number is chosen,
an error message ("invalid STATE number: ") is thrown.}

\item{year}{The year. If there are no accidents in a given year and state,
the message "no accidents to plot" is thrown.}
}
\value{
This function returns a map of the a given state, showing the distribution of accidents in a given year.

 @examples
\dontrun{fars_map_state(18, 2013)}
}
\description{
This function plots the geographical boundaries of a state and the distribution of reported accidents within the state.
State number and reporting years are required for the input.
}
