% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create of Map on the Number of incidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numerical value that represent the state selection in the function}

\item{year}{a numerical value that represent the year when the accidents happened}
}
\value{
The function will return a map graphic with points to show incidents at particular state and year
}
\description{
The function return a map with data points of the number of incidents happened over
             over a particular year and time. It he specified character(s) are plotted, centered
             at the coordinates and the map is limited to the state boundaries
}
\details{
The function will return an Error saying that the State is not found if the state.num parameter is not found within the dataset
         The function will return an Error saying that there is no accidents at that particular years and state if the length of Data is 0
}
\examples{
\dontrun {fars_map_state(12, 2008)}

}
