% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate a state map with FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number as used in the FARS dataset.}

\item{year}{Year for which to plot the accidents.}
}
\value{
Returns NULL. Prints a warning message if there are no accidents to plot (but not really since this can
    never happen). Generates an error if the state number is not valid.
}
\description{
Plots accidents for a given year on a map of the selected state
}
\examples{
\dontrun{
library(map)
fars_map_state(12, 2013)
fars_map_state(45, 2014)

# generates an error
fars_map_state(3, 2013)
}
}
