% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots observations of data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{\code{Integer} that represents the state}

\item{year}{integer}
}
\value{
Returns a plot of the state with plot accidents.  Returns error if state number is not valid.
Returns a message if no data is avaialable for that state.
}
\description{
User provides a state and the function will create a plot of observations of accidents
}
\examples{
\dontrun{
fars_map_state(50,2013)
}

}
