% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the accident for a specific year accross a state}
\usage{
fars_map_state(state.num, year,fars_path)
}
\arguments{
\item{state.num}{number identifying a statse in US character, integer or  numeric}

\item{year}{4 digit year character, integer or  numeric}

\item{fars_path}{charcter string correspondinf to the directory contianing the fars data}
}
\value{
a map object
}
\description{
Match a state number and year to the Fatality Analysis Reporting System dataset.
And produce a scatter plot of incident within a map af the state
}
\note{
\itemize{
\item both arguments should only include digits otherwise the program will stop with an error
\item if the state number is absent from the file an error is produced
}
}
\examples{
\dontrun{
# need to be in the same directory as the data
setwd("data")
my_map <- fars_map_state(1,2013)
}

}
\seealso{
make_filename

fars_read
}
