% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state - Create a map to display accidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state index as an integer}

\item{year}{A string, or an integer, with the input \code{year}}
}
\value{
This function shows a geographical map of the state with each point
     representing individual travel accidents for a given year
}
\description{
Displays a map of the reported accidents in the selected state for a given year
If the \code{state.num} is invalid the function shows an error
}
\examples{
\dontrun{
fars_map_state(state.num = 10, year = 2013)
fars_map_state(49, 2015)
fars_map_state("1",2014)
}

}
\seealso{
\link{fars_read}

\link{make_filename}
}
