% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Graphical map of fatalities in FARS data STATE and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer of state}

\item{year}{A character or integer of year}
}
\value{
This function returns NULL.
}
\description{
fars_map_state is used to create a graphical map of number of fatalities per
STATE and year.
}
\details{
The function draws a maps with dots displaying the number of fatalities per STATE and year
}
\note{
Error message: If an incorrect year or state is entered an error is thrown
}
\examples{
\dontrun{
fars_map_state(1,2013)
}

}

