% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function \code{fars_map_state}}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric the number of the state alphalbeticly}

\item{year}{the year of the death}
}
\value{
A map of a state, with each traffic death location shown on the map
}
\description{
This function displays the locations of traffic deaths
for a given state and year on a graphic map
}
\note{
Will display a warning message if no file for the given year not can be found

Will display a warning message a corresponding state can not be found
}
\examples{
fars_map_state(1, 2015)
fars_map_state(10, 2013)

}
