% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS map states.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that represents the state number.}

\item{year}{a vector of integer or strings that indicated the year of interest.}
}
\value{
This function draws a map of the State and shows one point for each fatality
  that occurred in that State. Will throw an error if an invalid
  state number is chosen or the chosen year's data does not exist.
}
\description{
Draws the location of accidents in a US State for a given year
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{filter}} in the \code{dplyr} package.
  \item \code{\link[maps]{map}} in the \code{maps} package.
  \item \code{\link[graphics]{points}} in the \code{graphics} package.
}
}

\examples{
\dontrun{
fars_map_state(3, 2014)
fars_map_state("3", "2014")
}

}
