% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS Map State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value for the STATE in the data frame}

\item{year}{A numeric value for the year in the filename}
}
\value{
This function returns a map with the accidents plotted on it
by longitude and lattitude.
}
\description{
This is a simple function that filters the data from a specific year
by a state. The filtered data is then plotted by to show the accident
data graphically by longitude and lattitude values.
}

