% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a data map for a given year and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric number}

\item{year}{a numeric number to indicate year -4 digt}
}
\value{
an error if the state.num given doesn't exist. a message if there is no oberservation in the year and state.num given. Otherwise, a map
}
\description{
This function read in a fars data file for a given year, and take a state.num. If state.num exist, it draws a map for a given year and the state
If state.num does not exist, throw a error. If there is no observation in the given year and state, throw a message
}
\examples{
\dontrun{fars_map_state(1,2015)}

}
