% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state - plot location of all accidents for a specific state in a specific year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The specific number whcih represend the state where the accidents occur.}

\item{year}{The specific year when the accidents occur, need to be able to converted to an integer}
}
\value{
a graph which shows the locations of all accidents happened in a specific state
   in a specific year. The locations are shown in dots on the map of the state.
}
\description{
This function takes a a specific state number and a specific year,
read the file of that year and plot the location of all accident happened
in that state in that year on a map.
}
\note{
Error when the state number is invalid or there's no accidents in that state in that year
}
\examples{
\dontrun{fars_map_state(23,2013)}
\dontrun{fars_map_state("10","2014")}

}
