% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Given a number representing a US State and a value representing a year, this function
extracts the FARS data for the given State at the given year, removes non-sensical
observations (where longitude > 900 or latitude > 90) and plots all the remaining
observations on a map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A value convertible to an integer, which represents the US State for
which to extract FARS data.}

\item{year}{A value convertible to an integer, which represents the year for which
to extract FARS data.}
}
\value{
No return value, but has the side effect of plotting the FARS observations for
        the given US State (\code{state.num}) at the given \code{year} on a map.
}
\description{
Given a number representing a US State and a value representing a year, this function
extracts the FARS data for the given State at the given year, removes non-sensical
observations (where longitude > 900 or latitude > 90) and plots all the remaining
observations on a map.
}
\note{
\code{state.num} and \code{year} must both be castable to integers.
}
\examples{
\dontrun{
fars_map_state(4, 2015)
}
}
