% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize Accidents on State Map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Valid State number as integer selecting which state to plot}

\item{year}{Year for which to plot the data as integer}
}
\value{
Returns plot displaying longitude and latitude of the accidents on the map
}
\description{
For a selected year and state the locations of all accidents are shown on the map of that state
}
\examples{
fars_map_state(2013,1)
}
