% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS Year Summarize}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character, numeric, or integer that gives the number
corresponding to the state for which data should be plotted.}

\item{year}{A character, numeric, or integer that gives the year for which
data should be plotted.}
}
\value{
a plot of the selected state corresponding to the inputted state.num,
 and points indicating the locations of the traffic fatalities of the
 selected year.
}
\description{
This function plots traffic fatalities for a specified year onto a specified
 state.
It uses functions from the 'maps', 'graphics', and 'dplyr' package, as well as
 the function 'fars_read' and 'make_filename' defined earlier.
}
\note{
Inputting years or states for which there is no data, or a year or
 state.num that are improperly formatted, will result in errors.
}
\examples{
\dontrun{fars_map_state(2,2013)}
\dontrun{fars_map_state("23","2014")}

}
