% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map of the events
This function will read in a state (by number index) and a year and return a map of the event locations.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number index in the fars data}

\item{year}{Four digit year}
}
\value{
a map of the event locations for the inputted state.
}
\description{
Create a map of the events
This function will read in a state (by number index) and a year and return a map of the event locations.
}
\examples{
\dontrun{
fars_map_state(1,2013) #Print events in Alabama for 2013
}
}
