% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots a map of the accidents for a US state of a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The integer code of a US state}

\item{year}{The integer value of a year. The \code{year} should be a positive decimal.
In order to use with the fars_data.zip dataset, the year should be in the range
2013..2015}
}
\value{
This function returns NULL
}
\description{
This function plots a map of a US state and displays the accidents in that state
for a year.
}
\examples{
\dontrun{
fars_map_state(state.num=1,year=2015)
}

}
