% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot of accidents by geographic location}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The desired state field which becomes an integer input.}

\item{year}{The desired year field which becomes an integer input.}
}
\value{
Returns a rendered state map based on a selected dataframe with the geographic data point of an accident.
}
\description{
This function provides a geographic summary of accident locations with a simple state map.
}
\details{
Uses the function make_filename and fars_read.  If no data is available, a message is issued.
}
\note{
This function uses the map and graphics functions.
}
\examples{
fars_map_state(13,2014)
}
\references{
For a list of state numbers, reference https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/812449
}
