% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function maps state number and returns data for the state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A state identification number}

\item{year}{A four digit valid year}
}
\value{
This function returns data for a state and year
}
\description{
This function maps state number and returns data for the state
}
\details{
This function raise errors when external library is not installed or invoked using library() function
}
\examples{
\dontrun{
fars_state_data <- fars_map_state(10, 2013)
}

}
