% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots map of number of fatalities}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a state number, when invalid an error is shown}

\item{year}{a number or character string containing numbers}
}
\value{
a tmap containing the number of casualties with months as rows and years as columns
}
\description{
From given statenumber and year, a map is plotted with the number of fatalities.
Longitude must be more than 900 and latitude more than 90.
When there are no fatalities, nothing is returned.
}
\examples{
fars_map_state(1,2013)

}
