% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot data of accidents in specific state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of state}

\item{year}{specific year}
}
\value{
plot with map or message
}
\description{
Takes state number and yeasr and make plot with map. If there are no state throw message "invalid STATE number: "
If there no accidents show message "no accidents to plot"
}
\examples{

\dontrun{
 fars_map_state(1, 2013)
 }

}
