% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate a map of fatalities for a given state and year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number corresponding with the desired state.}

\item{year}{A vector of years for which data is desired.}
}
\value{
A map of fatalities for a given state and year.
}
\description{
Generate a map of fatalities for a given state and year.
}
\section{Errors}{

\itemize{
 \item \code{state.num} is not a valid STATE number.
 \item \code{years} There are no fatalities in the given state and year.}
}

\examples{
#Basic usage:
fars_map_state(1, 2013)
}
