% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the accidence location in a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of the state}

\item{year}{the name of the year}
}
\value{
a map graph with the accidence location
}
\description{
Plot a the accident location in a specified state and year in the map.
if the LONGITUD > 900 or LATITUDE > 90, the location will be ignored.
}
\note{
if state.num is invalid, the program will exit with an error
}
\examples{
\dontrun{
fars_map_state(1,2015)
}

}

