% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{print("fars_map_state")
This function plots a map visualizing the data corresponding to the requested
state and year from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (FARS). The state is provided as a number
using the \code{state.num} argument. and the year is provided using the \code{year}
argument. If the state number does not exists, it returns an error message.
If there are no accidents for the selected state and year, it provides a
message.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number associated to the state for which data is to be
displayed}

\item{year}{the year for which the data is to be displayed}
}
\value{
This function plots a map displaying the FARS data corresponding to
   a particular year and state.
}
\description{
print("fars_map_state")
This function plots a map visualizing the data corresponding to the requested
state and year from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (FARS). The state is provided as a number
using the \code{state.num} argument. and the year is provided using the \code{year}
argument. If the state number does not exists, it returns an error message.
If there are no accidents for the selected state and year, it provides a
message.
}
\examples{
fars_map_state(12,2015)
fars_map_state(state.num = 12,year = 2015)

}
