% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw the Map of a State and Plot the Location of the Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number designating a state}

\item{year}{a numeric value indicating the year}
}
\value{
the function does not return anything. It plots a map if the input
  are in the expected range.
}
\description{
Select a state and a year and draw the map of the states and overlay the
locations of accidents
}
\note{
If the state number used as input does not exist in the data set, the
  function will throw an error indicating the selection was not valid. If for
  the chosen combination of state and year there are no accidents the
  function will display a message informing there are no accidents to plot.
}
\examples{
\dontrun{
fars_map_state(state.num = 1, year = 2013)
}
}
