% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generating a map of accidents for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numerical or character variable indicating the numerical
code of a given state in the U.S.. The coding used is available at
\url{https://crashstats.nhtsa.dot.gov/#/DocumentTypeList/6}.}

\item{year}{A numerical or character variable indicating the year whose data
one wants to plot.}
}
\value{
The function (invisibly) returns \code{NULL} and prints a map with
the locations of accidents in year \code{year} and state \code{state.num}
indicated by points drawn on a map of said state. An error is thrown if
either of the arguments is invalid. If there are no accidents reported for
state \code{state.num} in year \code{year}, a message is printed, the
function (invisibly) returns \code{NULL} and no plot is drawn.
}
\description{
This function takes a \code{state.num} argument, indicating the numerical
code of a given state, and a \code{year} argument, indicating the year
whose data one wants to plot. It then reads in the corresponding FARS file
and generates a plot with accident locations in that year and state,
superimposed on a map of said state. \code{fars_map_state} uses the
\code{map} function from the \code{maps} package to draw state boundaries.
It also uses the \code{filter} function from the \code{dplyr} package and
the \code{points} function from the \code{graphics} package.
}
\note{
Some of the states in a FARS file may not be present in the database
used by the \code{map} function in the \code{maps} package, which uses the
\code{\link[maps]{state}} database, containing 48 states. For example,
accident locations for the state of Hawaii, with \code{state.num = 15}, cannot
be plotted and calling the function for this state will result in an error.
}
\examples{
fars_map_state(1, "2013")
fars_map_state("12", 2014)
\dontrun{fars_map_state(14, 2014)}
\dontrun{fars_map_state(15, 2014)}

}
