% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_original.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Produce a plot of the location of the accidents on a state map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number, numeric integer value.}

\item{year}{A year, numeric value in the form yyyy of the accidents.}
}
\value{
This function returns a plot of a state map with points representing the location of the accidents in the year.
   If the state is not a valid number or it is not found in the data then the function stops and returns an error
   with the message "invalid STATE number: ". If there is no accident for the specified state and year then the
   function returns the message "no accidents to plot".
}
\description{
This function returns the location of the accidents for a specified state and year.
}
\examples{
\dontrun{
fars_map_state(40, 2014)
}

}