% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Title Function for creating a map of occurrence by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer containing unique state numbers}

\item{year}{An integer from \code{\link{year}}}
}
\value{
A map based on plots by latitude and longitude or Returns NULL if \code{data.sub}==0
stops if \code{state.num} is invalid, returns "invalid STATE number"
}
\description{
Title Function for creating a map of occurrence by state and year
}
\examples{
fars_map_state(6, 2013)
\dontrun{
fars_map_state(0, 2013)
}
}
