% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot te accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The code of the state}

\item{year}{The desired year in YYYY format}
}
\value{
A plot with the accidents occurred by latitude and longitude
}
\description{
Plot the accidents occured into an state along one year
}
\details{
Print "fars_map_state"
}

