% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plot fars data on map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer referencing a state.}

\item{year}{An integer, or a string}
}
\description{
\code{fars_map_state} maps fars data for each year and to a state
}
\details{
takes a state number and a year, and creates a map image plotting
the fars data point by long and lat, require mapdata package.
}
\note{
If the \code{state.num} provided is not in the FARS file, an error is produced.

If there are no fatal crashes for the state and year, a message is provided.

LONGITUD values > 900 are set to NA, LATITUDE values > 90 are set to NA
}
\examples{
fars_map_state(22, 2015)

\dontrun{
fars_map_state(1066, 0)   # error
}

}
