% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw the accident rate map of the state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{years}{The state number and year(2013-2015)}
}
\value{
A map object showing the accident numbers on the map
}
\description{
The function uses the accident, longtitude and latitude to draw the accident rate map for each state. If you enter an invalid state number or year, there will be an error
}
\examples{
fars_map_state(1, 2013)
}
