% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Displays a map of a state with the accident locations in the given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A value representing the state
numerical code. It will be coerced to integer,
therefore it does not need to be of a specific type,
as long as the coercion is feasible.}

\item{year}{A value representing the year. It will
be coerced to integer, therefore it does not need
to be of a specific type, as long as the coercion
is feasible.}
}
\description{
Checks for state code validity, accident presence in the
specified subset and latitude/longitude range.
Returns NULL invisibly.
}
\examples{
fars_map_state(6, 2015)

}
