% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number indicating a state of interest}

\item{year}{A number of year of interest}
}
\value{
this function returns a map. If state number does not exist,
a warning message would print. If no accidents in the state in the year,
no plot would be generated and a message would print.
}
\description{
this is a function that creates a map of a specified state
at a specified year
}
\examples{
\dontrun{fars_map_state(1,2013)}

}
