% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An Integer with the State Code.}

\item{year}{An integer, a numeric or string of the year.}
}
\value{
None.
}
\description{
This function provide a plot with a state map including the accidents
location by year.
}
\examples{
\dontrun{
fars_map_state(35, 2013)
}

}
