% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the accident in a given state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value of State number you want to check}

\item{year}{A numeric value of the year you want to check.}
}
\value{
A plot of US map where the dots represent the amount of accident in a given year
}
\description{
Filter out the data based on state number and year.
Will throw out error if the state number is invalid
Will return NULL in case there is no accident data in a given state
}
\examples{
\dontrun{fars_map_state(4124,  2015)}
}
\seealso{
\code{\link[=fars_read]{fars_read()}}
}
