% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{ID of the state.}

\item{year}{Year.}
}
\value{
The function returns a printed map of the state with marked accident locations.
}
\description{
The function prints a map of accident locations for a given state
}
\details{
The function can give an error if the STATE number is invalid.
It can also produce an error if the year is invalid and so the data file cannot be
read. The function uses make_filename(), fars_read(), so it can produce an error also
whenever these funtions would.
}
\examples{
\dontrun{
fars_map_state(6,2013)
}
}
