% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function to reads a file based on year parameter and plots accident data based on state parameter.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Valid State number as string or integer.}

\item{year}{Valid year as integer or string.}
}
\value{
A plot of accidents by longitude and latitude for state number and year.
}
\description{
This function takes state and year as parameters.
It reads file based on year parameter and gathers data to plot based on state number parameter.
It converts state number parameter as integer and stops if it cannot find state number is invalid.
It outputs a message if does not find any accidents to plot.
It returns the data plotted for state number by LONGITUDE > 900 and LATITUDE > 90
It uses functions: make_filename and fars_read
}
\examples{
\dontrun{
fars_map_state(25, 2015)
fars_map_state(state.num=25, year=2015)
fars_map_state("25", "2015")
fars_map_state(state.num="25", year="2015")
fars_map_state(25, "2015")
fars_map_state(state.num=25, year="2015")
fars_map_state("25", 2015)
fars_map_state(state.num="25", year=2015)
}

}

