% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map accidents in a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numerical code for a US state.}

\item{year}{The year of interest.}
}
\value{
This function returns a map of accients in a given state and year.
}
\description{
This is a function that takes a state and a year, and generates a map of the
state and then plots the areas where accidents occur.
}
\examples{
fars_map_state(state.num = 28, year = 2014)

}
