% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the Accidents for a State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the state of interest}

\item{year}{An integer representing the year of interest}
}
\value{
This function returns a map object got a given state, with the accident locations
   plotted on it
}
\description{
This function maps the accidents for a given state (represented by \code{state.num}) and
\code{year} onto a map of that state.
}
\details{
The function checks that the state is a valid state in the data set.

If there are no accidents, no plot is produced.
}
\examples{
fars_map_state(12,2015)
fars_map_state(21,2007)

\dontrun{
fars_summarize_years(87,"aaa")
}

}
