% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots a map for a specific state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{give number of state}

\item{year}{a string}
}
\value{
plots a graph for defined state
}
\description{
Plots a map for a specific state
}
\details{
This function plots a graph for definded state.
}
\examples{
 \dontrun{fars_map_state(29, 2015)}
}
\author{
James Bertschy \email{james.bertschy@gmail.ch}
}
