% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{create a map of fars events given a state number and a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{is an integer}

\item{year}{an integer}

\item{STATE}{a character string}
}
\value{
This functions returs a map with a point in every place there has been a fatality in a motor vehicle accident
}
\description{
create a map of fars events given a state number and a year
}
\examples{
\dontrun{fars_map_state(1,2013)}
}
