% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the accidents of one specific state and a range of years.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of the state to analyze.}

\item{year}{The year to analyze.}
}
\value{
A plot with the accidents of one specific state and a year.
}
\description{
Returns a plot of the accidents based on the \code{state.num} and \code{year} inputs. Returns an error if the state or year do not exist in the data set.
}
\examples{
fars_map_state(15, 1992)
fars_map_state(state.num = 15, year = 1992)

}
