% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plot the map of accidents with the state ploted on the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of a state is a numeric or an integer in the US as used in the FARS datasets}

\item{year}{field which becomes an integer input}
}
\description{
plot the map of accidents with the state ploted on the map
}
\note{
this function is to create map, which plots the location of the record on the map for the specified state and year

the state number should be integer or numerical and should exist in the data, otherwise the function terminates with an error
}
