% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accident data on a map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that represents a state. For more information,
please see the \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/812449}{FARS data manual}, pg. 60.
Single digit numbers should be formatted as single digit, not double digit
as indicated in the manual.}

\item{year}{A character string representing a year,
four-digit format (i.e. 2013)}
}
\value{
None
}
\description{
\code{fars_map_state} takes a state and a year and plots on a map
all accidents from the FARS dataset for that state and year.
}
\details{
This function will raise an error if there is no file
that matches \code{year}. This function will also raise an error
if the \code{state.num} doesn't match a number from the data.
}
\examples{
\\code{fars_map_state(36,'2013')}

}
