% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map fatalities within a state for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer specifying which state you want to see traffic
fatalities for.}

\item{year}{a four digit integer representing the year you want to see traffic
fatalities for.}
}
\value{
This function returns a map plot of the state fatalities in a year.
}
\description{
This function plots a map showing where traffic fatalitites occurred in a state,
specified by argument \code{state.num}, for a given year, specified by argument
\code{year}.
}
\details{
This function uses \code{make_filename()} and \code{fars_read()} to read in the
.csv that corresponds with parameter \code{year}. Parameter \code{state.num} is
coerced to be an integer and checked for valid existence.
Logitudes of greater than 900 and latitudes of greater than 90 coerced into NAs.
\code{maps::map()} is used to create the map. \code{graphics::points()} is
used to plot traffic fatalities on the map.
}
\note{
If you supply an invalid \code{state.num} to this function, you'll receive error
message: "invalid STATE number: ". If there are no traffic fatalities for the
state and year, you'll receive a message: "no accidents to plot".
Because this function coerces longitudes greater than 900 and latitudes
greater than 90 to NA, you could receive a message saying, "nothing to draw:
all regions out of bounds". \code{fars_map_state(2, 2013)} for example will
return this message.
}
\examples{
fars_map_state(48, 2015)

}
