% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer id of US state}

\item{year}{integer vector of years}
}
\description{
Maps yearly accidents in a US state as points.
}
\note{
Throws an error if the numerical id of a state is invalid.
}
\examples{
if(interactive()){
   fsum <- fars_map_state(1, 2013:2015)
}

}
