% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing a US State}

\item{year}{A integer that represent a year we are looking the data for}
}
\value{
map a map object. If the state id is not correct the function stops.
    No measures return as message.
}
\description{
Function that return the map of a state accidents in given year(s)
}
