% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map of fatal accidents}
\usage{
fars_map_state(state.num = 1, year = 2013)
}
\arguments{
\item{state.num}{state number (numeric, 1-56)}

\item{year}{(numeric)}
}
\description{
From specified year and state number, create a map of the accidents
as included in the (FARS) data.
Only data for the years 2013, 2014 or 2015 is available.
}
\details{
Throw an error if the state number is invalid or there are no accidents to plot
}
\examples{
 \dontrun{fars_map_state(1,2013)}
}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
