% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents on US state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number of a state.}

\item{year}{A year.}
}
\value{
A plot is created where each accident of the specific state and the
specific year is shown on a map. In addition, \code{NULL} is returned. If the
type of \code{state.num} or \code{year} is not correct or their values
are invalid, an error message is generated and nothing is plotted
nor returned. If there are no accidents, an appropriate message is generated
and nothing is plotted again.
}
\description{
The function \code{fars_map_state} reads data from the US National Highway
Traffic Safety Administration's Fatality Analysis Reporting System and plots
all accidents of a specific state and year on a map.
}
\examples{
setwd(system.file("extdata", package = "fars"))
fars_map_state(1, 2015)

}
