% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Overlay accidents locations onto map of given state in given year}
\source{
Coursera, Building R Packages
}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the index number of a state as an integer corresponding to a required state.
The function throws an error if no fatality information is available for
the requested state in the FARS data for the requested year.}

\item{year}{A \code{Year} provided as a string or an integer. If
the year is absent from the FARS data, this throws an error.}
}
\value{
NULL return value. The function plots a graph as a side-effect
  If no accidents were observed in the state/year, no map is plotted.
}
\description{
The accident locations are plotted on the map of STATE given as input argument (integer).
Function checks whether there is such STATE for given YEAR, and throws "invalid STATE number" if it is missing.
If there are no accidents for given state, then "no accidents to plot" information is presented.
Read single FARS file from working directory with given YEAR argument into tbl_df objects.
}
\note{
\itemize{ \item State number 2 do not work: "nothing to draw: all regions out of bounds". Possible bug
\item If LONGITUD is bigger than 900 or LATITUDE bigger than 90, their value is set to NA.
\item Function uses make_filename(year) and fars_read(year) functions, and readr, dplyr, and maps packages.
\item See also \code{fars_read}
}
}
\examples{
fars_map_state(44,2013)
\dontrun{fars_map_state(2,2015)} # do not works for this year and 2013,2014 as well. Returns: nothing to draw: all regions out of bounds

}
\keyword{FARS,}
\keyword{NHTSA,}
\keyword{Road}
\keyword{safety}
\keyword{traffic}
