% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This is a function that generates a state map showing the locations of all accidents for a specific year. It uses \code{\link{make_filename}}, \code{\link{fars_read}}.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number to generate the map for. It has to be an integer or a character string that can be converted into an integer, otherwise an error will be thrown and the function will halt. It must be present in the data files too, otherwise the function will stop and print an error message.}

\item{year}{An integer or character string that can be converted into an integer.}
}
\value{
A state map showing the locations for all accidents for a specific state.
}
\description{
This is a function that generates a state map showing the locations of all accidents for a specific year. It uses \code{\link{make_filename}}, \code{\link{fars_read}}.
}
\examples{
\dontrun{
year = 2015
fars_map_state(1, 2015)
fars_map_state(state.num = 1, year = 2015)
fars_map_state(state.num = "1", year = "2015")
}

}
