% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function \code{fars_map_state} plots the contour of the state
in question and points with accident locations for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number, a value between 1 and 56}

\item{year}{Four-digit year, as integer or as a string (i.e. 2001, "2001").}
}
\value{
Plot of a contour of the state in question with geographical locations
        of accidents (as dots) for a given year.

If the requested year input is not present in available data, the function stops
and returns an error:
"file 'filename_YEAR.csv.bz2' does not exist".

If the requested state number is not present in data, the function stops and
returns an error:
"invalid STATE number: number"
}
\description{
Function \code{fars_map_state} plots the contour of the state
in question and points with accident locations for a given year.
}
\examples{
\dontrun{
fars_map_state(1,2013)
fars_map_state("1","2013")
}
}
