% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_pack.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{- State number to provide a plot [Integer]/[String] 2 digits}

\item{year}{-   Years as a 4 digit number - The year  can be provided as single year, as a vector or list of years}
}
\value{
dataframe -  of filtered data, if there are no observations for a given state then returns (NULL)
}
\description{
fars_map_state plots the state given as an integer argument and shows accidents
for a given year using the LONGITUD and LATITUDE points stored within the data.
}
\note{
This function depends on map and points functions from the maps library.
Errors occor when an invalid State.num is provided
Other errors can be linked to the fars_read function.
}
\examples{
   fars_map_state(10,2013)


}
\references{
The US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census, provides the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
You can download the FARS data at \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\seealso{
\link{make_filename}
 \link{fars_read}
}
