% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{The final function produces a picture plotting
the observations (accidents) that happen in a state
in a given year.
It takes only the observation that belong in the state (using filter)}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{input character}

\item{year}{input numeric}
}
\value{
plot of observations
}
\description{
The final function produces a picture plotting
the observations (accidents) that happen in a state
in a given year.
It takes only the observation that belong in the state (using filter)
}
\examples{
fars_map_state("21",2014)
fars_map_state("30",2015)
}
