% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display accidents map by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An Integer with the \code{State Code}}

\item{year}{A string, or an integer, with the input \code{year}}
}
\value{
Returns a map, or an error if the \code{state.num} is invalid the
function shows an error.
}
\description{
Draw a geographical map of accident sites for a specific US state and year.
}
\examples{
\dontrun{
fars_map_state(49, 2015)
}

}
\seealso{
\link{fars_read}

\link{make_filename}
}
