% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents data by state num and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{An integer with the value of the year}
}
\value{
Map with the points where the accidents occured
}
\description{
This function filters the data set by state number and year and plot the
results in a map
}
\examples{
\dontrun{
fars_map_state(1,2013)
}
}
