% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map showing locations of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric constant giving the state id}

\item{years}{A numeric constant giving the year in four-digit 
format, e.g., 2013.}
}
\value{
This function plots a map and does not return anything.
}
\description{
This function uses FARS datasets to show a map of where accidents 
have occurred in a specified state and year.
}
\note{
The function stops if state.num is not a valid state id.
If there are no accidents, a message is printed to the console.
This function uses the functions \code{make_filename} and 
\code{fars_read}.
}
\examples{
fars_map_state(state.num=1, year=2013)

}
