% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents in a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric scalar. The state number of the accidents.}

\item{year}{A numeric scalar. The year of the accidents.}
}
\description{
\code{fars_map_state} plots the accidents in a state for a given year. The
state is plotted as a map and the accidents as points on the map. Throws an
error if the state number does not exist in the data and NULL if the
accidents are zero for the given year in the specified state.
}
\examples{
\dontrun{
fars_map_state(1, 2014)
fars_map_state(5, 2014)
}
}
