% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{faras_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number corresponding to the state of interest}

\item{year}{year of interest}
}
\description{
Plots the accidents from a state in a specific year
}
\examples{
\dontrun{
fars_map_state(13, 2014)
}

}
