% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number representing the FARS code of the desired state.
This function returns an error, if the state number if invalid.}

\item{year}{A number representing the desired year.}
}
\value{
This function returns a map with the locations of accidents for
the selected state and year.
}
\description{
This function plots a map with the locations of accidents for the selected
state and year.It calls the functions \code{make_filename} and
\code{fars_read} from within. T
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))
fars_summarize_years(2013:2015)

}
