% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the location of the accident}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{Year (in numbers)}
}
\value{
Returns "NULL". The function plot the accidents.
}
\description{
Plot the location of the accident for the specified state number and year.
}
\examples{

fars_map_state(21, 2013)

}
