% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the number of accidents by State and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{vector or list representing the state number}

\item{year}{vector or list representing the accidents years}
}
\description{
This function plot the number of accidents in the FARS dataset by states and years.
}
\section{\code{Error}{
}
If \code{years} contains an invalid year, the function gives a warning message.
if \code{years} does not contain a valid year, the function returns an error
if \code{state.num} does not contain a valid state number, the function returns an error
Error in get(dbname): object 'stateMapEnv' not found. It seems the object is missing in the maps library
}
\examples{
fars_map_state(1,2013)
fars_map_state(c(1,2),c(2013,2014))
}

