% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map with state shape and accidents.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{year}{An integer or character string that can be converted to integer
 meaning year}

\item{state}{An integer or character string that can be converted to integer
 meaning state}
}
\value{
This function returns a map with state shape and marked accidents.
}
\description{
This is a simple function that reads csv file for particular year.
Filter data for particular state. In case if state does not exist
appropriate message will be shown. If there is no accidents in particular state
within specified year appropriate message will be shown.
Function plots state shape and marks accidents as a dots.
}
\examples{
fars_map_state(1,2013)
}

