% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of a USA state in form of integer of character}

\item{year}{year as a number in form of integer of character}

\item{STATE}{selected state}
}
\value{
This function returns a map of a state, with points as location of
accidents
}
\description{
This function gets a number of of USA stat and a year, and visualize
state map, with accidents as points in it on that year.
}
\examples{
\dontrun{fars_map_state("53","2013")}
\dontrun{fars_map_state("4",2015)}
\dontrun{fars_map_state(10,2014)}

}
