% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Crear un mapa de eventos}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{valor entero que indica el estado}

\item{year}{valor entero que indica un año}
}
\value{
Esta funcion regresa un mapa con los datos proporcionados
}
\description{
Esta funcion permite crear un mapa de eventos dado years
}
\details{
Debe escribir un numero de estado existente, sino generará un mensaje de advertencia
}
\examples{
\dontrun{
fars_map_state(1,2013)
fars_map_state(state.num=1,2013)
fars_map_state(state.num=1,year=2013)
}

}
