% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map distribution of accidents for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer between 1 and 56 representing the state number}

\item{year}{An integer representing an year}
}
\value{
This function plots a map of a state with accident data represented by point.
It does not return anything. Not inputing an integer in \code{state.num} will result
in an error.
}
\description{
This function can be used to draw a map of a state, with a dot for each registered
accident in a given year. The state is selected via the \code{state.num} argument,
the year through the \code{year} aergument.
}
\examples{
\dontrun{
fars_map_state(1,2015)
}

}
