% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize the accidents in the US map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{Selected year(s)}
}
\value{
\code{fars_map_state} Output the plot of accidents for the selected
  state and year. The error occured if the state or year does not exist in the
data set.
}
\description{
Plot the accidents on the US map for a given state and year.
}
\examples{
\dontrun{
fars_map_state(5, 2013)
fars_map_state(20, 2015)
}

}
\seealso{
\code{\link{make_filename}}
\code{\link{fars_read}}
}
