% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Functions that read traffic crash from FARS dataset.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{US state number (represented as numerical alphabetical order,
i.e. Alabama=0, Alaska=1, Arizona=3, ... )}

\item{year}{4 digit year number/string}
}
\value{
This function returns a message indicating no valid state number \code{satate.num}
or unavailable data if no crashes occurred.
}
\description{
This function plots a map with traffic crashes location based on
FARS data, occurred in a specific US state. It will check if no data is available
and if location coordinates are valid. To explore traffic crashes,
\itemize{
\item Check at \href{https://www.nhtsa.gov/node/97996/251}{avaiable datasets}, data to plot/explore.
\item Download CSV file locally, extract contents , i.e. for National data from 2000
curl::curl_download(url = "https://www.nhtsa.gov/filebrowser/download/152076", destfile = "FARS2000NAtionalCSV.zip");
extract its contents, rename and bzip2 file "accidents.csv" file as "accident_2000.csv.bz2"
}
}
\details{
Available data starts from 1975 (as a reference:\href{https://www.nhtsa.gov/node/97996/251}{avaiable datasets}).

Print "fars_map_state"

It will present a error if  parameter represents an invalid
state number according to the local data base.
}
\note{
This information is provided by FARS (Fatality Analysis Reporting System),
which became operational since 1975. Data is based on the 50 states, Puerto Rico
and the District of Columbia, containing census of fatal traffic crashes.
Trained employees from FARS gather more the 140 data elements per traffic crash,
for full details, check \href{https://www.nhtsa.gov/crash-data-systems/fatality-analysis-reporting-system}{FARS data})
}
\examples{
\dontrun{
fars_map_state(1, 2011) # generates a crash location map for Alaska in 2011
}

}
