% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function creates a map of observations (fatal accidents) that have occurred in a given state during a specified year.}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System \url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number describing the state that is to be mapped}

\item{year}{the year that will be included in the string output}
}
\value{
The function returns a map of the specified state and the location of each observation (fatality) using longitute and latitude.
}
\description{
This function creates a map of observations (fatal accidents) that have occurred in a given state during a specified year.
}
\note{
This function directly calls make_filename() and fars_read(), so any notes and errors that apply to those functions
listed above also apply here.

The state.num must exist in the dataset.  If it does not an error will appear reading "invalid STATE number: "
}
\examples{
fars_map_state(28,2013)
}
\references{
The function calls make_filename() and fars_read() from this package
}
