% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS Entries for a State and Year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric indicating the state for the map. Will be coerced to an integer.}

\item{year}{Numeric indicating the year to map. Will be coerced to an integer.}
}
\value{
As a side-effect, plots a map of FARS entries for the state and year
}
\description{
This function takes as input a state and a year and plots a map of
the location of the entries in the US National Highway Traffis Safety
Administration's
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Anlaysis Reporting System}
(FARS) for the given state and year. The function throws an error if the state/year
combination do not exist in the FARS database.
}
\examples{
library(FARS)
fars_map_state(1,2013)


}
