% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a map of fatal accidents in a given US state and year from FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number for the state for which to map accident data. State numbers can be found
in the \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/812315}{FARS Analytical User's Manual},
page 26: "C1/V1/D1/PC1/P1/NM1 State Number" section.}

\item{year}{The numeric year value for the data to map.}
}
\value{
This function does not return a value (i.e., returns \code{NULL}).  As a side effect, this
   function causes a scatterplot map of the fatal accidents in the given state during the given year
   to be plotted together with the state boundary.
}
\description{
This function generates a map of fatal accidents in the given state from
FARS data for the requested years.
}
\details{
FARS is the US National Traffic Safety Commission's
\href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System}.
}
\note{
This function checks that the files for the requested years exist in the working directory,
and will exit, generating an error, if the file does not exist.  It also checks in the loaded
data for the validity of the state number given, and will generate an error if the state number
is invalid, or a message if there is no data for that state during that year.
}
\examples{
\dontrun{fars_map_state(22, 2014)     # Louisiana, 2014 data}
\dontrun{fars_map_state(11, 2015)     # District of Columbia, 2015 data}

}
\seealso{
\code{\link{make_filename}}, \code{\link{fars_read}}
}

