% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a FARS data map for a given year and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of a state in the US as used in the FARS data
sets. It should be numeric or integer.}

\item{year}{The year of analysis (numeric or integer)}
}
\value{
Returns a plot of the accidents based on the \code{state.num} and
\code{year} inputs. Returns an error if the state or year do not exist in the
data set.
}
\description{
Draws a map for a given year and state, one point for observation. It will
throw an error if state.num is not in the dataset. If no observations are found,
it will throw a message.
}
\examples{
\dontrun{fars_map_state(45, 2015)}
}
