% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the noted accidents on a map of the US by state}
\usage{
fars_map_state(state.num, year)
}
\value{
This function returns a plotted chart of the accidents on a
map. If the state or year input is invalide, and error will be given.
}
\description{
This function takes the numeric identifier of each state
from our source data files and plots the accidents on a map of
the United States. A stop clause exists if any of the state
numbers are not unique. A warning also exists in th event
that a row has 0 data points.
}
\examples{
\dontrun{
fars_map_state(22, 2014)
}

}
