% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state code of interest}

\item{year}{The year of interest}
}
\value{
This function returns a map of accidents that occurred within a specified
state and during a specified year. If the state code is not valid, the code stops
and an error is generated. If there are no accidents for the given year and month,
a warning is produced.
}
\description{
This function creates a map of accidents that occurred within a specified state
and during a specified year.
}
\note{
Examples: fars_map_state(12,2013), fars_map_state(1,2013)

If the file "accident_2050.csv.bz2" does not exist, attempting to create a map
from this file will generate an error.
try(fars_map_state(6,2050))
}
