% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates a map of number of fatalities for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Identifier for state number}

\item{year}{a single year for which to plot the data}
}
\value{
This function returns a map of fatalities for a given state and year
}
\description{
This function uses state number and year as input and if fatality data exists, it puts this data on a map
}
\examples{
fars_map_state(1, 2015)

}
