% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map with geolocated observations (cars accidents)}
\usage{
fars_map_state(state.num, year)
}
\description{
This functions takes a valid state number and a valid year
and produces a map with geolocated observations (accidents)
plotted as dots on the map.
}
\details{
@param state.num valid State number (between 1 and 51)
 @inheritParams fars_read_years

@return A map with geolocated accidents as dots on the map

@examples
 fars_map_state(51, 2015)
}
