% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create Fatality Analysis Reporting System incident map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{(\emph{integer}) An identifier specifying the state map to be drawn.
(see also the \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/812827}{dataset documentation}) \cr
\tabular{llll}{
\strong{state.num} \tab \strong{State Name} \tab
\strong{state.num} \tab \strong{State Name} \cr
1 \tab Alabama \tab 31 \tab Nebraska\cr
2 \tab Alaska \tab 32 \tab Nevada\cr
4 \tab Arizona \tab 33 \tab New Hampshire\cr
5 \tab Arkansas \tab 34 \tab New Jersey\cr
6 \tab California \tab 35 \tab New Mexico\cr
8 \tab Colorado \tab 36 \tab New York\cr
9 \tab Connecticut \tab 37 \tab North Carolina\cr
10 \tab Delaware \tab 38 \tab North Dakota\cr
11 \tab District of Columbia \tab 39 \tab Ohio\cr
12 \tab Florida \tab 40 \tab Oklahoma\cr
13 \tab Georgia \tab 41 \tab Oregon\cr
15 \tab Hawaii \tab 42 \tab Pennsylvania\cr
16 \tab Idaho \tab 43 \tab Puerto Rico\cr
17 \tab Illinois \tab 44 \tab Rhode Island\cr
18 \tab Indiana \tab 45 \tab South Carolina\cr
19 \tab Iowa \tab 46 \tab South Dakota\cr
20 \tab Kansas \tab 47 \tab Tennessee\cr
21 \tab Kentucky \tab 48 \tab Texas\cr
22 \tab Louisiana \tab 49 \tab Utah\cr
23 \tab Maine \tab 50 \tab Vermont\cr
24 \tab Maryland \tab 52 \tab Virgin Islands (since 2004) \cr
25 \tab Massachusetts \tab 51 \tab Virginia\cr
26 \tab Michigan \tab 53 \tab Washington\cr
27 \tab Minnesota \tab 54 \tab West Virginia\cr
28 \tab Mississippi \tab 55 \tab Wisconsin\cr
29 \tab Missouri \tab 56 \tab Wyoming\cr
30 \tab Montana\cr
}}

\item{year}{(\emph{character} or (\emph{integer}) ) The year for which the
filename will be created.}
}
\description{
Draws a map of Fatality Analysis Reporting System (FARS) incidents of a
given \code{year} for a state.
}
\details{
If FARS file for the given \code{year} does not exist in the current
working directory or if an undefined \code{state.num} is selected, an error
is thrown.
}
\examples{
\dontrun{
# draw map of Alabama for year 2014
fars_map_state(1, 2014)
# draw map of Nevada for year 2015
fars_map_state(32, 2015)
}

}
\references{
US National Highway Traffic Safety Administration \cr
(\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\seealso{
\code{\link{fars_read}}, \code{\link{make_filename}}
}
