% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a map of accidents of a specific state during a specific year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{(numeric) The state number}

\item{year}{(numeric) The year}
}
\description{
Not suitable for plotting multiple years or multiple states. FARS files with
standard naming are expected to be found in the working directory. States
that can not be found in a file or state.num year combinations without
accidents will raise errors.
}

