% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{FARS map states.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The State identifier as an \code{integer} or a \code{character} .}

\item{years}{A vector containing the years of interests as \code{integer} or
\code{character}.}
}
\description{
Plots on a USA State map the fatal injuries suffered in motor vehicle traffic crashes for
the requested year and state.
}
\details{
This function performs the steps below:
\enumerate{
  \item Build the FARS file name from \code{year} (using \code{\link{make_filename}}).
  \item Read the FARS data file in R as a data frame tables (using \code{\link{fars_read}}).
  \item Plot on a USA State map all the incidents related to the selected State (\code{state}).
}

If the provided \code{state} is not valid then execution is stopped and an
error is returned.

If the provided \code{state} is valid and does not have any incidents then
execution is stopped and a message is returned.
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{filter}} in the \code{dplyr} package.
  \item \code{\link[maps]{map}} in the \code{map} package.
  \item \code{\link[graphics]{points}} in the \code{graphics} package.
}
}

\examples{
\dontrun{
fars_map_state(4, 2013)
fars_map_state("4", "2013")
}

}
