% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accident data per state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric or integer value representing the state in the Fatality Analysis Reporting System}

\item{year}{A numeric or integer valaue representing the year that needs to be plotted}
}
\value{
Returns a plot of all accidents using the \code{state.num} and \code{year} parameters.
Returns an error if the state number or year is invalid.
}
\description{
This function will plot the accidents for a selected state and year. If the state number is not a valid integer or numeric,
the function will stop with an error message. The function will also return an error if the year is invalid
}
\examples{
\dontrun{
fars_map_state(25,2014)

# Error
fars_map_state(111,2014)
fars_map_state(25,2016)
}

}
