% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map with Fatality Analysis Reporting System data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value with the state number to plot}

\item{year}{A numeric value with the year for which the data will be plotted.}
}
\value{
Plots a map with points marking the locations of accidents.
}
\description{
This function creates a map with fatalities in motor vehicle crashes for a
specific year and state. The function fails if an invalid year or state is
passed as input.
}
\note{
The csv files with the data need to be downloaded previously from the
  US National Highway Traffic Safety Administration's Fatality Analysis
  Reporting System website
  (\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\examples{
fars_map_state(3, 2010)

}
