% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Fars Map State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the state number that corresponds to either a US state or territory. If an invalid state number
is used, an erorr will occur.}

\item{year}{the year of interest. can be string or integer.}
}
\value{
the corresponding state map with the fatal accidents represented as points
}
\description{
Displays the state map for a given year, highlighting the area with points to represent a fatal accident occuring
at that location
}
\examples{
fars_map_state(50,2013)

}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read}}
}
