% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw accident locations on the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{year}{specify the year of data}

\item{stat.num}{a state number shown in the "state" column data}
}
\value{
a map based on the Longitude and Latitude in the data
}
\description{
Draw accident locations on the map
}
\note{
"maps" package is required
}
