% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents on state-level map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric which represents the state number. If no accidents
occurred in the state during the specified year, than a "no accidents to
plot" message appears. If the state number is invalid, an error message,
"Invalid state number" is generated.}

\item{year}{A numeric which represents the year of the data. The value must
correspond to a year included in the filename of a CSV file in the working
directory such as "accidents_2013.csv.bz2" where the 2013 in the filename
represents the year of the data. No corresponding file generates and error
message.}
}
\value{
A plot of points representing accidents in a given year within a
state-level map.
}
\description{
The \code{fars_map_state()} function plots the locations of accidents on a map for
a specific state and year using data from the Fatality Analysis Reporting
System (FARS). The function takes two parameters: \code{state.num}, which represents
the state number and \code{year}, which represents the year of the data.
}
\section{External Functions Not Imported}{

\itemize{
\item \code{\link[dplyr:filter]{dplyr::filter()}} used to create a subset of the data.
\item \code{\link[maps:map]{maps::map()}} used to plot a state map.
\item \code{\link[graphics:points]{graphics::points()}} used to plot points on a map.
}
}

\examples{
\dontrun{
fars_map_state(10, 2013)
fars_map_state(5, 2015)
}

}
\seealso{
\itemize{
\item \code{\link[=make_filename]{make_filename()}} to create a file name from a year
\item \code{\link[=fars_read]{fars_read()}} to read CSV file to tibble
}
}
