% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Print "fars_map_state"}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A int value that represent a state}

\item{year}{A int value contained in file name}
}
\value{
This function returns a plot with points that are composed of LATITUDE and LONGITUD in state.num and year
}
\description{
This is a simple function that graphs geographic points given the state number and the year
}
\details{
If state.num or year are not in the data then this function fails
}
\examples{
\dontrun{
year <- 2013
state.num <- 1
fars_map_state(state.num, year)
}

}
