% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot points at all fatal accident sites for a given year and state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number representing a state.  The number must be
coercible to an integer.}

\item{year}{year as a four digit number, or anything coercible to
four digit number.}
}
\value{
NULL - this function is called for its side-effects.
}
\description{
This function plots points on a map, showing the location of all
fatal accidents that occured in the given year and state.
\code{fars_map_state} takes two arguments, \code{state.num} and \code{year}.
\code{state.num} specifies the state while \code{year}
specifies the year for which fatal accident, location points
should be plotted.  If the data doesn't have \code{state.num}
then \code{fars_map_state} throws an error.  If no fatal accidents
were recorded in the given state and year then \code{fars_map_state}
prints an associated message, without plotting any points.
}
\examples{
fars_map_state(15, 2012)
fars_map_state(5, "2012")

}
