% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data to a state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An \code{intger} represents the identifier of a state.}

\item{year}{An \code{integer}, or any other type could be coerced to an integer, represents the year of the \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data file.}
}
\value{
A plot of a state from \code{\link[maps]{map}} and points correspond to the coordinates of recorded accidents.
}
\description{
This function plots individual \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} accidents' coordinates as points on a stat map
}
\details{
The function returns an error \code{invalid STATE number: state.num} if the provided number does not correspond to a STATE in the data of this particular year and a message \code{no accidents to plot} if the number of accident for this year is 0.
otherwise it constructs a file name using the provided year, read the data file and plot points of individual accidents on the STATE map.
}
\examples{
\dontrun{
fars_map_state(50, 2013)
}

}
