% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function generate a map of all FARS events in one or more US states, with each event displayed as a dot on the map.
User may specify the state(s) by unique ID# and year from which FARS event data should be drawn.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer vector that uniquely identifies one or more US states in the FARS dataset}

\item{year}{An integer representing the year that will be inserted into the file name}
}
\value{
Returns map plot of one or more US states with black dots representing FARS events in the selected year
}
\description{
This function generate a map of all FARS events in one or more US states, with each event displayed as a dot on the map.
User may specify the state(s) by unique ID# and year from which FARS event data should be drawn.
}
\note{
This function throws an error if an invalid \code{year} is selected (not a number or not a year in the available dataset(s))

It also throws an error if an invalid \code{state.num} is specified

More than one state may be selected at a time, generating a country-wide map with data plotted for selected states

If a valid subset of data is selected, but not FARS events are avaialble, the function will print a message indicating that
}
\examples{
\dontrun{fars_map_state(1, 2013)}
\dontrun{fars_map_state(1:50, 2013)}
\dontrun{fars_map_state(75, 2013)}

}
