% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generates a plot of all the accidents for a state and year combination}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number (1 to 12) (integer/numeric) as used in the data files}

\item{year}{Year (integer/numeric) in the format yyyy}
}
\value{
A NULL, however it generates a plot of all the accidents in the state for
the year
}
\description{
This function takes in an argument of a state number (1 to 12) and
year (yyyy) and generates a plot of all the accidents in the state,year
combinaiton.
}
\note{
If the state number does not exist in the CSV file for the year
then it stops with an error.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
fars_map_state(10, 2014)
}

}
