% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot state and points where accidents occurred.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An atomic numeric vector representing the state in the
USA of interest.}

\item{year}{An atomic character or numeric vector representing the year of
interest.}
}
\value{
Outputs an outline of the state with points representing the
accidents that occured during the given year in that state.
}
\description{
\code{fars_map_state} plots the outline of a state given as an integer
argument and plots the points where all the accidents occured in that state
for a given year.
}
\details{
The function filters out the relevant points of longitude and latitude
where accidents occured and plots them within an outline of the state
in which they occurred. Further, the points represent where the accidents
occured within a given state for a given year.
}
\section{Required functions to import from other packages}{

maps::map; note that graphics::points is part of the base graphics
and does not need to be imported.
}

\examples{
fars_map_state(1, 2013)
fars_map_state(1, "2013")

\dontrun{
## The following won't work as the year argument must be an element
## that can be converted into an integer.
fars_map_state(1, 2/3/2011)
}
}
