% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize the accident summary in US map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{Integer of year to use}
}
\value{
This function returns a plot of the number of accidents
   for the specified state.
}
\description{
Accident events are shown on the US map for a chosen state and year.
}
\examples{
\dontrun{fars_map_state(12, 2012)}

}
