% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Make a plot for a particular state for a year.}
\source{
This function uses the function make_filename() inside.
}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{Input}{class 'integer' value for the 'STATE number'}

\item{Input}{class 'integer' value for the 'year'}
}
\value{
This function returns a plot displaying density of accidents for a
certain state number in a particular year.
}
\description{
This function shows a geographical plot of 'dots' according to number of accidents
in a state within a year.
}
\note{
The two packages 'stats' and 'maps' must be preloaded before compiling
this function. Otherwise there would be a situation to get an error as: "Error
in get(dbname) : object 'stateMapEnv' not found"
}
\examples{
\dontrun{
fars_map_state(1,2015)
fars_map_state(56,2013)
}

}
