% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function plots a point at the state number where the accidents reported for each year
happened. The year and state must be provided in order to plot the accidents.
If no accidents were reported in a state for a given year,then there will be no points to plot.
If the number provided for the state is  not in the desired range then an error will be given.}
\usage{
fars_map_state(state.num, year)
}
\description{
@param state.num an integer indicating the id for the state to be plotted
}
\details{
@inheritParams make_filename

 @return a map of the state and location plotting the points where the accidents were reported for the
 given state by year

 @importFrom dyplyr filter

 @importFrom maps map

 @importFrom graphics points

 @examples
 \dontrun{
 fars_map_state(01,2013)
 }

 @export
}
