% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Returns a map of an specified state including the accident locations.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer value representing the state.}

\item{year}{The integer value representing the year to be considered.}
}
\value{
Returns a map of an specified state given by \code{state.num} including the accident locations.
}
\description{
Returns a map of an specified state including the accident locations.
}
\examples{
\dontrun{fars_map_state(30, 13)}
}

