% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state(): Maps accident data of specific state in a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric value of US State (Min:1, Max:56)}

\item{year}{An year of interest in which data to be read (2013-2015)}
}
\value{
Displays a plot of state map with accidents data. If there's no accidents in the state, 'no accidents to plot' is printed instead.
}
\description{
This function accepts a specific year and a numeric value of US State.
Afterwards, the accident data of that year is read.
If the state number is invalid, error msg is displayed.
Otherwise, accident out of that given state is filtered out.
Further, when no accidents data are available, 'no accidents to plot' is to be shown.
Otherwise, accidents data is to be mapped within the state geography.
}
\examples{
\dontrun{
fars_map_state(1,2013)
}
}
