% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Maps accidents onto states}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numerical integer denoting the US state as is shown in the data set}

\item{year}{A numerical integer denoting the year}
}
\value{
Returns a plot of states with the number of accidents on each states
the accidents based on the \code{year}. Returns an error if \code{state.num} or if 
\code{year} do not exist in the data set.
}
\description{
Takes input of a \code{state.num} and \code{year} as arguments and plots the 
accidents onto a map of the states. If the state number is invalid, an error is thrown. 
If there are no accidents in that state, a message is returned that there are no 
accidents to plot.
}
\examples{
\dontrun{
fars_map_state(20, 2013)
fars_map_state(10, 2014)
fars_map_state(30, 2016)

fars_map_state(72, 2013) #Error because \\code{state.num} doesn't exist
fars_map_state(20, 2020) #Error because \\code{year} doesn't exist
}

}
