% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate a graphical map of accident data for a U.S. state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a string or integer representing the position of a state in alphabetical order. One of the states in the R standard dataset state.names}

\item{year}{a string or integer representing the year for which to load highway accidents.}
}
\value{
a graphical map with points for individual highway accidents
}
\description{
This is a function that creates a graphical map of highway accidents given a state number and year
}
\details{
If a accident file with standard naming convention is not in currect working, a "invalid year" warning is generated. If state.num is not between 0 and 50, a "invalid STATE number" error is thrown. If a particular state had 0 accident records for a given year, a "invalid STATE number" error is thrown.
}
\examples{
\dontrun{
fars_map_state(1,1990)
fars_map_state("1","1990")
}

}
