% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS accident data on maps of US states}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The numeric code (as numeric or integer) of the US state to be analysed. Numeric codes can be found e.g.
\href{https://en.wikipedia.org/wiki/Federal_Information_Processing_Standard_state_code}{on Wikipedia}.}

\item{year}{The year (as numeric or integer) from which the data is to be plotted.}
}
\value{
The function returns a plot with a graphical representation of the selected US state with the
       accidents of the indicated year plotted as dots on the map according to their geographic
       location. If the data set contains no accidents for the indicated year and state,
       the message \code{"no accidents to plot"} is printed to the console and if an invalid
       state number is given as input, a warning will be displayed.
}
\description{
This function makes use of the geographic data contained in FARS accident data sets to plot
accidents on maps of US states.
}
\details{
The function internally uses the \code{\link{make_filename}} and \code{\link{fars_read}}
        functions to read in the FARS data for the given year and than filters for the given state.
        The geographic data is extracted and plotted using the \code{maps} package.
}
\examples{
\dontrun{fars_map_state(31,2015) #Plot all accidents that happened 2015 in Nebraska.}

}
