% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map Years Fatal Accidents onto State Map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{year}{Numeric, integer or numeric/integer string corresponding to "accident_<year>.csv.bz2" file in the
working directory}

\item{state}{Numeric, integer or numeric/integer string corresponding with a US state code.}
}
\description{
Draws statemap and plots fatal accidents from fars file for specified year and state.
}
\note{
If state number does not exist in file will generate an error
If no records for state exist will write message and return an a invisible(NULL)
}
\examples{
\dontrun{
fars_map_state(1 , 2018)
fars_map_state('1' , '2018')
}

}
