% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the accidents on a US state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of a state in the US as used in the FARS data
sets. Must be integer}

\item{year}{The year of analysis. Must be integer.}
}
\value{
a plot of the accidents based on the \code{state.num} and
\code{year} inputs.
}
\description{
This function plots the accidents in a simple map. The state number should be
integer or numerical and should exist in the FARS data, otherwise the
function terminates with an error. Also returns an error if the data file for
the year input does not exist.
}
\examples{
\dontrun{
fars_map_state(45, 2015)

# Results in an error
fars_map_state(45, 2016)
fars_map_state(60, 2015)
}

}
