% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map state-wide injuries in given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer number representing a state}

\item{year}{An integer number for a given year}
}
\description{
This function shows a map that indicates injuries distribution in a state of given year.
If misinput the state number, error message will be printed.
If no accidents is reported in this state of given yerar, "no accidents to plot" will be
printed.
}
\examples{
\dontrun{
fars_map_state(17, 2015)
}

}
