% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_W4.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Import data of the years requested}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer indicating the state to plot}

\item{year}{The year to plot}
}
\value{
A plot of the state showing the accidents
}
\description{
This function accepts a vector of years of the data requested and
summaries the number of accidents by month and year.
}
\note{
An error will occur if the data does not contain the state requested.
}
\examples{
\dontrun{
  fars_map_state(1,2015)
}

}
