% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Map of fatalities by location}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number}

\item{year}{The year to use the FARS data}
}
\value{
a map object
}
\description{
Plot the location of fatalities on a state map
}
\details{
The function checks for a valid state.num. Then it filters the accidents for the state. If there aren no
accidents, then a message is provided.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}
}
