% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents on a state map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer, or an object that can be coerced into integer,
representing a state number}

\item{year}{an integer, or an object that can be coerced into integer,
representing a calendar year}
}
\value{
a plot of a state accidents map
}
\description{
This function takes a state number and year as its argumants, and returns a
map of all all accidents in that state and year. The function requres FARS
raw datsets saved in \code{getwd()}.
}
\details{
An invalid state number throws an error. If no accides found in that state,
the function returns \code{NULL} silently. The function uses imported
\code{maps::map()} tp plot the data.
}
\examples{
my_file <- system.file("extdata", "fars_data.zip", package = "fars")
unzip(my_file, junkpaths = TRUE)
fars_map_state(16, 2015)


}
