% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show accidents on the map of a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An atomic numeric value that can be coerced to a two-digit integer
representing the state's code. If there is no states matching such a code,
an exception is thrown.}

\item{year}{An atomic numeric value convertible to a four-digit integer
representing a year to process.}
}
\value{
\code{NULL}
}
\description{
The function shows on the map the accidents that happened during a year
on the territory of the state.
It uses \code{LATITUDE} and \code{LONGITUDE} registered for each accident in FARS
to plot it on the state's map.
For more information about FARS data, see
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}.
}
\examples{
fars_map_state(2, year = 2015)
fars_map_state(state.num = 1, year = 2013)

}
