% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize accident data on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character or numeric specifing the state code.}

\item{year}{A character or numeric specifing the year.}
}
\value{
This function returns a map of the selected state with a point drawn
   on the map for each accident observation
}
\description{
This is a function that vizualizes accident data on a map of the selected state.
}
\examples{
fars_map_state(1, 2014)
fars_map_state("28", "2015")

}
