% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map showing all accidents in a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer}

\item{year}{an integer}
}
\value{
This function returns a graphical object containing locations of accidents.
}
\description{
This function plots a map with all the accident recorded in the given state and year, when the state number and year are provided.
}
\details{
If the provided state number does not exist in the data for the given year, the function
returns an error message.

If there are no accidents in a given state and year, the function returns an error message.
}
\examples{
\dontrun{
fars_map_state(16, 2017)
}

}
