% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents happened in a certain State and year}
\usage{
fars_map_state(state.num, year, demo = FALSE)
}
\arguments{
\item{state.num}{a number/character string for the State you are looking for}

\item{year}{a number/character string for the year}

\item{demo}{logical if you want to use the function with demo data that comes with the package.
Default to FALSE}
}
\value{
This function plots a map with points where accidents happened.
 Nothing is returned in the environment
}
\description{
This function reads the correspondent file for the specified year
checks the specified State number and if it is present outputs the State map
and where the accidents have happened.
Files should be in working directory.
}
\section{Warning}{

 The function checks for valid \code{state.num} otherwise it throws an error.
 The function doesn't accept more than one state number or year.
}

\examples{
 \dontrun{fars_map_state("1","2013")
fars_map_state("3","2014") #this shows the error}

}
