% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number code}

\item{year}{Year of data file}
}
\value{
plot of map

if state number or file corresponding to year does not exist
}
\description{
Create a state map
}
\examples{
\dontrun{
fars_map_state(1,2013)
fars_map_state(4,2015)
}
}
