% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Return a map plot of accidents by State and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number}

\item{year}{The valid year of recorded data}
}
\value{
a map plot of all accidents in the state in that year.
}
\description{
This function takes a state number ID and a year and output a map plot shows all
the accidents in that year.
}
\examples{
\dontrun{fars_map_state(50, 2014)}
}
\seealso{
\code{\link{make_filename}} , \code{\link{fars_read}}
}
