% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map of Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer giving the number of the state the function will display.}

\item{year}{An integer giving the year of which the function will display the map of accidents.}
}
\value{
This function returns a graphical object. In case of an invalid state number an error
message is displayed. A warning tells you if there are no accidents to plot.
}
\description{
Create a map of fars events for given state and year.
}
