% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{\code{fars_map_state} takes in a specific state and specific year and plots accident points according to latitude and longtitude on a state map.
If the input state is invalid, displays error message, "invalid STATE number: ". If there is no accident in that state, displays message,
"no accident to plot".}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number specified}

\item{year}{year specified}
}
\value{
a date.frame of filtered data for a particular state and year; if there are no rows, it returns invisible(NULL)
}
\description{
\code{fars_map_state} takes in a specific state and specific year and plots accident points according to latitude and longtitude on a state map.
If the input state is invalid, displays error message, "invalid STATE number: ". If there is no accident in that state, displays message,
"no accident to plot".
}
\examples{
\dontrun{
fars_map_state(1,2011)
}
}
