% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents location for a given state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer specifying state}

\item{year}{An integer specifying year}
}
\value{
This function doesn't return anything. The side effect is a plot of accidents location.
}
\description{
This is a function that, given a state number and a year, plots accidents locations
on a map. There are 51 states, entering \code{state.num} larger than 51 (or lower than 1)
will generate an error
\code{fars_map_state} is dependent on \code{dplyr} and \code{maps} packages.
}
\examples{
\dontrun{fars_map_state(state.num = 2, year=2013)}

}
