% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the numbers stand for state}

\item{year}{a numeric value}
}
\value{
A map plot for input state and year will be produced.
}
\description{
plot accidents by lattitude and longitude for state and year inputed by users.
}
\details{
the input state.num should be in the STATE column of data, otherwise
error message "invalid STATE number" will be given; If there is no accident
for the input state.number and year, a message will state "no accidents to plot".
}
