% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show map of accident number by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer of state codes}

\item{year}{an integer or string with input year}
}
\value{
map graphics
}
\description{
generate a map of fatal accidents for a given state and year. An error is raised for invalid state numbers.
}
\examples{
\dontrun{fars_map_state(1,2013)}

}
