% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Title}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state.num state number to plot accidents for}

\item{year}{year of data to plot accidents that occured}
}
\value{
a data.frame of filtered data, if there are no rows then returns invisible(NULL)
}
\description{
Title
}
\examples{
 \dontrun{
  fars_map_state(1, 2013)
}

}
