% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the locations of accidents happend in a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer that represents a state}

\item{year}{An integer of 4 digits that represents a year}
}
\description{
This function plots a map with dots representing the locations of
accidents happened in a given state and year.
}
\note{
This function will throw an error when the input \code{state.num} value
is not valid.
}
\examples{
\dontrun{
fars_map_state(4, 2013)
}

}
