% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state Function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{The}{code number of a state and the year}
}
\value{
a plot that depicts a map of the state
}
\description{
This is a function that takes as input the state you are interested in and the state you want to study
and gives as output a map with dots in all the places were a fatal accident took place.
}
\note{
A message appears when an invalid state number is given when there are no accidents.
}
