% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS data by state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric number from 1 to 56 that specifies the state to plot on the map}

\item{year}{numeric number that specifies the year to read the data for}
}
\value{
This function returns a plot
}
\description{
This is a function that reads FARS data for specified \code{year}
and then plots it on a contour map of the  state specified by \code{state.num}
}
\details{
Map of the state with dots of casualities will be ploted
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{filter}} in the \code{dplyr} package.
  \item \code{\link[maps]{map}} in the \code{maps} package.
  \item \code{\link[graphics]{points}} in the \code{graphics} package.
}
}

\examples{
\dontrun{
fars_map_state(1, 2014)
fars_map_state("3", "2014")
}

}
