% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_fars_fun.r
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number identifier for state of interest}

\item{year}{A year of data to graph locations of accidents.}
}
\value{
fars_map_state returns a plot object.
}
\description{
This function expands on the previous functions by now including a state.num argument. The
Here, the data are read in, and then checked for the presence of the state, given by state.num.
If the state is not in the data, the function stops. If state is present, data are filtered to
this state only and checked for number of accidents that occurred. If none, message is printed,
but if there is data, then the locations by latitude and longitude are plotted.
}
\note{
returns an error if state number is invalid.
}
