% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Create tibble data from csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string for filename along with its path}
}
\value{
This function returns tibble dataframe.
}
\description{
This function reads file given and converts to tibble format
}
\details{
It will give error if file is not present
}
\examples{
\dontrun{
fars_read("C://abc/file2.csv")
fars_read("file1.csv")
}

}
