% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to be read as a character string}
}
\value{
This function returns a tibble of the filename.
}
\description{
This is a function that will read in a file if it exist and return a tibble (data frame tbl). This function needs to import read_csv
from the readr package as well as tbl_df from the dplyr package.
}
\examples{
fars_read("accident_2013.csv")

}
