% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Building R Packages: Week 2 Peer-Graded Assignment  Adrian Accurso 5/25/2022}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the file to read, see
notes.}
}
\value{
A data frame with data readed from the csv file, or an error if the
file does not exists.
}
\description{
The assignment will be graded based on how closely the documentation reflects
the actual functioning of the code presented in the script file. In particular,
you will be expected to document:
-what each function does, in general terms;
-the function arguments (inputs);
-each function's return value;
-conditions that may result in an error;
-functions that need to be imported from external packages;
-examples of each function's usage.
############################################################################
fars_read: Read a file with FARS (Fatality Analysis Reporting System) data.
Create exception if file does not exist, otherwise changes the file contents
from a csv to a tibble
}
\details{
For more information, see:
\itemize{
\item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
\item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
\dontrun{library(dplyr)
library(readr)
yr <- 2013
data <- yr \%>\%
  make_filename \%>\%
  fars_read
head(data)}
}
\seealso{
\link{make_filename}
}
