% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Building R Packages: Week 2 Peer-Graded Assignment  Adrian Accurso 5/25/2022}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string representing a file name}
}
\value{
A tibble with fars data from the file named 'filename', or an error
if the filepath is incorrect or if the file is not in csv format.
}
\description{
The assignment will be graded based on how closely the documentation reflects
the actual functioning of the code presented in the script file. In particular,
you will be expected to document:
-what each function does, in general terms;
-the function arguments (inputs);
-each function's return value;
-conditions that may result in an error;
-functions that need to be imported from external packages;
-examples of each function's usage.
############################################################################
fars_read: Read a file with FARS (Fatality Analysis Reporting System) data.
Create exception if file does not exist, otherwise changes the file contents
from a csv to a tibble
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
\dontrun{x<-fars_read('farsFile.csv')}
}
