% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{fars_read
fars read first check if the filename provided exists, then loads the file suppresing all messages 
and  transfor the data into a tibble.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path and name of the file to be downloaded and treated}
}
\value{
If the file exists it returns a tible, otherwise it informs that the filename does not exists
}
\description{
fars_read
fars read first check if the filename provided exists, then loads the file suppresing all messages 
and  transfor the data into a tibble.
}
\examples{
\dontrun{
data<-fars_read("inst/extdata/accident_2013.csv.bz2")
}
}
