% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Fast csv data Reading}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(character) correspond to the full name.}
}
\value{
A object of the tbl_df class.
}
\description{
This is a simple function for reading csv
}
\details{
In case the file does not exist the file a message warning appear
}

