% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Loads a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{A}{character string providing the name of the csv file to load}
}
\value{
The function returns a tibble (data.frame) from the CSV file loaded
}
\description{
This function reads in raw data from a csv file
}
\details{
Will return an error message if the file does not exist
}
\examples{
\dontrun{
accident_2014 <- fars_read("./data/accident_2014.csv.bz2")
accident_2015 <- fars_read("./data/accident_2015.csv.bz2")
}
}
