% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{fars_read}
\alias{fars_read}
\title{Read the Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string with a file name and filepath to the file
containing data}
}
\value{
if the file is succesfully loaded it returns \code{data.frame} object
   with a given data, otherwise is "file does not exists" occur.
}
\description{
The \code{fars_read} function takes only one argument \code{filename}
the name of the csv file that contains the Fatality Analysis Reporting
System data \url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}.
It uses \code{readr::read_csv} to load data so other formats are also supported.
}
\examples{
\dontrun{fars_read("data.csv")
fars_read("accident_2013.csv.bz2")}

}
