% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Print "Fars Read"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the full path and filename to the accident data.}
}
\value{
This function returns a data.frame from a delimited file using readr library. Future versions will return a tibble.
}
\description{
This function reads a CSV into a data.frame by passing in the full path
and file name like (\code{filename = "accident_2015.csv.bz2"}).
}
\note{
Incorrect path to filename will result in the message [File Name] does not exist.
}
\examples{
\dontrun{
fars_read("/data/accident_2015.csv.bz2")
fars_read(filename = "/data/accident_2015.csv.bz2")
}

}
