% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a delimited file into a tibble object.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the compressed file you want to read in (a string value).
You can provide the full file path rather than setting the working directory before running this function.}
}
\value{
A tibble object with the contents of the filename.
}
\description{
This function reads the contents of a compressed file in .bz2
format and returns them in a
\href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble} object.
}
\section{Warning}{

   If the file does not exist, the function will show an error.
}

\examples{
\dontrun{
fars_read("YourComputer/FolderWithData/accident_2013.csv.bz2")}
}
