% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load CSV file into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path and name of the CSV file}
}
\value{
This function returns a tibble (data.frame) of the input file
}
\description{
This function reads CSV file defined by \code{filename}, a main argument of the function,
and returns a tibble (data.frame).
}
\note{
This function will give error if the file path is incorrect or file does not exist
}
\examples{
fars_read("./data/accident_2013.csv.bz2")
accident_2014 <- fars_read("./data/accident_2014.csv.bz2")

}

