% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string giving the name of a file in table format}
}
\value{
Function \code{fars_read} returns a data frame, with with cases corresponding to
lines and variables to fields in the file.
}
\description{
Functions \code{fars_read} reads a file with \code{filename} in table format
to create the data frame.
}
\note{
If the file with \code{filename} does not exist,
the error message would be displayed.
}
\examples{
far_read("accident_2013.csv.bz2")
far_read("accident_2014.csv.bz2")

}
