% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a delimited file into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string to be supplied to function to read and it
does not have default value.}
}
\value{
This function returns a tibble.
}
\description{
This function reads a comma separated value file data.
}
\details{
it first checks whether the file exists in the current working directory. If
 the file doesn't exist it throws an error, otherwise it reads the file and transforms
 it into a tibble.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}

}
