% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to the file containing the FARS data}
}
\value{
This function returns a tibble of FARS data
}
\description{
Read in FARS data
}
\note{
error returned if file does not exist
}
\examples{
\dontrun file<-make_filename(2014); fars_2014<-fars_read(file)

}
