% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from the "Fatality Analysis Reporting System (FARS)"
This is a function that reads in data from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System. It checks if the the data file exists and if so reads in the csv data,
(using the \code{fars_read} argument). If the file does not exists, it prints "file "filename" does not exist".
The data is turned into a dataframe table. The function requires functions from the packages readr and dyplr.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file that is read in.}
}
\value{
This function reads in csv data, if the specified filename exists in the working directory,
   and turns it into a tibble data. If the filename does not exist, "file "filename" does not exists" is printed.
}
\description{
Read data from the "Fatality Analysis Reporting System (FARS)"
This is a function that reads in data from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System. It checks if the the data file exists and if so reads in the csv data,
(using the \code{fars_read} argument). If the file does not exists, it prints "file "filename" does not exist".
The data is turned into a dataframe table. The function requires functions from the packages readr and dyplr.
}
\examples{
\dontrun{fars_read("accident_2013.csv")}

}
