% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Loads a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the CSV file (character)}
}
\value{
Returns a tibble (data.frame) of the CSV file.
}
\description{
Loads as argument a CSV file \code{filename} and returns
a tibble. In case of incorrect path definition provides an error message and ends.
}
\examples{
\dontrun{
accident_2015 <- fars_read("./data/accident_2015.csv.bz2")
}

}
