% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Fars Read: Reading a Fatality Traffic dataset}
\usage{
\code{"fars_read(filename)"}
}
\arguments{
\item{filename}{A file with csv extension}
}
\value{
Return a tidy tibble dataset
}
\description{
This function return a tidy tible dataset of from the US National
             Highway Traffic Safety Administration's Fatality Analysis Reporting System in a zip format
}
\details{
The function will return a message of "File not exists" if the file requested
         in the argument does not exist
}
