% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fars_read}
\alias{fars_read}
\title{Create a dataframe table from the given csv filename.
Should the filename not exist, then, an error is raised.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename:}{The file to read from.}
}
\value{
dplyr tbl_df instance of the file content.
}
\description{
Create a dataframe table from the given csv filename.
Should the filename not exist, then, an error is raised.
}
\note{
No progress bar will be shown while reading the file.
}
\examples{
csv <- fars_read('avalid_filename.csv')
> class(csv)
[1] "tbl_df"     "tbl"        "data.frame"
invalid <- fars_read('does_not_exist.csv')
Error in fars_read("does_not_exist.csv") :
 file 'does_not_exist.csv' does not exist
}
