% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string with the path of the .csv file to read.}
}
\value{
A tibble containing the data, or an error if the file does not exist
}
\description{
This function reads a .csv file providing the \emph{American public yearly
data regarding fatal injuries suffered in motor vehicle traffic crashes}, and
returns a tibble.
}
\examples{
\dontrun{

library(dplyr)
library(readr)
fars_read("accident_2015.csv.bz2")

}

}
