% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a CSV file with FARS data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to FARS CSV file}
}
\value{
A tbl_df of the data in the CSV file for which the path was provided. If the path was invalid, execution
   will stop
}
\description{
Read Fatality Analysis Reporting System (FARS) CSV file
}
\examples{
\dontrun{
accident_2013 <- fars_read("data/accident_2013.csv.bz2")

# stops executing
accident_fail <- fars_read("data/this.file.does.not.exist.csv.bz2")
}
}
