% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read in file with data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A \code{string} with filename to be read}
}
\value{
A dataframe created from csv file or will return an error if file does not exist
}
\description{
Function takes in a csv from the \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}
called a FARS file for a user-specified year.
}
\examples{
\dontrun{
fars_read("accident_2015.csv")
}
}
