% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a Fatality Analysis Reporting System dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string corresponding to the path of the file}
}
\value{
data frame tbl
}
\description{
Reads a dataset from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census providing the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\note{
stop the execution and print an error message if the file does not exists
}
\examples{
\dontrun{
data2013 <- fars_read("fars_data/data/accident_2013.csv.bz2")
}

}
