% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read - Read FARS .csv file into a data frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of data file as a character string}
}
\value{
A data frame containing the data read from the csv file, or an error
        if the file does not exists.
}
\description{
This function reads a (compressed) csv file containing data from the US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System into
a data frame. See also:
\itemize{
  \item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
  \item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
library(dplyr)
library(readr)
yr <- 2015
data <- yr \%>\%
  make_filename \%>\%
  fars_read
head(data)
}
