% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading FARS csv data files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the path and the file name. For example
"./A/B/C/xyz.csv.bz2"}
}
\value{
This function returns a data frame.
}
\description{
fars_read is used to read FARS (Fatality Analysis Reporting System) csv files.
}
\details{
Converts a FARS bz2 file into a dataframe
}
\note{
Error Message: If a wrong file name is entered then an error is thrown
}
\examples{
t <- system.file("extdata", "accident_2013.csv.bz2", package="fars")
fars_read(t)

}

