% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function \code{fars_read}}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to load}
}
\value{
A data.frame object
}
\description{
This function reads the contents of a csv file.
If the file can not be found, an error message is returned
}
\note{
This is an internal function not to be used from outside the package
}
