% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{FARS Data Input.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename of the csv file to be
read}
}
\value{
returns a data frame table if the file exists, throws an error if the file
  doesn't exist.
}
\description{
This functions reads a file in table format, and returns it as an object of the "tbl_df" class. If
     the file does not exist in the current directory, this function stops
     execution with an error message. While reading all messages are suppressed.
}
\details{
This function uses the \code{read_csv()} function from the
         \code{readr} package, with \code{progress = FALSE} parameter.
         Then the \code{tbl_df()}, function from the \code{dplyr} package
         is called to convert the data into the tabl_df format before returing it.
}
\section{Depends on}{

\enumerate{
  \item \code{\link[readr]{read_csv}} in the \code{readr} package.
  \item \code{\link[dplyr]{tbl_df}} in the \code{dplyr} package.
}
}

\examples{
\dontrun{
  fars_read("accident_2014.csv.bz2")
}

}
