% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name along with the file path for the data}
}
\value{
This function returns a data table of the class tbl_df
}
\description{
This is a simple function that reads in FARS data files. If the
file doesn't exist an error message will be returned:
Error in fars_read("test.csv") : file 'test.csv' does not exist
}

