% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Import fars file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The directory to the file to be imported. Data dictionary can be obtained as:
\itemize{
    \item{\code{system.file("extdata", "accident_2013.csv.bz2", package = "FARS")}}
    \item{\code{system.file("extdata", "accident_2014.csv.bz2", package = "FARS")}}
    \item{\code{system.file("extdata", "accident_2015.csv.bz2", package = "FARS")}}
}}
}
\value{
A data frame with data stored from the csv file, or an error if the file does not exist.
}
\description{
\code{fars_read} import the fars file with the directory as argument.
}
\examples{
require(readr)
require(dplyr)

dat <- year \%>\%
    make_filename \%>\%
    fars_read

}
\seealso{
\code{\link{make_filename}}
}
