% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{read in csv data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with a route to the csv file that needs to be read}
}
\value{
This function return a tibble object of the csv file data
}
\description{
This is a simple function that checks for existence of a csv file and . If exists, read in csv file by its name. If not, throw an erro
}
\examples{
\dontrun{fars_read(system.file("extdata", "accident_2013.csv.bz2", package = "fars"))}

}
