% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read - reading data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that specific the file to read}
}
\value{
This function returns the data file as a \code{\link{dplyr}} \code{\link{tbl_df}}.
}
\description{
This is a function that read the dedicated file in CSV format and output
as a \code{\link{dplyr}} \code{\link{tbl_df}}.
}
\note{
Error when file does not exist
}
\examples{
myfile <- "accident_2014.csv.gz2"
\dontrun{dataset <- fars_read(myfile)}
\dontrun{dataset <- fars_read("accident_2013.csv.gz2")}

}
