% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Return a data frame representing the data in the csv file pointed to by the \code{filename} given as input.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character vector representing the filename of a csv file to be read in.}
}
\value{
A data frame representing the data from the \code{filename} given as input.
}
\description{
Return a data frame representing the data in the csv file pointed to by the \code{filename} given as input.
}
\note{
The \code{filename} must exist or an error is raised.
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")
}
}
