% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in csv-File as dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to the file to read in}
}
\value{
This function returns the data from the given file as a dataframe
}
\description{
This function checks if a given csv-file exists and if so reads it in silently as a dataframe
}
\note{
Errors can result from the given filename not being a valid path to an existing file or the file not being encoded in the right format
}
\examples{
fars_read('my_csv.csv')
fars_read('/Users/User/my_csv.csv')

}
