% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{FARS Dataframe Creator}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing a CSV file path that will be
checked and used to create a data frame. If the file is in the same
directory as the R project, it could simply be a CSV file name.}
}
\value{
A data frame with the contents of the CSV for which a file path was
 given.
}
\description{
This function checks for the presence of a file, and outputs
 a data frame object of the data if the file exists. If the file
 does not exist it will output an error message.
It uses functions from the 'readr' and 'dplyr' packages.
}
\note{
Inputting a file path improperly, or a file path to a file that does not
 exist, will result in an error.
}
\examples{
\dontrun{fars_read("myfile.csv")}
\dontrun{fars_read("C:/Users/Username/Documents/directoryfolder/myfile.csv")}

}
