% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read  in a file.
This function will check to see if a file exists based on the file path.}
\usage{
fars_read(filename, filepath = system.file("extdata", filename, package =
  "fars"))
}
\arguments{
\item{filename}{The name of the file to read.}

\item{filepath}{a character string with the directory of the package's data files.}
}
\value{
This file returns the data in the csv file as a tibble.
}
\description{
Read  in a file.
This function will check to see if a file exists based on the file path.
}
\examples{
\dontrun{
#' fars_read('accident_2013.csv.bz2') #only if in the same directory as data
fileName <- system.file("extdata","accident_2013.csv.bz2",package="fars") #when using provided data in package
fars_read(fileName)
}
}
