% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads the data from a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to process. The file should exist and should
be a CSV data file.}
}
\value{
This function returns a data frame containing the data of the file processed.
}
\description{
This is a simple function that, reads the data from a file specified by the
 \code{filename} argument. The functions stop in case of no files found.
 The files should contain a CSV formatted data. The data read is returned
 as a data frame.
}
\examples{
\dontrun{
fars_read(filename="accident_2014.csv.bz2")
}

}
