% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Display file content in tabular format}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file you want to read in and display.}
}
\value{
This function dsiplays a data frame in tabular format.
}
\description{
This function takes a file name and if it does not exist,
it exits with an error message, saying the file does not exist.
Otherwise, it reads in the file sing the read_csv function from readr
and then displays it in tabular format as a data frame.
The function could be made better because the file itself may not be a csv file.
}
