% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read the Fatality Analysis Report Data}
\source{
The source data is hosted on https://d3c33hcgiwev3.cloudfront.net
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A csv file that contains the Fatality Analysis report from the US National Highway Safety Administration}
}
\value{
This function returns a dataframe with Fatality Analysis report
}
\description{
Read the Fatality Analysis Report Data
}
\details{
This function raise errors when external library is not installed or invoked using library() function
}
\examples{
fars_data <- fars_read("accident_2013.csv.bz2")

}
