% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path of the file you want to read in. Can be a csv or zipped csv.}
}
\value{
A tibble containing your dataset.
}
\description{
Read in the desired data file from FARS.
}
\section{Errors}{

This function will throw an error if:
\itemize{
 \item \code{filename} is not a valid path or does not exist.
 \item \code{filename} is not a valid input type.}
}

