% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file}
}
\value{
a tibble
}
\description{
Read a csv file and convert it into a tibble
}
\note{
if file did not exit, the prgram will exit with an error
}
\examples{
fars_read("accident_2015.csv.bz2")

}

