% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function for reading your csv files.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{file to be read}
}
\value{
This function will return the loaded data in the form of a tibble.
}
\description{
This function reads your csv file (can be in bz2 zipped format also). If the filename you want to open does not exist,
function returns error message and stops.
}
\details{
filename you are giving to the function fars_read (or the path) must be
correct and the file must exist. Otherwise the function returns error and stops.
}
\examples{
\dontrun{
fars_read(system.file("extdata", "accident_2013.csv.bz2", package = "fars"))
}

}
