% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{print("fars_read")
This is a function that reads a file containing data from the US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System
(FARS). A filename must be provided as an argument to the function, which
can be done using the \code{filename} argument. If the file does not exists,
it generates an error message.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from.}
}
\value{
This function returns a tibble containing the data from the file.
}
\description{
print("fars_read")
This is a function that reads a file containing data from the US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System
(FARS). A filename must be provided as an argument to the function, which
can be done using the \code{filename} argument. If the file does not exists,
it generates an error message.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
fars_read(filename = "accident_2013.csv.bz2")
}

}
