% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System (FARS) data into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string giving either the name of or the path to a
file containing FARS data}
}
\value{
This function returns a tibble data frame.
}
\description{
The function takes a file name or a file path pointing to a FARS data set ,
checks that the file exists and then reads it into a data frame format. This
is a lightweight wrapper around \code{readr::read_csv()}.
}
\note{
An error is thrown if the file indicated does not exist.
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
}
}
