% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read motor vehicle crash csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path, name, and extension of a file to read in.}
}
\value{
This function returns a data frame table containing data for motor vehicle crashes for a particular year if the input csv file exists.
		An error message is returned if the file does not exist.
}
\description{
This function (fars_read) validates that an input file exists.
An error message is returned if the file does not exist.
If the file exists then the file is read using the read_csv function from the readr package (\code{\link[readr]{read_csv}}).
A data frame table is created using the tbl_df function from the dplyr package (\code{\link[dplyr]{tbl_df}}).
}
\examples{
\dontrun{
fars_read("D:/Data/accident_2013.csv.bz2")
fars_read(filename="D:/Data/accident_2013.csv.bz2")
}

}
