% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading a single FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string indicating the name of the file to be read. This
may be an absolute or relative file path, as well as a short file name,
in which case the search for the file will occur in the current working
directory.}
}
\value{
The function returns a tibble containing the data in the file. If
the file indicated by \code{filename} doesn't exist, an error is generated.
}
\description{
This function takes one argument, \code{filename}, and, if the corresponding
file exists, reads it using the \code{read_csv} function from the
\code{readr} package, suppressing any messages (as well as progress
indication) generated by this function. The data in the file is then
converted to a tibble, using the \code{tbl_df} function from the
 \code{dplyr} package.
}
\examples{
\dontrun{fars_read("mydata.csv")}
\dontrun{fars_read("C:/Users/JaneDoe/R/ExampleData/mydata.csv")}
\dontrun{fars_read("~/Examples/mydata.zip")}

}
