% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read() - read accidents statistics by the given file name}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{- character string with file name}
}
\value{
data frame as dplyr::tbl_df object. If file does not exist print error message.
}
\description{
open file with accidents statistics by the given file name and read it into dplyr::tbl_df variable.
}
\examples{
name <- system.file("extdata", sprintf("accident_\%d.csv.bz2", 2014), package = "Fars")
data <- fars_read(name)


}
