% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Title A simple function for reading files or path from the system into the package}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the name of the file or path to be read. An errors occurs if filename does not exist}
}
\value{
A dataframe from \code{\link{tbl_df}}
}
\description{
Title A simple function for reading files or path from the system into the package
}
\examples{
fars_read("accident_2013.csv.bz2")
}
