% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv-data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Csv file containing data}
}
\value{
This function will search within the specified path for the
  filename provided. If the file exists, it will be imported and returned as
  a data frame tbl.  If it does not exist an error message will be returned.
}
\description{
This is a function that reads in csv data .
}
\details{
\code{fars_read} reads in csv  data.
}
\examples{
\dontrun{
data <- fars_read(filename = "./accident_2013.csv.bz2")

file<-make_filename(year=2013)
data<-fars_read(filename = file)
}

}
