% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function to read fras files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to file}
}
\value{
The dataframe generated
        Error if file doesn't exist
}
\description{
Function to read FARS file.

This function uses readr and dplyr packages.
It takes a CSV file as an input filename parameter.
If the file does not exist in the working directory, this function stops with an output message that the file does not exist.
If the file exists, this function reads it into dataframe object.
It suppresses messages and progress bar while reading the file.

Function to load a CSV dataset into a dataframe
}
\details{
Dataset must accomplish the FARS structure
}

