% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from a file}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string providing the location of the file to be read.}
}
\value{
The function returns a tbl_df, tbl, data.frame object.
}
\description{
This function extracts data from the users current work directory for data in csv format.
The user must provide file location which is
the only agrument of this function. The data is extracted using the readr package.
Once the data is read into R the dplyr package alters the look of the data frame
using the tbl_df function.
}
\details{
an error will be thrown if the filename does not exist, or is not in the
current working directory.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
}
