% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{loads or "reads" in fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{The}{name of the file to be loaded.}
}
\value{
dataframe containing the selected file's data.
}
\description{
\code{fars_read} loads data from csv.
}
\details{
This function loads data from a comma separated file taken from the  US
 National Highway Traffic Safetyc Administration's Fatality Analysis
 Reporting System (FARS).
}
\note{
If the file or permission to it does not exist, an error is produced.
}
\examples{

\dontrun{
fars_read(filename = full_filename)
}

\dontrun{
fars_read(filename = 'filedoesnotexist')
}

}
