% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read one year dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A filename generated by the make_filename function.}
}
\value{
This function returns a dataframe in the year you choose.
}
\description{
This is a function that read the dataset in the year you choose. The data are
from the US National Highway Traffic Safety Administration's Fatality Analysis
Reporting System, which is a nationwide census providing the American public yearly data
regarding fatal injuries suffered in motor vehicle traffic crashes
}
\examples{
\dontrun{filename=make_filename(2013)}
\dontrun{df=fars_read(filename)}
\dontrun{head(df)}

}
