% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read and Import File into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of your file.}
}
\value{
It returns a data frame
}
\description{
Returns a tibble data frame after loading the file
}
\details{
If your file does not exist or you type the wrong name, the
          function will throw an error. If you are using relative path,
          the file need to be within relevant directory. Otherwise
          there will be an error
}
\examples{
fars_read("Happy.csv")
}
