% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Helper function to read a data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string with the exact name of
the file to read}
}
\value{
A tibble with the full data file contents
}
\description{
Checks file for existence in the current
directory and returns its content as a tibble.
Used by \code{fars_read_years} and \code{fars_map_state}.
}
