% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Title: fars_functions
fars_read function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file name and path}
}
\value{
this function returns a data frame. If file name or path does
not exist, it will give a warning message.
}
\description{
This is a function that read in existing data files with readr package
without showing progress or printing any messages
}
\examples{
\dontrun{fars_read("accident_2013.csv")
fars_read("C:\mydata\accident_2013.csv")
fars_read("http:\\\\abc.org\accident_2013.csv")}

}
