% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{This is a assignment to Coursera Building R Packages week2.
fars_function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the filename to read.}
}
\value{
This function returns a data frame from csv file, or an error with
        wrong filename.
}
\description{
The functions will be using data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System, which is a
nationwide census providing the American public yearly data regarding fatal
injuries suffered in motor vehicle traffic crashes.
}
\examples{
fars_read(filename = 'accident_2013.csv.bz2')

}
