% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV files on accident data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of file you want to read}
}
\value{
A tibble data frame
}
\description{
A Helper function to read CSV files as tibble data frame, should be used with other function.
If the file does not exist, it will throw out error message
}
\examples{
 \dontrun{dat <- fars_read("accident_2014.csv.bz2")}
}
\seealso{
\code{\link[=make_filename]{make_filename()}} to easily make the file name
}
