% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of the file.}
}
\value{
The function returns a table of classes "tbl_df", "tbl", "data.frame".
}
\description{
The function reads a file.
}
\details{
The function checks if the file exists. if it exists,
it reads it (also suppresses any messages that might be printed),
if it doesn't exist, it throws an error message.
}
\examples{
data(accident_2013)
write.csv(accident_2013,"test.csv",row.names=FALSE)
fars_read("test.csv")
}
