% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file into R}
\source{
\url{https://www.coursera.org/learn/r-packages/peer/25XGp/documenting-code}
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename.csv}{The name of the csv file to load in R}
}
\value{
a data frame tbl 
Returns \code{file filename does not exits} if the file can not be found or the directory is
not properly set
}
\description{
This function loads a csv file in R and converts it to a data frame tbl
}
\section{Warnings}{

This function can only handle csv files from 
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\examples{
\dontrun{  
fars_read("accident_2013.csv.bz2")
}
}
\author{
Bella V. (Documentation only)
}
\references{
Peer Assignment in Coursera Course "Building R Packages"
}

