% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in FARS Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The file path and/or file name of the csv file with FARS data
to read into R}
}
\value{
This function returns a tibble data frame containing FARS data, as
identified by the user.
}
\description{
This is a simple function that reads in csv data and coerces it to a tibble
data frame object.
}
\examples{

\dontrun{
fars_read(filename = "accident_215.csv")
}

}
