% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading CSV into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name filename}
}
\value{
Returns a tibble data frame
}
\description{
Reading CSV into R
}
\details{
This function reads a csv file and returns a tibble data frame. If the path does not exist it will produce an error message.
}
\examples{
\dontrun{fars_read("accident_2013.csv")}
}
\author{
James Bertschy \email{james.bertschy@gmail.ch}
}
