% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This is a function that reads a data file and displays it as a dataframe.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{is a character string giving the name of the data file}
}
\value{
This function returns the output of the data file
}
\description{
This is a function that reads a data file and displays it as a dataframe.
}
\details{
If the file doesn't exist it will result in an error
}
\examples{
\dontrun{fars_read("data/accident_2013.csv")}
}
