% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Reads a delimited csv file into a data frame table}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that indicates the file to read.}
}
\value{
A data frame table.
}
\description{
Reads a delimited csv file specified by \code{filenME}into a data frame table that wraps the local data frame.
}
\examples{
fars_read('Table.csv')
fars_read(filename = 'Table.csv')

}
