% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in data from the Fatality Analysis Reporting System (FARS).}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with a valid file name.}
}
\value{
This function returns a dataframe.
}
\description{
\code{fars_read} reads in the data from a csv file, returning a
dataframe
}
\details{
This function will raise an error if the file referenced
in \code{filename} doesn't exist.
}
\examples{
df <- fars_read(file)
head(df)
}
