% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in NHTSA FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string indicating the name of the *.csv file to be read in}
}
\value{
The fars_read function returns a tbl_df object
}
\description{
The fars_read function takes as an input the filename of a *.csv file.  It reads in the file
data as a data frame.
}
\details{
The fars_read function first checks whether the file exists and prints a
warning if the file cannot be found.  The function generates an error if the data
present within the file is invalid.
}
\examples{
\dontrun{
  fars_read("accident_2015.csv.bz2")
}

}
