% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import .csv flat files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the .csv file to be
imported.}
}
\value{
This function returns a tibble.
}
\description{
This function imports a .csv file. The parameter \code{filename} is meant to
be passed a string from the function \code{make_filename()}. This function is
used within the \code{fars_read_years()} function.
}
\details{
This function first checks to make sure the file exists, and then,
if it does exist, uses \code{readr::read_csv()} to import the file, suppressing
import messages and the progress bar in the process.  Finally, it uses
\code{dplyr::tbl_df()} to convert the imported file to be a tibble data frame.
}
\note{
if a .csv file matching the \code{filename} you supply does not exist,
the function will stop execution with message
"file \code{filename} does not exist."
}
\examples{
fars_read("accident_2015.csv.bz2")

}
