% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{path to file}
}
\value{
data.frame
}
\description{
Reads a data file given a path.
}
\note{
A non-existent file path will result in an error.
}
\examples{
if(interactive()){
   df <- fars_read("path_to_file")
}

}
