% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load packages}
\usage{
fars_read(filename)
}
\arguments{
\item{A}{CSV file that contain all the data}
}
\value{
A tibble (data.frame) dataset or an error message if the dataset
    does'nt exist!
}
\description{
This function is been used to load the required dataset(s) that is used
in other functions of the package
}
