% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(character)}
}
\value{
data frame tbl
}
\description{
Read Fatality Analysis Reporting System (FARS) data in a data data frame tbl.
This function is not exported.
}
\details{
This function throws an error message if the file does not exist.
}
\examples{
\dontrun{fars_read(system.file("extdata","accident_2013.csv.bz2",package="fars"))}
}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
