% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read files from Fatality Analysis Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A path to a csv file. Files ending in .gz, .bz2, .xz,
or .zip will be automatically uncompressed.}
}
\value{
A data frame with the values of the csv file. If type of
\code{filename} is not correct or the file does not exist, an error message
is generated and nothing is returned.
}
\description{
The function \code{fars_read} reads csv files from the US National Highway
Traffic Safety Administration's Fatality Analysis Reporting System and
creates a data frame out of the values.3
}
\examples{
setwd(system.file("extdata", package = "fars"))
fars_read("accident_2015.csv.bz2")

}
