% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character with the path to the fars file you would like to read in}
}
\value{
This function returns an R data frame with the data from the filename parameter
}
\description{
This function reads in a fars file and returns the data stored in an R data frame
}
\examples{

\dontrun{fars_data <- fars_read("accident_2013.csv.bz2")}
\dontrun{fars_data <- fars_read("accident_2014.csv.bz2")}
\dontrun{fars_data <- fars_read("accident_2015.csv.bz2")}

}
