% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function to read Fatality Analysis Reporting System (FARS) data files}
\source{
website \url{http://www-fars.nhtsa.dot.gov/Main/index.aspx}
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{CSV or CSV.BZ2 filename with path leading to it as Character.}
}
\value{
A tibble (tbl_df, tbl, data.frame) containing the FARS dataset
}
\description{
FARS is a nationwide census providing NHTSA, Congress and the American public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes
The data files are in CSV format, often compressed. Function could read CSV and BZ2 files as well.
It has 50 variables=columns (eg. STATE, CITY, COUNTY, DAY, MONTH, YEAR, FATALS etc.), and around 30k records.
}
\note{
Existence of the file is checked at runtime. Internally read_csv function is used, which is part of package "readr", and need to be installed beforehand.
Second package dependence is dplyr, which is used for formatting output data.frame.
}
\examples{
\dontrun{fars_file<-fars_read("path_to_your_file")}
\dontrun{summary(fars_file)}

}
\keyword{FARS,}
\keyword{NHTSA,}
\keyword{Road}
\keyword{safety}
\keyword{traffic}
