% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This is a function that creates a data.frame from a csv file (can be zipped .bz2) specified using \code{filename} argument.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the location of the csv file}
}
\value{
A data.frame object if the csv file exists. Otherwise the function stops with a "file does not exist" error.
}
\description{
This is a function that creates a data.frame from a csv file (can be zipped .bz2) specified using \code{filename} argument.
}
\examples{
\dontrun{
fars_read("data/accident_2013.csv")
fars_read(filename = "data/accident_2013.csv")
}

}
