% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function \code{fars_read} loads input data file into memory and transforms it
into dplyr tibble.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{file name in string format}
}
\value{
tbl_df data frame

If the input file is not found, function stops returning an error:
"Filename does not exist"
}
\description{
Function \code{fars_read} loads input data file into memory and transforms it
into dplyr tibble.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
