% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file}
\usage{
fars_read(filename)
}
\value{
This function returns a tibble.
}
\description{
This function reads a csv file, zipped or unzipped, and stores it in a tibble.
}
\note{
This functions throws an error if the file does not exist.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
