% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_pack.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{-  A character string defining the CSV file to be imported}
}
\value{
This function returns a dataframe of the CSV file being read
}
\description{
This is a simple function that reads aswirl CSV file creating returning a dataframe.  It
has specifically been designed to read the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (FARS).
The function is used throughout the set of fars_ functions
}
\note{
The function first checks for the existence of a file so will stop if the file is not found
}
\examples{
   fars_read(filename)
   fars_read(data/filename.csv)
   my_data<- fars_read(data/filename.csv)

}
\references{
The US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census, provides the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
For more information including data that can be downloaded:
  \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\seealso{
\link{fars_read_years}
 \link{fars_summarize_years}
 \link{fars_map_state}
}
