% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a data file and creates a dataframe with dplyr.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{input character vector}
}
\value{
dplyr::tbl_df by reading the whole file
}
\description{
Reads a data file and creates a dataframe with dplyr.
}
