% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read .csv file whic contains FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of either a path to a file or file.}
}
\value{
A data frame with data readed from the csv file, or an error if the
file does not exists.
}
\description{
This function reads data from .csv file. The data is from the \strong{US
National Highway Traffic Safety Administration's} \emph{Fatality Analysis
Reporting System (FARS)}. For more information enter hear:
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
\dontrun{
accident_2015 <- fars_read("./data/accident_2015.csv.bz2")
}

}
\seealso{
\link{make_filename}
}
