% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the file to be read}
}
\value{
This function returns the data as tbl_df(data)
}
\description{
This is a function that reads a file and transforms it on a data frame
}
\examples{
\dontrun{
fars_read()
}

}
