% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in a data set}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file}
}
\value{
This function returns a \code{tbl_df} object.
}
\description{
This function reads in data from a csv file.
}
\examples{
my_data <- fars_read(filename="accident_2013.csv.bz2")

}
