% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{filename of the data in csv archive}
}
\value{
returns a tibble containing the data from the csv
}
\description{
This function reads the data from a csv into a dataframe to then coerce it into a tibble.
}
\examples{
fars_read("data.csv")

}
