% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data downloaded from FARS}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to read}
}
\value{
The function returns a data.frame
}
\description{
Reads the data downloaded from
the \href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System}
into an R data.frame. A csv.bz2 file must exist in the working directory.
Otherwise the function throws an error.
}
\examples{
setwd("inst/extdata")
fars_read(accident_2013.csv.bz2)
fars_read(accident_2014.csv.bz2)
}

