% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in a FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file as a string}
}
\value{
Returns a tibble (also a data.frame)
}
\description{
This function reads in a file that has been downloaded from the US National Highway
Traffic Safety Administration's Fatality Analysis Reporting System (FARS).
}
\note{
If the file does not exist, or if filename incorrectly specified, this function will return an error.

This function imports \link{read_csv} and \link{tbl_df}
}
\examples{
fars_read("accident_2013.csv.bz2")
fars_read(make_filename(2015))

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
