% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to file}
}
\value{
A tibble data frame
}
\description{
Check if file exists and read it. If file doesn't exist it throw an error.
}
\examples{

df <- fars:::fars_read(fars:::make_filename(2015))
head(df)
}
