% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of file to be read.}
}
\value{
This function returns a table created from csv file.
 Returned object has following classes:tbl_df, tbl, data.frame.
}
\description{
This is a simple function that reads csv file. In case if file does not exist
appropriate message will be shown.
}
\examples{
fars_read("accident_2013.csv.bz2")
}

