% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read  file into a data frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{characters vectors, containing file names or paths}
}
\value{
This function returns a data frame.
}
\description{
This is a support function to read a delimited file into a data frame.
}
\section{\code{Error}{
}
If \code{filename} does not exist, the function will throw an error.
}
\examples{
fars_read("accient_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
fars_read("accident_2015.csv.bz2")
}

