% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving directory and name of the csv
file which is needed to be read}
}
\value{
This function returns a data.frame, if the filename does not exist
either for wrong name or wrong directory, returns error.
}
\description{
This function reads a csv file form a given directory, to a data table
data frame
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
\dontrun{fars_read("data/accident_2013.csv.bz2")}

}
