% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Lee un archivo de datos csv}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{nombre del archivo que se quiere leer}
}
\value{
Esta función regresa en la consola los datos leidos del archivo seleccionado
   por el usuario como data frame tbl
}
\description{
Esta funcion permite leer un archivo de datos csv y mostrar en la consola
}
\details{
Debe escribir el nombre del archivo junto con la extensión .csv y todo esto
entre comillas.
Si ingresa un nombre incorrecto me devolverá que el archivo no existe : does not exist
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
fars_read(filename= "accident_2013.csv")
}

}
