% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the input file}
}
\value{
This function returns a tibble (a data frame with class \code{tbl_df}.
If the file does not exist, a message error is produced instead.
}
\description{
This function can be used to import a dataset in .csv format. The file name
needs to be provided via the \code{filename} argument.
}
\examples{
fars_read("accident_2015.csv.bz2")

}
