% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from FARS}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The csv data file}
}
\value{
Outputs a tibble data.frame from csv file. The error occured if the file does not exist.
}
\description{
Reads the data sourced from Fatality Analysis Reporting System that provided from
the US National Highway Traffic Safety Administration
}
\examples{
\dontrun{
accident_2015 <- fars_read("data/accident_2015.csv.bz2")
}

}
