% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Print "fars_read"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{CSV file name with naming convention "accident_nnnn.csv.bz2",
where "nnnn" is a 4-digit year number}
}
\value{
This function returns a tibble with traffic crash data.
}
\description{
This function reads the content of a csv file
}
