% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{This function imports a specified csv file and returns a dataframe.  It also checks to see if the file exists.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The csv file, or file path, that is imported and manipulated into a dataframe.}
}
\value{
This function returns a dataframe containing the information from the imported csv file.
}
\description{
This function imports a specified csv file and returns a dataframe.  It also checks to see if the file exists.
}
\note{
if the file does not exist in the working directory an error message will display "file does not exist"
}
\examples{
fars_read("accident_2013.csv.bz2")
}
