% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector). See \code{\link[readr:read_csv]{readr::read_csv()}}}
}
\value{
If succesful, returns a tibble with the contents of the file. If the file
doesn't exist, an error message is displayed
}
\description{
This function takes as input a data set from the US National Highway Traffis Safety
Administration's
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Anlaysis Reporting System}
(FARS) and loads it into a data frame
}
