% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data from a single file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The filename of the data file to load}
}
\value{
This function returns the data frame of data loaded from the CSV.
}
\description{
This function reads data from the FARS CSV file (compressed or uncompressed) having the filename provided
and returns a data frame of the data. FARS is the US National Traffic Safety Commission's
\href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System}.
}
\note{
This function is usually used with the \code{\link{make_filename}} function.

This function checks that the given filename exists in the working directory,
and will exit, generating an error, if the file does not exist.
}
\examples{
\dontrun{fars_read(make_filename(2014))}

}

