% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV format files silently}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A path to the CSV file (A character string)}
}
\value{
At tibble object (data.frame) of the CSV file data
}
\description{
This function
1) checks for an existent CSV file and,
2) reads it silently.
}
\details{
If the file does not exists throws an error
}
\examples{
path <- system.file("extdata", "accident_2013.csv.bz2", package = "CourseraFARS")

fars_read(path)
}
