% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Loads csv file into memory}
\usage{
fars_read(filename)
}
\value{
A tbl_df (tibble) will be returned containing the
information in the csv file.
}
\description{
This function will import a csv file. If the file does
not exist, an error will be generated. If the file does
exist, it will be imported as a tbl_df, or tibble data frame.
}
\examples{
\dontrun{
foobar_data <- fars_read("c:/User/foobar/foobar.csv")
}

}
