% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file that the function will read and coerce into
a tibble.}
}
\value{
This function returns a tibble derived from the input CSV file.
If the input CSV file does not, an error is generated.
}
\description{
This function checks whether the CSV file exists. If the file exists, then it is read
and coerced into a tibble.
}
\note{
Examples: fars_read("data2013.csv"), fars_read("data2015.csv")

If the file "data2050.csv"" does not exist, attempting to read this file
will generate an error.
try(fars_read("data2050.csv"))
}
