% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data from a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the file(character string) from which to read the data.}
}
\value{
This function returns a tibble contianing the data.
}
\description{
This function first checks if the file exists. If it does, it reads the data and converts it into a tibble. If the file does not exist, it will throw an error.
}
\examples{
fars_read("filename")

}
