% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read file into a dplyr's data frame tbl}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file's name the function
will read}
}
\value{
This function returns a error message either if the file's name
does not exits or a data frame table representing the data of the file
}
\description{
This function reads a file by firstly checking whether a files with that name
exists and then converting in into a data frame table
}
\examples{
fars_read("./data/accident_2015.csv.bz2")
fars_read(filename = "~/accident_2015.csv.bz2")
fars_read("mistake_file_name")

}
