% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(\emph{character}) The file to be read.}
}
\value{
FIBS data is returned as tibble (see \code{\link[dplyr]{tbl_df}})
}
\description{
Read Fatality Analysis Reporting System (FARS) data from 
US National Highway Traffic Safety Administration. \cr 
\code{fars_read} reads a file defined by \code{filename} from the current 
working directory.
}
\details{
If \code{fars_read} cannot find the given filename in the current 
directory an error is thrown. It is recommended to use 
\code{\code{make_filename}} to create a filename for a given \code{year} 
according to the FARS filenaming scheme. \cr 
\code{\link{fars_read}} is a helper function for \code{\link{fars_read_years}}
}
\examples{
\dontrun{
year <- 2014 
df <- fars_read(make_filename(year))
}

}
\references{
US National Highway Traffic Safety Administration \cr
(\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\seealso{
\code{\link{make_filename}}, \code{\link{fars_read_years}}
}
