% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Dataset name as character string.}
}
\value{
This function returns a single dataset with
classes of 'tbl_df', 'tbl' and 'data.frame'.
}
\description{
This function reads data saved as csv.bz2 format.
}
\note{
Function throws an error if selected file can not be found.
}
\examples{
\dontrun{
fars_read(filename = 'accident_2013.csv.bz2')
}

}
