% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a file using readr::read_csv suppressing its output}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(character) The file to load}
}
\value{
A tbl_df
}
\description{
This function reads csv files using the read_csv function from readr.
Its messages and progress bar is suppressed.
Entering a non-existent file
while raise an error.
}

