% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{FARS Data Input.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file which the data are to be read from.}
}
\value{
A data frame table (see \code{\link[dplyr]{tbl_df}} in the \code{dplyr} package).
}
\description{
Reads a FARS data file in R as data frame table.
}
\details{
This function is the principal means of reading FARS tabular data into R.

The file must be in a CSV format. If the file, specified through \code{filename},
does not exist the execution is stopped and an error is returned. Otherwise the
content of the file is read. Please note that \strong{while reading the file all
warnings are suppressed}.
}
\section{Depends on}{

\enumerate{
  \item \code{\link[readr]{read_csv}} in the \code{readr} package.
  \item \code{\link[dplyr]{tbl_df}} in the \code{dplyr} package.
}
}

\examples{
\dontrun{
fars_read("/tmp/data/fars/accident_2013.csv.bz2")
}

}
