% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the path of the CSV file}
}
\value{
This function returns a tibble rendered from the CSV file
}
\description{
This function reads a CSV file specified by the /code{filename} argument and
returns a tibble containing the data. If the file does not exist, the function 
will stop and will return an error
}
\examples{
\dontrun{accident_2014 <- fars_read("./data/accident_2014.csv.bz2")}

}
