% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string indicating the path to the csv file with data.}
}
\value{
The function returns a tibble with the data in the  file.
}
\description{
This function checks if the input file exists and then reads it. If the file
does not exist the function will stop with an error.
}
\examples{
fars_read("accident_1995.csv.bz2")
fars_read("accident_2010.csv.bz2")

}
