% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{BUILDING R PACKAGES
fars_functions.R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string of the intended file to be read. Working directory must be correctly set.}
}
\value{
Data frame consisting of the data from the csv file supplied in filename parameter. If no such file exists
then an error is returned.
}
\description{
BUILDING R PACKAGES
fars_functions.R
}
\seealso{
\code{\link{make_filename}}
}
\author{
Yi Shen



Reading file from the Fatality Analysis Reporting System.
This function reads in a .csv file and outputs a dataframe of the csv file.
}
