% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in a file and return a tbl dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of a csv file to read in.}
}
\value{
a data frame tbl containing the data in the file
}
\description{
Read in a file and return a tbl dataframe
}
