% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV file to tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of a CSV file to be read; if the filename is not
valid, an error message "file filename does not exist" is generated.}
}
\value{
A tibble is returned as the output
}
\description{
The \code{fars_read()} function checks if the specified file exists, reads the CSV
file into a data frame using \code{read_csv()}, suppresses any messages, and
converts the data frame into a tibble before returning it as the output
}
\section{External Functions Not Imported}{

\itemize{
\item \code{\link[readr:read_delim]{readr::read_csv()}} used to read the CSV file.
\item \code{\link[dplyr:tbl_df]{dplyr::tbl_df()}} used to convert the data frame to a tibble.
}
}

\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
}

}
