% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Print "fars_read"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving path file}
}
\value{
This function returns a data frame tbl
}
\description{
This is a simple function that trasform your data in data frame tbl.
}
\details{
this function may fail if file is not supported by readr::read_csv#'
}
\examples{
\dontrun{
fars_read(filename)
}

}
