% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS, csv data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the
Fatal Analysis Reporting System (FARS) file to load.}
}
\value{
This function returns a \code{data.frame} (\code{tbl_df}).
   The \code{data.frame} reflects the data in the file given
   by the \code{filename} argument.
}
\description{
This function reads Fatal Analysis Reporting System (FARS) data
from fars, csv files. \code{fars_read} takes a single argument;
\code{filename}; the name of the fars file to read. \code{fars_read}
will stop and report an error, if the given \code{filename} doesn't exist.
}
\note{
This function also reads compressed, csv files, so there's
no need to uncompress them beforehand.
}
\examples{
\dontrun{
fars_read("accident_2015.csv")
fars_read("accident_2015.csv.bz2")
}

}
