% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\value{
This function returns a dataframe of the FARS data.
}
\description{
This function reads data from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System 
into R. It takes a filename as an argument.
}
\note{
The function suppressess messages generated during the data read process and throws an error if the \code{filename} does not exist in the working directory.
}
\examples{
fars_read("accident_2013.csv.bz2")
\dontrun{fars_read("does_not_exist.csv")}

}
