% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads and Loads a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name (with the path) (character) of the CSV file to be read}
}
\value{
A data.frame object with data from the read CSV file
}
\description{
This function reads a CSV file with the name passed as an argument
in the argument <filename>. It returns a data frame with the data
read from the file.
}
\note{
If the file does not exist it stops with an error
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
}

}
