% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in data from a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{An atomic character vector holding the path name of the file.}
}
\value{
A tbl_df data structure with contents of the given file.
}
\description{
\code{fars_read} returns a data structure with data read from a given file.
}
\details{
The function takes an atomic character vector holding the name of the file
and given that it exists will return the data from the file in a tbl_df
data structure. If the file does not exist the function will exit and an
error will be thrown.
}
\section{Required functions to import from other packages}{


readr::read_csv();
dplyr::tbl_df()
}

\examples{
fars_read("data/fname.csv")

fname <- "data/fname.csv"
fars_read(fname)

\dontrun{
 ## The argument must be a character vector. The following will not work.
 fars_read(data/fname.csv)
 }
}
