% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv from fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string that will include the path to the file location}
}
\value{
Returns a data.frame containing the data.
}
\description{
This is a simple function that checks if the file exists.
If it does, it returns a data frame from the input csv. If it does not exist, the function with
stop with an error message
}
\examples{
\dontrun{
fars_read("data/accident_2015.csv.bz2")
}
}
