% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This function reads an existing CSV file and converts it into a
data table in R by returning a tbl_df. The file must be imported from your
your directory. If not the function will produce an error.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{provides the name of the csv file}
}
\value{
a table data frame - tbl_df
An error will occur if the file does not exist
}
\description{
This function reads an existing CSV file and converts it into a
data table in R by returning a tbl_df. The file must be imported from your
your directory. If not the function will produce an error.
}
\examples{
\dontrun{
x <- fars_read("accident_2013.csv.bz2")
fars_read("accident_2013.csv.bz2")
}

}
