% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads the data in a csv-file given by filename into an data.frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A filename as a character. The corresponding file will be read.}
}
\value{
An object of the type tbl_df. If the file does not exist, an error will occur.
}
\description{
Reads the data in a csv-file given by filename into an data.frame.
}
\examples{
\dontrun{fars_read(make_file(2015))}
\dontrun{fars_read("accident_2015.csv.bz2")}
}

