% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read(): Reads FARS (Fatality Analysis Reporting System) Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{String type file name provided to be read into dataframe table}
}
\value{
This function returns a dplyr::tbl_df class object.
}
\description{
This function accepts a filename and reads FARS file (csv format) and creates dataframe table.
The function checks if the file exists and prints error msg when it doesn't.
}
\examples{
fars_read(make_filename(2013))
}
