% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load CSV File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that specifies a the location and name of the file
to be loaded by this function}
}
\value{
This function returns a dataframe or tibble.
}
\description{
This is a simple function that loads a csv file (using the \code{filename} 
argument) and converts it into a dataframe. An error is thrown if the file specified by 
\code{filename} does not exist.
}
\examples{
\dontrun{
accident_2013 <- fars_read("~/data/accident_2013.csv.bz2")
accident_2014 <- fars_read("~/data/accident_2014.csv.bz2")
accident_2015 <- fars_read("~/data/accident_2015.csv.bz2")
}

}
