% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The file name that has the data that should be loaded.}
}
\value{
This functions returns a data frame with file data loaded, if it exists.
   If the file doesn't exist, it stops the function and return a message telling that the file doesn't exist.
}
\description{
This function checks if the file exists in the work directory.
If does, it load it to a data frame.
}
\details{
This function uses read_csv function from readr package.
   In order to load the file, it should be in the working directory.
}
\examples{
\\code{fars_read("data/accident_2013.csv.bz2")}

}
