% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load data from the "Highway Fatality Analysis Reporting System"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file name/path to load the data from. This can be a local file or url to remote file. More details can be found at \code{\link{read_csv}}.}
}
\value{
This function returns the contents of filename as a tbl_df
}
\description{
This is a function that reads a data file from 
the "US National Highway Traffic Safety Administration's Fatality Analysis Reporting System"
}
\details{
The file/path/url specified in \code{filename} must be accessable or a error is thrown.
}
\examples{
\dontrun{
fars_read("fars_data.zip")
fars_read("/home/user1/Download/fars_data.zip")
}

}
