% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the file you want to read in}
}
\value{
This function returns the data from the input file in \code{tbl_df} class
        (see R documentation for the \code{tibble} package).
}
\description{
The \code{fars_read()} function uses the \code{read_csv()} function from
the \code{readr} package to read in a data set and converts it to the
\code{tdl_df} format known from the \code{dplyr} and \code{tibble} packages.
The function has been developed to read in FARS data but works with every data format
that is accepted by \code{readr::read_csv()}. It gives a warning
if the specified file name does not exist.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
