% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file and return it as a tibble dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A filepath character string of a csv file to be read to a tibble}
}
\value{
Returns a tibble of the csv
}
\description{
Reads data from bzipped csv files from the US National Highway Traffic Safety Administrations's
Fatality Analysis Reporting System

Will create an error if the file does not exist at the path
}
\examples{
\dontrun{
path <- "accident_2013.csv.bz"
fars_read(path)

fars_read("accident_2013.csv.bz")
}

}
