% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Loads a CSV file to a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the CSV file.}
}
\value{
a tibble result from loading file.
}
\description{
This function loads a CSV file defined by \code{filename} argument to a tibble.
If the path is incorrect (file not existed), an error will be thrown.
}
\examples{
\dontrun{
accident_2015 <- fars_read("accident_2015.csv.bz2")
}

}
