% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS.R
\name{fars_read}
\alias{fars_read}
\title{Read a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string giving a pathname of a file}
}
\value{
This function returns a dataframe to store data in the file
}
\description{
This is a simple function that reads a file from a path. If the file doesn't exist,
then print a error information. Otherwise, read this file into a dataframe.
}
\examples{
\dontrun{
fname <- system.file("extdata","accident_2013.csv.bz2", package="FARS")
fars_read(fname)
}

}
