% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_W4.R
\name{fars_read}
\alias{fars_read}
\title{Import data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the file name to be imported}
}
\value{
A tibble of the data requested
}
\description{
This is a function to import data in CSV format given a file name.
}
\note{
An error will occur if the file does not exist in the working directory.
}
\examples{
d <- fars_read("accident_2015.csv.bz2")

}
