% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Read a FARS dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The filename of the file to read}
}
\value{
This function returns a tibble containing the FARS data for the given filename
}
\description{
This function reads the specified file and returns a tibble containing the FARS data
}
\details{
Function throws an error os file with the given filename does not exist
}
\examples{
\dontrun{
fars_read('accident_2013.csv.bz2')
}
}
