% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the source file the function will
read}
}
\value{
This function returns the FARS data in a tibble dataframe.
}
\description{
This is a simple function that, given a filename, opens the data and returns
it in a dataframe. Throws an error if the filename is invalid.
}
\details{
The function uses imported \code{readr::read_csv()} to open the file and
\code{dplyr::tbl_df()} to coerce the dat into a tibble object.
}
\examples{
library(readr)
my_file <- system.file("extdata", "fars_data.zip", package = "fars")
unzip(my_file, junkpaths = TRUE)
fars_read("accident_2013.csv.bz2")

}
