% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A name of the file to read (either a single string or a raw vector).
Please use the \code{\link{make_filename}} function to get a correct file name.}
}
\value{
Returns a tibble containing the FARS data loaded from the source file
 (or the several source files).
}
\description{
The function accepts the filename (using \code{filename} parameter)
of the file containing FARS data for a year to process,
checks if the file exists, and returns its data.
For more information about FARS data, see
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}.
}
\examples{
data <- fars_read("accident_2013.csv.bz2")

}
