% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Character string}
}
\value{
This function returns a table data frame.
}
\description{
This function reads the data in a csv file and transforms it to a R table dataframe.
}
\details{
If the file does not exist the function returns an error message.
}
\examples{
\dontrun{
fars_read("filename.csv")
}

}
