% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.r
\name{fars_read}
\alias{fars_read}
\title{Read external csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to the csv file.}
}
\value{
This function returns the csv file in tibble format.
}
\description{
This function check the existence of the csv file and if it exists read it into a tibble.
You need to specify the name of the file in the \code{filename} argument.
The function looks for the file in the wd.
}
\section{Warning}{

This function check for the existence of the specified file.
If the file doesn't exist it stops and throws an error.
}

\examples{
\dontrun{df <- fars_read("name_of_the_file.csv")}

}
