% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the data file (as string)}
}
\value{
data frame named data

if the filename input is not found, returns an error message "file *filename* does not exist"
}
\description{
Read fars data file
}
\examples{
\dontrun{ fars_read("accident_2013.csv.bz2") }
}
