% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string gives the location to the .csv file or
.csv.bz2 file}
}
\value{
This function returns a dataframe. As a side effect, it will also
wrap the tibble format around the dataframe.
\ @examples
\dontrun{fars_read("./data/accident_2013.csv.bz2")}
}
\description{
This function reads in the .csv file containing the data. It will check if
the file exists in the current directory so set it correctly. Lastly, it
will wrap the dataframe with tibble format.
}
\details{
This function will use the read_csv from readr package and tbl_df
from dplyr package.
}
\seealso{
\code{read_csv}
}
