% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Data Input}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name (or path) of the (possibly compressed) file.}
}
\value{
This function returns a tibble. A warning tells you if the requested file does not exist.
}
\description{
Reads a csv file (sep = ,) into a tibble.
}
