% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read NHTSA FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A path to a file or a connection.
Compressed files in with  .gz, .bz2, .xz, or .zip formats supported.}
}
\value{
Data frame.
}
\description{
Reads a Fatality Analysis Reporting System (FARS) data file in CSV format and creates a tidyverse tibble from it.
}
\details{
Returns an error if \code{filename} does not exist.
}
\examples{
\dontrun
{
accident_2013 <- fars_read("inst/extdata/accident_2013.csv.bz2")
}

}
