% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read input data from a given file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving name of a file containing data.}
}
\value{
This function returns data formatted as data.frame.
}
\description{
This function reads data from a file specified in \code{filename} argument.
\code{fars_read} works when the appriopriate data files are in the working
directory, otherwise returns an error.
}
\examples{
\dontrun{fars_read("accident_2001.csv.bz2")}

}
