% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{read in file function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string of characters shows the name of data file}
}
\value{
The returned file is a data frame tbl.
}
\description{
This is a function to read in data file in csv format and its output
is a data frame tbl.
}
\details{
If the file does not exist, it will give an error message.
}
