% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file of FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A charactor string with the name of the file to read}
}
\value{
a data frame with data readed from the csv file, or an error if the file does not exists.
}
\description{
Read FARS data file of csv format and return tbl_df.
If there is no file, error message comes up
}
\examples{
yr <- 2015
file <- make_filename(yr)
fars_read(file)

}
