% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from a csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string specifing the file name.}
}
\value{
A tibble object containing the data from the specified input file.
}
\description{
This function checks the existance of a specified file in the current
directory and returns the data in a tibble when the file exists.
}
\note{
This function will throw an error when the file is not found.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
