% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read Function}
\usage{
fars_read(filename)
}
\arguments{
\item{A}{character with the name of the file you want to study}
}
\value{
data.table object with all the data
}
\description{
This is a function that reads the cvs file with the data.
}
\note{
An error occur when the file does not exist'
}
