% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_fars_fun.r
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to read into R.}
}
\value{
fars_read returns a tibble or lets user know if file doesn't exist.
}
\description{
This function checks for the existence of the file in folder, and if the file
does not exist, stops to let the user know that file does not exist. Otherwise,
the file is read in and converted to a data table.
}
\details{
This collection of functions is used for the US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census providing the American public yearly data
regarding fatal injuries suffered in motor vehicle traffic crashes.
}
