% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads files with given}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{A vector with years}
}
\value{
This function returns list of dataframe with years and month
}
\description{
This function reads all files with years given and creates list of dataframe with year and month
}
\details{
if there is no file with year prefixed after accident_ and suffixed with .csv.bz2, it will give error
}
\examples{
\dontrun{
fars_read_years("2021")
fars_read_years(c(2021,2022))
}

}
