% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARs data for years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{This is a list/vector of years in YYYY format}
}
\value{
This function returns a list of FARS data for
    each year, if the data is found. NULL is returned
    for entries that are not found.
}
\description{
This function loads in FARS data for multiple years.
}
\examples{
fars_read_years(c("2013", "2014", "2015"))

}
\author{
John J Smith, \email{fakeemail123@duvman.ie}
}
\keyword{FARS}
