% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character or integer vector of the years a user wants to read in.}
}
\value{
Returns a list of tibbles with the columns MONTH and year
}
\description{
This is a function that creates files from a vector of years. 
This function needs to import mutate and select from the dplyr package.
This function will throw an error if it is unable to read in the input vector.
}
\examples{
fars_read_years(c(2000,2001,2002))

}
