% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years
fars_read_years Read and store the month and year column for each of the years provided as argument. Function make_file and fars_read name needed.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Year or range of years from which we want to extract the information}
}
\value{
A list with as many tibles as valid years provided. If invalid years are provided the code will create a warning specifying the unvalid year.
}
\description{
fars_read_years
fars_read_years Read and store the month and year column for each of the years provided as argument. Function make_file and fars_read name needed.
}
\examples{
\dontrun{
setwd(system.file("extdata", package = "fars"))
data<-fars_read_years(2013:2015)
}


}
