% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data of specific years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer o chracter string.}
}
\value{
A object of the tbl_df class with two variable: the months and the years.
}
\description{
This is a simple function that filter data in function of the years.
}
\details{
In case in the year does not exist data a error message apprear.
}
