% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads the month and year from accient files and generates tables with month and year from data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of the years in numeric or integer format}
}
\value{
Returns a list of tibbles (data frames) with the same number or rows as the data in "accident_<year>.csv.bz2' files and two columns-month and years.
}
\description{
This function generates tables subtracting month and year from data
}
\details{
Returns NULL and warning if the file does not exist
}
\examples{
\dontrun{
fars_read_years(2013:2015)
fars_read_years(list(2013,2014))
#Results in a warning 
fars_read_years(2016)
}
}
