% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Print "Fars Read Years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of one or more character string years for the desired Fars data.}
}
\value{
This function returns a list with year and month from valid Fars files or will return
a warning and a \code{NULL} value if the file is invalid.
}
\description{
This function loops through one or more years and calls \code{make_filename} to
assemble the  file name for a year and calls \code{fars_read} to read the data.
}
\note{
An invalid year will result in a warning message with nothing returned. If working directory
has not been set, a warning message will result for valid file names.
}
\examples{
fars_read_years(c('2015'))
fars_read_years(c('2014','2015'))
fars_read_years(years = c('2013','2014','2015'))

}
