% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads data from a compressed file based on a specific year, or a set of years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector representing the years of a time period.}
}
\value{
A \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
   object with the data of car fatalities by month for each year in the
   specified period.
}
\description{
This function will read the data for a specific period, and it will
create a new variable (MONTH) in the dataset. Fatalities are then shown per year and month.
}
\section{Warning}{

   If a file does not exist for a specific year, the function will
   show a warning that the information was not read for that year.
}

\examples{
# Data with the car fatalities in 2013.
fars_read_years(2013)
# Data with the car fatalities in 2013 and 2014.
fars_read_years(c(2013, 2014))
}
