% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get months and years from the respective record year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list (numeric or integer) with the years of accident records.}
}
\value{
This function returns a list of tibbles (data.frames) from the respective years of
the records. This function returns a warning and NULL if the file(s) does not exist.
}
\description{
This function accepts a vector or list of years \code{years}, where the records from US
National Highway Traffic Safety Administration's Fatality Analysis Reporting System should
be available. It returns a list of the data.frames with two columns (MONTH, year).
}
\note{
This function will give error if the files are not located in the working directory.
}
\examples{
fars_read_years(2013)
fars_read_years(2013:2015)

}

