% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector or list of years}
}
\value{
Function \code{fars_read_years} returns a list, with the variables \code{MONTH}
and \code{year} for each element in argument \code{years}.
}
\description{
Function \code{fars_read_years} takes a list of years and return the data
of specific years, which only remains the variables of \code{MONTH} and \code{year}.
}
\note{
If a element \code{year} in argument \code{years} is invalid,
function \code{fars_read_years} returns \code{NULL}
and be reported as error in the invalid \code{year}.
If \code{year} is not integer, it would be transfered to \code{\link{as.integer}}.
}
\examples{
fars_read_years(c(2013,2014))
fars_read_years(c(2014,2015))

}
