% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple file names}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numeric vector containing four-digit numeric values.}
}
\value{
A list with length equal to the length of numeric vector \code{years} in case
 \code{year} names are correctly provided. Elements of the list are tibbles containing
 2 columns, \code{MONTH} and \code{year}.
}
\description{
\code{fars_read_years} function takes a vector of numeric values and read their
corresponding file names into a tibble.
}
\details{
\code{fars_read_years} first takes a numeric vector containing four-digit
 elements. It then calls \code{\link{make_filename}} to turn every elements of
 \code{years} vector into a character vector of formatted file names. The file
 names are used to read the corresponding date files into tibbles. After that
 it calls to \code{\link[dplyr:mutate]} and \code{\link[dplyr]{select}} to set
 the name of \code{year} column to the corresponding year and select \code{MONTH}
 and \code{year} columns. for the purpose of error handling the function calls
 \code{\link[base]{tryCatch}} in order to check whether the provided filename
 can be evaluated and read into tibble or not. In case an invalid year name
 is provided an error will be thrown.
}
\examples{
fars_read_years(c(2013, 2014, 2015))

\dontrun{
 fars_read_years(2013, 2014, 2015)
 }

}
