% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple file names}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numeric vector containing four-digit numeric values.}
}
\value{
A list with length equal to the length of numeric vector \code{years}
 in case year names are correctly provided. Elements of the list are
 tibbles containing 2 columns, month and year.
}
\description{
This function takes a vector of numeric values and read their
 corresponding file names into a tibble.
}
\details{
\code{fars_read_years} first takes a numeric vector containing four-digit
 elements. It then calls \code{\link{make_filename}} to turn every elements of
 years vector into a character vector of formatted file names. The file
 names are used to read the corresponding date files into tibbles.
 for the purpose of error handling the function calls\code{\link[base]{tryCatch}}
 in order to check whether the provided file name can be evaluated and read
 into tibble or not. In case an invalid year name is provided an error will
 be thrown.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
}

}
