% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in multiple years of FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer years, if data for any years don't exist, no data will be returned}
}
\value{
this function returns a tibble with data from all years of interest
}
\description{
Read in multiple years of FARS data
}
\note{
NULL returned if no data for any input years
}
\examples{
\dontrun{fars<-fars_read_years(c(2012,2013,2014))}

}
