% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in "FARS" data by specifying the year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The input year. If the input year does not exist,
a warning message is thrown: invalid year "year".}
}
\value{
This function creates a list of the months and year for data specified with the input year.
}
\description{
This is a function that reads in FARS data by using as an input the year.
It returns the months and year as a list. An warning message is thrown, when an invalid year is chosen
as an input.
}
\examples{
\dontrun{fars_read_years(2013)}

}
