% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading the Fars Dataset}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Integer or numerical index for the years}
}
\value{
The function return a list of FARS dataset with corrected names and time signal columns
}
\description{
This function takes a list of years for FARS dataset and convert it to the  tidy dataset with
             additional column of the year.
}
\details{
The function will return an error if the years inserted is not available within the list
}
\examples{
\dontrun {years <- c(2006:2010)}
\dontrun {fars_read_years(years)}
}
