% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create a dataframe only for the specific years.
Each year would have a filename of the form \code{accident_<year>.csv.bz2}}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers}
}
\value{
list of dplyr tbl_df instances each for the year. If data for the year
specified cannot be found, a warning is printed and NULL is returned for that entry
in the year.
}
\description{
Create a dataframe only for the specific years.
Each year would have a filename of the form \code{accident_<year>.csv.bz2}
}
\note{
This call assumes that the data files are
directly accessible under the current directory. Use \code{getwd()} to confirm.

MONTH refers to the numeric representation with January being 1.
}
\examples{
\dontrun{
data <- fars_read_years(c(2014, 2015, 2021))
Warning message:
In value[[3L]](cond) : invalid year: 2021

> head(x=data[1])
[[1]]
A tibble: 30,056 x 2
MONTH  year
# <int> <dbl>
 1     1  2014
2     1  2014
3     1  2014
4     1  2014
5     1  2014
6     1  2014
7     1  2014
8     1  2014
9     1  2014
10     1  2014
# ... with 30,046 more rows
> head(x=data[2])
[[1]]
# A tibble: 32,166 x 2
MONTH  year
<int> <dbl>
  1     1  2015
2     1  2015
3     1  2015
4     1  2015
5     1  2015
6     1  2015
7     1  2015
8     1  2015
9     1  2015
10     1  2015
# ... with 32,156 more rows
> head(data[3])
[[1]]
NULL
}
}
