% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read fars datafiles for a given list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list with a list of years (given as strings)}
}
\value{
A dataframe with entries in data by month, or NULL (and a warning) if the
 \code{year} is not a valid year in the dataset.
}
\description{
This function is used by \code{fars_summarize_years}
}
\examples{
\dontrun{
fars_read_years(2013)
}

}
\seealso{
\link{fars_read}

\link{make_filename}

\link{fars_summarize_years}
}
