% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create a filename format}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years for which to read the FARS data. See \code{\link{make_filename}}}
}
\value{
Returns a list of tbl_df containing the MONTH column from the FARS data and an additional column for the
    year. Returns NULL and a warning for every FARS data file that does not exist.
}
\description{
Create a filename for a Fatality Analysis Reporting System (FARS) CSV file
}
\examples{
\dontrun{
fars_2013_2016 <- fars_read_years(2013:2016)

# Returns NULL and a warning
fars_2022 <- fars_read_years(2022)
}
}
