% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Tables of Yearly Accident Data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{\code{integers}}
}
\value{
Returns a \code{list} of \code{data.frame}s, each item in list corresponds to one of the years in the FARS data.
}
\description{
The function will generate a list of dpylr dataframes where each table represents a year of the data.
Must be specified by the user.
}
\examples{
\dontrun{
# invalid input, year doesn't exist
fars_read_years(c(2013, 2014, 1900))
}

}
