% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years from a list of Fatality Analysis Reporting System}
\usage{
fars_read_years(years,path)
}
\arguments{
\item{years}{a list of values corresponding to year}

\item{path}{directory containing the Fars data set to "data" by default}
}
\value{
a list of dat frame tbl with MONTH and year columns
}
\description{
This function generates lists of a projection of the data along the column month and year.
filtered by the year as written in the file name.
}
\note{
\itemize{
\item file names are generated with the \code{make_filename} function
\item require to be run from the directory containing the data
\item each list element may contain duplicated month year tuple
}
}
\examples{
\dontrun{
# need to be in the same directory as the data
setwd("data")
# with one year
fars_read_years(2013)
# with multiple years
fars_read_years(2013:2015)
# with characters
fars_read_years(c("2013","2015"))
}

}
\seealso{
fars_read

make_filename
}
