% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years - Read fars datafiles for a given list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years of the accidents as a list or vector of integers}
}
\value{
The accidents for given years sorted by month as list of data frames
    with integers and numeric elements. Returns NULL (and a warning) if the
    \code{year} is not a valid year in the dataset
}
\description{
This function is used by \code{fars_summarize_years}
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years("2014")
fars_read_years(c(2013,2014))
fars_read_years(years = 2013:2015)
fars_read_years(years = list(2013,2014,2015))
}

}
\seealso{
\link{fars_read}

\link{make_filename}

\link{fars_summarize_years}
}
