% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function \code{fars_read_years}}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The specified year}
}
\value{
A data frame with number of deaths for each month in the given year
}
\description{
A function that returns the number of deaths in a year per each month of the year
Will return with a warning message, if the year can not be found
}
\note{
This is an internal function not to be used from outside the package
}
