% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data for Multiple Years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer or strings that indicated the year of the FARS data files.}
}
\value{
a list of FARS data for the years in the parameter \code{years}.
  Invalid years will have  a NULL entry in the returned list.
}
\description{
This function takes as its only parameter a vector of years, reads the FARS
   files for those years and returns the MONTH and year columns as a list.
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{mutate}} and \code{\link[dplyr]{select}} in the \code{dplyr} package.
}
}

\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
fars_read_years(2013:2015)
fars_read_years(list(2013, 2014, 2015))
}

}
