% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character vector of years which will be cycled through
to form the list of data sets}
}
\value{
This function returns a list of tbl_df objects, each containing
data for the years given in the input. An error will be returned for invalid
years.
}
\description{
This is a simple function that creates filenames for each of the values
in the input and then forms a list of tables from the files which match the list
of filenames.
}

