% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return \code{month} column at each year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of years}
}
\value{
A list of \code{month} under specific \code{year} for all input \code{years}, or an error message \code{invalid year: year}"
}
\description{
Return \code{month} column at each year
}
\examples{
setwd(system.file("extdata", package = "FARS"))
dat_list <- fars_read_years(c(2013, 2014, 2015))

}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read}}
}
