% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{batch import csv file}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
A list of tibbles. Each tibble has 2 columns- month and year
}
\description{
Iterate over a year vector that read csv data for each given year. if the csv file for that year does not exist, throw a warning.
If exist, create a year column and select only year and month columns
}
\examples{
\dontrun{fars_read_years(2013:2015)}

}
