% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years - read files with all the years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{List of years whcih the files will be read.}
}
\value{
a list of \code{\link{dplyr}} \code{\link{tbl_df}} which shows
   the month there is an accident in that year.
}
\description{
This function takes a list of years of the accidents 
and read all the coorisponding files.
}
\note{
Error when file does not exist or cannot be read
}
\examples{
\dontrun{recent_accident <- fars_read_years("2015")}
\dontrun{other_accident <- fars_read_years(c(2013,2014))}

}
