% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Given a list or vector of years, return the month/year combinations for all observations in the FARS
data belonging to the given years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector of values that can be coerced to integers and that each represents a year.}
}
\value{
A list of tibbles, one tibble for each year in \code{years}, which each contains the month/year
        combination for each observation in the FARS files for the given \code{years}.
}
\description{
Given a list or vector of years, return the month/year combinations for all observations in the FARS
data belonging to the given years.
}
\note{
The elements in \code{years} must be castable to integers.
}
\examples{
\dontrun{
fars_read_years(list(2014))
fars_read_years(list("2013",2014,2015))
}
}
