% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in data for several years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Collection of years as vector or list where each element is in integer format or can be converted with as.integer}
}
\value{
This function returns a list of dataframes where each dataframe contains the data for one year
}
\description{
This function reads in the accident data for multiple years as a list of dataframes. If one of the years can not be 
read in the function skips that year with a warning
}
\note{
Errors in this function can result from errors thrown by the fars_read function or invalid years parameter being passed
}
