% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{FARS Year Reader}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character vector, numeric vector, or integer vector containing
the years for which data will be read.}
}
\value{
A vector containing a data frame for each inputted year's data,
 where each observation only contains its month and year.
}
\description{
This function takes a vector of years, and returns a vector of data frames
 that correspond to the inputted years' data. Each data frame is simplified
 such that each observation only gives the month and year of the observation.
It uses functions from the 'magrittr' and 'dplyr' package, as well as the functions
 'make_filename' and 'fars_read' defined earlier.
}
\note{
Inputting years for which there is no data, or years that are improperly
 formatted, will result in errors.
}
\examples{
\dontrun{fars_read_years(c(2013,2014,2015))}
\dontrun{fars_read_years(c("2013","2014"))}

}
