% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of data
This function will take a vector of years and return a list of data frames.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
If the years are valid, the function will return a data frame. Otherwise the function will return null.
}
\description{
Read multiple years of data
This function will take a vector of years and return a list of data frames.
}
\examples{
\dontrun{
fars_read_years(c(2013,2014,2015))
example_years <- 2013:2015
fars_read_years(example_years)
}
}
