% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads the data from a file for some years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{List or vector of the years to be extracted from the datafile.}
}
\value{
This function returns a list of data frames containing the data of the file processed by years.
Each data frame has two columns for the month and year of a sample in a data file.
}
\description{
This function tries to collect the data for one or more years from the datafile.
 \code{years} argument should be a list of the required years. If no data found for a
 specific year, a warning message is displayed and the element in the output listis NULL.
}
\examples{
\dontrun{
fars_read_years(years=c(2014,2015))
}

}
