% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create new data files from the FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Desired range of years as numeric values. If a year is incorrect, a message is issued.}
}
\value{
Returns a filtered list of dataframes with a year and month variable or NULL if it doesn't exist.
}
\description{
Read the desired yearly data file and create a new data set with two variables, month and year.
}
\details{
Uses the dplyr function to create a new data files with month and year variables.  If the year is not valid, an error essage will be issued.

Uses the function make_filename and fars_read
}
\examples{
df<-fars_read_years(2013:2015)
}
