% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads in list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years}
}
\value{
edits the data files with names ending in provided year range
}
\description{
Reads in a list of years and for each file whose name ends in a year within the
provided year range changes the year in that file.
Throws an error if the year range contains a year which doesn't correspond to
a file name with that year in its name.
}
