% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function will reset the year column with provided year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years for which data will be loaded}
}
\value{
This function returns a dataframe month and year after resetting the year column with provided year
}
\description{
This function will reset the year column with provided year
}
\details{
This function raise errors when year value is invalid
}
\examples{
\dontrun{
fars_data <- fars_read_years(c(2013,2014))
}

}
