% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Make list of data.frames for each year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of years. If there are no file with specific year NULL is returned
#'}
}
\description{
Takes years and makes list of data.frames. Each elements is data.frames with month and year
}
\examples{

\dontrun{
 fars_read_years(c(2013, 2014, 2015))
 }
}
