% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years Generate tibbles for each year of data needed.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years for which data is desired.}
}
\value{
A list of tibbles with the data for each of the years provided.
}
\description{
fars_read_years Generate tibbles for each year of data needed.
}
\section{Errors}{

This function will throw an error if an unavailable or invalid year is encountered.
}

\examples{
#Basic usage:
fars_read_years(years = c(2013, 2014, 2015))
}
