% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Open the file with specifying the list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{- variable containing only one year or many years}
}
\value{
tibble(s) showing month and year variables only.
}
\description{
This function allows you to open the accidents files for wanting year(s).
}
\details{
if there is no file which contains a year that you submitted, a warning will be shown and NULL will be returned
because the file with the given year does not exist
}
\examples{
\dontrun{
fars_read_years(2014)
fars_read_years(2013)
fars_read_years(c(2013,2014))
}

}
