% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{print("fars_read_years")
This is a function that reads and returns a list with the data from the US
National Highway Traffic Safety Administration's Fatality Analysis Reporting
System (FARS), provided a year or vector of years. Each year's data is stored
is a different component of the function return as a tibble object. The list
of years is passed onto the function as an argument, which can be done using
the \code{years} argument. If any of the provided years is not in the correct
format, it returns an error message.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{the year for which the filename will be constructed.}
}
\value{
This function returns a list of tibbles containing each the data
   corresponding to a particular year of the FARS database.
}
\description{
print("fars_read_years")
This is a function that reads and returns a list with the data from the US
National Highway Traffic Safety Administration's Fatality Analysis Reporting
System (FARS), provided a year or vector of years. Each year's data is stored
is a different component of the function return as a tibble object. The list
of years is passed onto the function as an argument, which can be done using
the \code{years} argument. If any of the provided years is not in the correct
format, it returns an error message.
}
\examples{
fars_read_years(2015)
fars_read_years(years = c(2013:2015)

}
