% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract the month from every accident in the data set}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numeric vector representing one or more years}
}
\value{
a list of data frames (a data frame for every year) with the month
  for every accident in the FARS data set data for each the selected years.
}
\description{
Extract the month from every accident in the data set
}
\note{
If one of the years used as input does not exist the function will
  throw a warning, but will read correctly for all the valid values. The
  invalid year will have a corresponding empty data frame in the output list.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}
}
