% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of motor vehicle crash data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A string or integer vector giving the years of motor vehicle crash files to read in.}
}
\value{
This function returns a list of data frames if the motor vehicle crash year file exists.
		An warning message is returned if a file does not exist in the current directory.
}
\description{
This function (fars_read_years) accepts a vector of years or a single year as input.
Each element of the input vector is treated separately and a list of data frames is returned.
The year is converted to a filename using the make_filename function from this package (\code{\link{make_filename}}).
The file is then read using the fars_read function from this package (\code{\link{fars_read}}).
A column containing year is set to the year value contained in the filename using the mutate function from the dplyr package (\code{\link[dplyr]{mutate}}).
Only the MONTH and year columns for each file are returned by using the select function from the dplyr package (\code{\link[dplyr]{select}}).
A warning message is returned if a file does not exist.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013, 2014))
fars_read_years("2013")
fars_read_years(c("2013", "2014"))
fars_read_years(years=2013)
}

}
