% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading a set of FARS files referring to multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric or character vector containing the years whose data
one wants to read.}
}
\value{
The function returns a list of tibbles, one per year. If a certain
year is not valid, a warning is generated and the list returned will
contain a value of \code{NULL} in the position corresponding to that year.
}
\description{
This function takes as argument a vector containing the years whose data
one wants to read and returns a list of tibbles, each consisting of two
columns, MONTH and year, containing the month (number) and year of each
accident reported in that year. The function uses the \code{mutate} and
\code{select} functions, as well as the \code{\%>\%} operator, all from
the \code{dplyr} package.
}
\examples{
fars_read_years(c(2013, 2014, 2015))
fars_read_years(c("2013", "2014", "2015"))
fars_read_years(2013)
fars_read_years("2014")

}
