% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data for years into a list of data frame tables}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A \code{numeric} vector of years for which data is to be read}
}
\value{
A list of data frame tables. For invalid inputs, a warning message is printed
and returns \code{NULL}.
}
\description{
This function reads data for the years provided and it returns
a list containing the data frame tables. The output contains the data
for all the months of the specified years.
}
\examples{
\dontrun{
fars_read_years(2010:2012)
}

}
