% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years read accidents data by given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{- vector of years.}
}
\value{
data frame of class dplyr::tbl_df with two columns (month ans year) per each accident. If data for any of given years do not exist
function stops working and return an error message.
}
\description{
Read accidents data by given years and return data frame with short info for each accident.
}
\examples{
fars_read_years(c(2013, 2014, 2015))

}
