% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_original.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract month and year for all observations from an accident file.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{One or more years, numeric integer value specified in form of a list.}
}
\value{
This function returns a list, one element for each specified year and each element is a tibble
   with 2 columns (MONTH and year), the number of observations is the number of accidents in the year.
   If the parameters are not legal integers the function return an error.
   If a specified year is not valid year, the function returns a warning message indicating the invalid year.
}
\description{
This function extract the month and the year for each observation (accident) of a file and
put the results in a list. In the list there is one element for each year and each element is a tibble
with 2 columns (MONTH and year) for each accident.
}
\examples{
\dontrun{
read_years<-fars_read_years(years=2013:2015)
}

}
