% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Obtains years into file}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector with desired years in format YYYY}
}
\value{
A list with an element for each file including MONTH and year
        Error if doesn't exist the file associated to year
}
\description{
Load the data files associated to each year and extract month and year
}
\details{
Print "fars_read_years"
}

