% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Testing if a vector specified by the user has data available.}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{is a numerical vector, or a vector that can be coerced to numerical, of years}
}
\value{
This function returns a list. Each element of the list contains the months and year that data is available.
If a year is not available an error is returned.
}
\description{
This function determines if data is available in the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System for the a years specified by the user.
}
\details{
An error will be generated if a year is given and no data is available for that year. This
allows the used to be aware of the years when data is available to this package.
}
\examples{
\dontrun{fars_read_years(2013)
fars_read_years(c(2013,2014))
fars_read_years(c(2013,2014,2015))}
}
