% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{reads fars files for one or more years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{One or more years as integer or string}
}
\value{
a list of data frames
}
\description{
\code{fars_read_years} produces a list of FARS data containing months and years.
}
\details{
for the given years fars file produces a dataframe of year and month observations.
}
\note{
If the file for a specified year, or permission to it, does not exist, an error is produced.
}
\examples{
\dontrun{
fars_read_years(years = 2013)
}

\dontrun{
fars_read_years(years = 1066) # error
}

}
