% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R

\name{fars_read_years}
\alias{fars_read_years}
\title{Grabbing the Year and Month From Data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years you are interested}
}
\value{
A list of data frames for each year
}
\description{
Returns a data frame containing specific year and month
}
\details{
If you entered a year that is not valid, there will be a warning.
}
\examples{
fars_read_years(2013:2014)
}
