% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Helper function to select months from a year list.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector of values representing years.}
}
\value{
A tibble with one row per accident and
columns year and MONTH reporting the accident date.
}
\description{
Since data is split in files one per year, a year
column is added and the results combined into a
single tibble.
Checks for year validity.
Used by \code{fars_summarize_years} and \code{fars_map_state}.
}
\examples{
fars_read_years(c(2013, 2015))

}
