% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years function}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of numbers specifying the years of interest}
}
\value{
this function returns a data frame. If the year of interest
doesn't exist, it will print a warning message.
}
\description{
this is a function that reads in complied file names with specified years
of interest and then returns data by month of the year of interest
}
\examples{
\dontrun{fars_read_years(2013)}

}
