% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of years' FARS files to open.}
}
\value{
A data.frame including entries in data by month, or NULL if the
 \code{year} is not valid.
}
\description{
This function takes a vector of years and produces a list of tibbles,
where each tibble is that year's FARS file year and MONTH observations.
}
\examples{
fars_read_years(years = c(2014, 2015))

}
