% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the CSV file of a given year(s) and add year column}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Numeric (including vector) for the years you want to check}
}
\value{
A tibble data frame with extra column in year
}
\description{
A function that compile all the years you want to look the data.
Currently only accept 2013-2015, if the year is not available, then it will throw up warning and return NULL
}
\examples{
\dontrun{dat <- fars_read_years(2013)}
}
\seealso{
\code{\link[=fars_read]{fars_read()}}
}
