% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function to reads multiple files based on years parameter.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Valid years as a vector (atomic or list) or an expression object.}
}
\value{
File data is read into a dataframe object by MONTH and year.
}
\description{
This function takes the years parameter.
It reads multiple files, one at a time, for each year and returns a dataframe object with data for all years.
It outputs a warning message if it cannot find a file for a year in the years parameter. It returns NULL.
It uses functions: make_filename and fars_read.
}
\examples{
fars_read_years(list(2013, 2014, 2015))
fars_read_years(year=list(2013, 2014, 2015))
fars_read_years(list("2013", "2014", "2015"))
fars_read_years(year=list("2013", "2014", "2015"))

}

