% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load MONTH and year into R}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An character vector, integers or a list containing years}
}
\value{
a list of data frames tbl containing MONTH and a newly created variable 
year for each year specified as parameter
Returns  \code{NULL} if the years can not be found in the files in the directory
or if the directory is not properly set
}
\description{
This function reads yearly and monthly data from 
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
and returns an error if the selected years are not written corrected, found in the 
working directory files or if the working directory is not set properly
}
\examples{
\dontrun{
fars_read_years(2013)
}
}

