% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS files for one or more years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The year(s) of FARS data to examine.}
}
\value{
This function returns a list of tibbles, each of which contains the
year and month from the observations in the corresponding FARS data.
}
\description{
This is a function that reads in FARS data and, based on the years provided
by the user, creates individual tibbles containing the month and year from
the observations in the corresponding FARS data.
}
\examples{
fars_read_years(years = c(2013, 2014, 2015))
fars_read_years(years = 2015)

}
