% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS year data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of intergers, representing the years to read in}
}
\value{
This function returns a data set of the years specified.
}
\description{
This reads in data for a number of years as specified in the vector \code{years}.
Each file is read for each year, and added to an overall data set.
}
\details{
If there is any year without an associated data file, the function will report a warning and cease working.
}
\examples{

\dontrun{
fars_read_years(c(1991,1992,1993))
fars_read_years(2001)
fars_read_years(c(1991,2002,2013))
fars_read_years("ababab")
}

}
