% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading CSV into list}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector}
}
\value{
Returns a list of each year
}
\description{
Reading CSV into list
}
\details{
This function puts all input table into one list. There will be an error message when the year is not invalid.
}
\examples{
\dontrun{fars_read_years(2013:2015)}
}
\author{
James Bertschy \email{james.bertschy@gmail.ch}
}
