% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function reads a list of years and returns a list of data.frames.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{an integer}
}
\value{
This function returs a list of data.frames
}
\description{
This function reads a list of years and returns a list of data.frames.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
}
