% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads month and year from accident files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The vector of years to be considered.}
}
\value{
A list of data frames with month and year of Returns NULL and a warning if the file does not exist.
}
\description{
The function takes a vector of years and returns a list of data frames with month and year based on data in "accident_<year>.csv.bz2".
}
\examples{
fars_read_years(1992:1996)

}
