% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{creates new data files from the FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Desired range of years as numeric values. if a year is incorrect, an error essage will generate}
}
\value{
Returns a filtered list of dataframes with a year and month variable or NULL if it doesn't exist
}
\description{
creates new data files from the FARS data
}
\details{
Uses the dplyr function to create a new data files with month and year variables. if the year is not valid, an error essage will generate

Uses the function make_filename and fars_read
}
