% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return the month and year of each FARS observation}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer or a string corresponding to a specific year}
}
\value{
The fars_read_years function returns a list of months and years.
}
\description{
The fars_read_years function takes as input an integer or a string that represents
a specified year.  It then reads in the data from the *.csv file corresponding to
the given year and returns each observation's month and year.
}
\details{
The fars_read_years function returns an error message for invalid years.
}
\examples{
\dontrun{
  fars_read_years(2015)
}

}
