% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Select the month and year columns from the files whose years you specify.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years for the files you're interested in.}
}
\value{
This function returns a list, with each element of the list being a
tibble for each year in the vector of years specified in argument \code{years}.
Each tibble in the list contains two columns, "MONTH" and "year".
}
\description{
This function is used inside of the \code{fars_summarize_years()} function
and input for the \code{years} argument comes from the \code{years} argument
of the \code{fars_summarize_years()} function.
}
\details{
This function takes a vector of years as input and uses \code{lapply()} to
make a filename for each year, read in that file as a tibble and select the
month and year columns in that tibble.
}
\note{
if even one of the years in the vector passed to argument \code{years}
does not exist (if there is no file for that year), no results, only an error
message, will be returned; "invalid year: 2017" for example.
}
\examples{
fars_read_years(c(2013, 2014, 2015))

}
