% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{integer vector}
}
\value{
data.frame
}
\description{
Reads in fars files given a vector of years.
Returns a data.frame with columns MONTH, year
}
\note{
Throws a warning and returns NULL when an invalid year is part of the vector.
}
\examples{
if(interactive()){
   df <- fars_read_years(c(2013, 2014, 2015))
}

}
