% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A integer vector containing 1 or more years datasets. If
a year or more do not exist it returns a error message!}
}
\value{
A list containing the values of month and year for each year(s) entered
}
\description{
This function read multiple year(s)/dataset(s) and check if they exist(s)
}
\examples{
\dontrun{
m_years <- c( 2011, 2012, 2013, 2014)
fars_read_years(m_years)

fars_read_years(2015)
}
}
