% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data for One or More Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of 4 digit years, coerced to integers.}
}
\value{
This function returns a \code{dplyr::tbl_df} with month and year columns.
}
\description{
Reads a series of Fatality Analysis Reporting System (FARS) files and returns a dplyr tbl_df of the month and year columns for each file.
}
\note{
If the file for a specified year, or permission to it, does not exist, an error is produced.
}
\examples{
fars_read_years(2003:2015)

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}

