% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of FARS data}
\usage{
fars_read_years(years = c(2013, 2014))
}
\arguments{
\item{years}{(integer vector)}
}
\value{
list of tibbles
}
\description{
For specified years, read the MONTH and year values of the Fatality Analysis Reporting System (FARS)
data in a list of tibbles.
Only data for the years 2013, 2014 or 2015 is available.
This function is not exported.
}
\details{
Throws an error if data for specified year does not exist.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
