% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract accidents per month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of data frames. Each data frame corresponds to one year and
consits of columns for month and year as well as rows for accidents.
If type of \code{years} is not correct or if its elements are invalid,
an error message is generated and nothing is returned.
}
\description{
The function \code{fars_read_years} reads data from the US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System
and extracts accidents per month and year.
}
\examples{
fars_read_years(2013:2015)
fars_read_years(c(2013, 2015))

}
