% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This is a function that reads a lists of data input files with an argument specifying the \code{years} to read. It uses \code{\link{make_filename}} function internally.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years in integer format to read. It might stop if the argument is not in integer format.}
}
\value{
A list of data.frames (tibble), in which each element contains data for a year.
}
\description{
This is a function that reads a lists of data input files with an argument specifying the \code{years} to read. It uses \code{\link{make_filename}} function internally.
}
\examples{
\dontrun{
years = seq(2013, 2015)
fars_read_years(years)
fars_read_years(years=years)
}

}
