% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function \code{fars_read_years} take as an argument a four-digit year and
makes a dataset containing columns month and year of an accident that occurred}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Four-digit year, as integer or as a string (i.e. 2001, "2001").}
}
\value{
Data frame with columns:
\describe{
   \item{MONTH}{a value between 1 and 12, 1 for January, 12 for December}
   \item{year}{four-digit value for a year}
}

If the requested year input is not present in the data, the function returns
a warning: "invalid year".
}
\description{
Function \code{fars_read_years} take as an argument a four-digit year and
makes a dataset containing columns month and year of an accident that occurred
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years("2013")
fars_read_years(c(2013,2014,2016))
}
}
