% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract months and years}
\usage{
fars_read_years(years)
}
\value{
This function returns a tibble with the columns MONTH and year.
}
\description{
This function reads csv files for given \code{years}, extracts their months, and adds a column with the years.
}
\note{
This functions throws a warning if a file for the given year does not exist.
}
\examples{
\dontrun{
setwd(system.file("extdata", package = "fars"))
fars_read_years(2013:2016)
}

}
