% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_pack.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{-   Years as a 4 digit number.
The year  can be provided as single year, as a vector or list of years}
}
\value{
Dataframe  - containing only the year and month from the underlying FARS data.
}
\description{
This reads one or more years Fatality Analysis Reporting System (FARS) data
created from the read_fars function containing only the Month and year from the data
}
\note{
Function requires the magittr library to be available.
If the underlying CSV file is not stored in the current working directory the process
will call an warning :
      In value[[3L]](cond) : invalid year: 2013

Code needs to be updated to unclude a directory option as an input parameter
}
\examples{
\dontrun{
   fars_read_years(2013)
   fars_read_years(c(2013,2014,2015))
   fars_read_years(c("2013","2014","2015"))
}
}
\references{
The US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census, provides the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
You can download the FARS data at \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\seealso{
\link{fars_read_years}
 \link{fars_map_state}
 \link{fars_summarize_years}
}
