% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with a list of years.}
}
\value{
A data.frame including entries in data by \code{month}, or \code{NULL} if the
year is not valid.
}
\description{
Helper function used by \code{fars_summarize_years}.
The function accepts a vector or list of years and returns a list of data
frames with \code{month} and \code{year} columns based on data in \code{"accident_<year>.csv.bz2"}
files. The files need to be located in the working directory.
}
\examples{
fars_read_years(2013)

}
\seealso{
\link{fars_read}

\link{make_filename}

\link{fars_summarize_years}
}
