% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read a range of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with the years value}
}
\value{
This function returns a data frame summarized by years and month
}
\description{
This function reads a range of years, from 2013 to 2015
}
\note{
If the year doesn't exist, the function throws an error
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013,2014))
fars_read_years(c(2013,2014,2015))
}

}
