% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create a list of FARS datasets separated by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of years for which data are desired.
Can alternatively be a list of numeric values. The years should be 
in four-digit format, e.g., 2013. Can alternatively be a vector of 
strings, e.g., c("2013","2014","2015")}
}
\value{
This function returns a list. Each element of the list is 
a \code{tbl_df} object showing the month and year columns of a  
FARS dataset corresponding to a given year.
}
\description{
This function reads in FARS datasets for select years and returns 
each dataset as a separate element in a list.
}
\note{
If an error occurs while trying to read in a FARS dataset, a 
warning is given and the list element for that year is NULL. This 
function uses the functions \code{make_filename} and \code{fars_read}.
}
\examples{
my_list <- fars_read_years(years=2013:2015)

}
