% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read-in fars data (Fatality Analysis Reporting System)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector. The years of the traffic data.}
}
\value{
A list of tibbles containing the traffic data for each year.
}
\description{
\code{fars_read_years} reads-in fars (Fatality Analysis Reporting System)
data for specified years.
}
\section{Warning}{

If an invalid year has been given an error will be thrown and \code{NULL}
returned.
}

\examples{
\dontrun{
fars_read_years(2014)
fars_read_years(c(2013, 2014))
}
}
