% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of the years that correspond to the data}
}
\value{
a list of tibbles containing de data from each year.
}
\description{
The function reads the data from every year in the vector input
}
\examples{
fars_read_years(c(2010, 2011))

}
