% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Aggregate FARS data across years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer, vector, or list of year(s)}
}
\value{
A list containing tibbles
}
\description{
This function reads in one of more years of FARS data and aggregates the data into a list. Each
list element is a tibble for a specific year. Only the month and year columns are retained.
}
\note{
If there is no FARS file for a specific year this function will return a warning and a NULL list element

Imports functions from the extenal packages dplyr and magrittr
}
\examples{
fars_read_years(2013:2015)

}
