% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files for indicated years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{List or vector of years}
}
\value{
List of tibble data frames
}
\description{
Read files for years passed as arguments. If some year doesn't exists,
it throw an error.
}
\examples{

fars:::fars_read_years(list(2013, 2015))

}
