% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read accident year records in a list}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{is a vector or list  containing the accidents years.}
}
\value{
This function returns a list containing the FARS months and years records.
}
\description{
This is a support function to read the FARS  years  records in a list.
}
\section{\code{Error}{
}
If \code{years} contains an invalid year, the function gives a warning message.
if \code{years} does not contain a valid year, the function returns a null list
}
\examples{
fars_read_years(c("2013","2014"))
fars_read_years(c(2013,2014))
fars_read_years(list("2013",2014))
}
\seealso{
\code{make_filename}
\code{fars_read}
}

