% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Filter and select data for given years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric or character vector that can be convereged to numeric vector,
 meaning years to be filter.}
}
\value{
This function returns a list with data frames for each year, containing two
 variables MONTH (int) and year (chr).
}
\description{
This is a simple function that filter data for given years. In case if data for given year
 does not exist appropriate message is shown and NULL value is returned for this year.
 Function selects only MONTH and year variables from a data source.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
\dontrun{fars_read_years(c("2013","2014"))}
\dontrun{fars_read_years(2014)}
\dontrun{fars_read_years("2014")}
}

