% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Lee años}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector entero que indica los años}
}
\value{
regresa variables añadidas a un conjunto de datos, se muestran solo las
  seleccionadas MONTH and year
}
\description{
Esta funcion permite aplicar a varios años la función anterior
}
\details{
Si no se introduce correctamente el argumento me regresa un mensaje de adevertencia:
  invalid year:year
}
\examples{
\dontrun{
fars_read_years(2013:2015)
fars_read_years(list(2013, 2014))

# Results in a warning
fars_read_years(2016)
}

}
