% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read month data from a dataset}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of integers representing years}
}
\value{
This function returns a list of dataframes. Each dataset contains two columns:
the column \code{MONTH}, extracted from the datasets contained in the package, and the
column \code{year}, containing the year to which the dataset refers.
An error is thrown in case an invalid year is inputed.
}
\description{
This function can be used to extract the MONTH column from the datasets provided
in the package. The packages to which the function will be applied are selected via
the the argument \code{years}.
}
\examples{
\dontrun{
fars_read_year(list(2014,2015))
}

}
