% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{numerical inputs to indicate the years of the required data}
}
\value{
The function uses the make_filename function and returns a tibble data.frame based on the
multiple provided years under csv format. The error occured if the input year does not exist.
}
\description{
Read the multiple year of Fatality Analysis Reporting System data files
 based on the provided years.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
\seealso{
\code{\link{make_filename}}
}
