% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Print "fars_read_years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector that contains POSIXlt date time format}
}
\value{
This function returns a data frame with the traffic crashes occurred on
each month as per input year. It generates warning message upon unavailable data.
}
\description{
This function reads data content from FARS dataset and summarizes
traffic crashes by year and month.
}
