% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function creates n data frames of months and years.  The months will vary within each data frame, but the year will remain constant.
n will vary with the number of years input.}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System \url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years in which a separate data frame will be created for each year.}
}
\value{
The function will return a list of data frames for each year input as an argument.
}
\description{
This function creates n data frames of months and years.  The months will vary within each data frame, but the year will remain constant.
n will vary with the number of years input.
}
\note{
If a year is input in which a corresponding file does not exist in the working directory, and error will occur.

Please make sure the csv files and R script are in the same directory, and that directory is the current working directory.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013,2014))
}
\references{
This function uses both the fars_read and make_filename function from this package as steps to create the data frames output.
}
