% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS data for Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{List of years. Each year will be coerced to an integer with \code{\link[=as.integer]{as.integer()}}}
}
\value{
A list of tibbles, one for earch year and containing only the Month and Year from each row of
the FARS data for the given years. If a year is invalid, then a warning message is issued
and NULL is returned.
}
\description{
This function takes as input a list of years and loads the data from
US National Highway Traffis Safety
Administration's
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Anlaysis Reporting System}
(FARS) for those years.
}
\examples{
fars_read_years(c(2013,2014,2015))

}
