% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data contents for several years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of numeric year values for which to load data.}
}
\value{
This function returns a list of data frames containing the
   columns \code{MONTH} and \code{year} for the requested years.
}
\description{
This function reads data from the FARS CSV files for the provided years,
and returns a list of data frames containing the month and year of each row
in the data for the given years.
}
\details{
FARS is the US National Traffic Safety Commission's
\href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System}.
}
\note{
This function checks that the files for the requested years exist in the working directory,
and will generate warnings for each of those files that does not exist.
}
\examples{
\dontrun{fars_read_years(c(2013, 2014))}
\dontrun{fars_read_years(list(2013, 2014))}

}
\seealso{
\code{\link{make_filename}}, \code{\link{fars_read}}
}

