% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Batch import of FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years in numeric or integer format.}
}
\value{
A list of tibbles. Each tibble contains two columns, MONTH and year.　
Returns NULL and a warning, if the file does not exist.
}
\description{
Iterate over a numeric year vector that reads FARS data for each #' year. If file does not exists, it returns a warning.
}
\examples{
\dontrun{fars_read_years(2013:2015)}
}
