% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Pulls month and year from an accident report csv}
\usage{
fars_read_years(years)
}
\value{
uses fars_read() function to create one or more
tibble data frames that matches the year of the file name
created with the make_filename() function. each tibble df
will contain two columns, month and year. In the event that
the year entered does not pull a valid data file, a warning
will be generated as well as a "NULL" value.
}
\description{
This function accepts a numberic input
of either one or multipe years, then
collects the month and year columns from the corresponding
data file. If the year or one of the years are not searchable,
the function will throw a warning and a NULL output.
}
\examples{
\dontrun{
fars_read_years(2013,2014)

fars_read_years(2018)

# This function will generate a warning
}

}
