% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Check input vector "years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years that will be checked against the make_filenames()
output.}
}
\value{
This function returns a data frame with MONTH and year variables for each
file with a valid year. If the year is not valid, then a warning is returned.
}
\description{
This function checks whether the input vector "years", contains elements that appear
in the output from make_filename() that produces file names in the format:
"accident_YYYY.csv.bz2", where YYYY represents the year.
}
\note{
Examples: fars_read_years(c(2013,2014)), fars_read_years(c(2013,2014,2015))

If the file "accident_2050.csv.bz2" does not exist, attempting to check the input
year for this file will generate an error.
try(fars_read_years(c(2013,2014,2050)))
}
