% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads data from a file for each year, subsets month and year and saves it into a list}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a single year or a vector containing multiple years}
}
\value{
This function returns a list containing month and year from each years data
}
\description{
This function reads in a vector of years. For each year, it makes a filename, reads the data from that file, adds a year column to the data, selects, month and year variables, and finally stores that into an element of list.
}
\examples{
fars_read_years(c(2001:2007))

}
