% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read and wrangle multiple Fatality Analysis Reporting System files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{(\emph{character} or (\emph{integer}) ) A vector containing the desired years. 
Both \emph{numeric} and \emph{character} vectors are allowed for \code{years}.
\code{years} is coerced to integer. If \code{years} cannot be coerced to integer,
\code{fars_read_years} will result in an error as filenames cannot be created
properly.}
}
\value{
A list containing the rearranged datasets for each given year.
}
\description{
\code{fars_read_years} reads Fatality Analysis Reporting System (FARS) files 
for multiple \code{years} and rearranges the data such that each row contains
one incident with columns \emph{MONTH} and \emph{year}.
}
\details{
\emph{list} is also allowed for \code{year} but not recommended.
Files must exist in current working directory. \cr
\code{fars_read_years} is a helper function for \code{\link{fars_summarize_years}}.
}
\examples{
\dontrun{
lst <- fars_read_years(c(2014, 2015))
lst <- fars_read_years(c("2014", "2015"))
# also usage of lists is allowed:
lst <- fars_read_years(list(2014, "2015"))
}

}
\references{
US National Highway Traffic Safety Administration \cr
(\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\seealso{
\code{\link{fars_read}}, \code{\link{make_filename}}, \code{\link{fars_summarize_years}}
}
