% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read yearly accident data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{One or several years in a vector of character string
or numeric values.}
}
\value{
This function returns the months and years of accidents.
Data is retured as yearly lists and with
classes of 'tbl_df', 'tbl' and 'data.frame'.
}
\description{
This function returns US traffic accidents data by year and month.
}
\note{
Years from 2013 to 2015 are available, function throws an error if
year(s) outside these years are selected.
}
\examples{
\dontrun{
fars_read_years(c(2014))
fars_read_years(c(2013, 2014, 2015))
}
}
\references{
US National Highway Traffic Safety Administration
}
