% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract month numbers from FARS files of specific years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{(numeric) A vector of years for which the month numbers should be returned}
}
\value{
A tbl_df with columns MONTH and year
}
\description{
This function will try to read in standard named FARS files from the
working directory using fars_read() and extract the month number of every observation.
Years for which no correspondingly named file can be found will raise an error.
}

