% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{FARS Read years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing the years of interests as \code{integer} or
\code{character}.}
}
\value{
A list containing the selected data as a data frame
table (see \code{\link[dplyr]{tbl_df}} in the \code{dplyr} package) for each provided year.
}
\description{
Reads the FARS data file(s) for the provided years in R as data frame table(s) and,
for each one of them, keeps only the \code{MONTH} and \code{year} variables.
}
\details{
For each provided year the following steps are executed:
\enumerate{
  \item Build the FARS file name from \code{year} (using \code{\link{make_filename}}).
  \item Read the FARS data file in R as a data frame table (using \code{\link{fars_read}}).
  \item Add a \code{year} variable to the data frame table.
  \item Keep \code{MONTH} and \code{year} variables only from the data frame table.
}

The data file names must be compliant with FARS pattern name. If a file does not exist
the associated data is set to \code{NULL}.
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{mutate}} and \code{\link[dplyr]{select}} in the \code{dplyr} package.
}
}

\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
fars_read_years(c("2013", "2014", "2015"))
}

}
