% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Fatality Analysis Reporting System data one or more years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of years. Uses a YYYY specification.}
}
\value{
A list including tibbles with the FARS data for each of the input
  years.
}
\description{
This function reads the Fatality Analysis Reporting System (FARS) data from
local csv files for one or multiple years. The function returns NULL if the
information is not available for a specific year.
}
\note{
The csv files with the data need to be downloaded previously from the
  US National Highway Traffic Safety Administration's Fatality Analysis
  Reporting System website
  (\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\examples{
\dontrun{fars_read_years(2010)}
\dontrun{fars_read_years(2010:2015)}

}
