% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
a dataframe of entries of the years sorted by month, or a warning if a year is invalid e.g., no file
exists for said year.
}
\description{
Reads FARS data for a list of years, sorted by month.
}
\examples{
fars_read_years(c(2013,2015))

}
\seealso{
\code{\link{fars_read}}

\code{\link{make_filename}}

\code{\link{fars_summarize_years}}
}
