% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Compute month and year information for given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers}
}
\value{
a list of tbl dataframes which have columns MONTH and year
}
\description{
Compute month and year information for given years
}
\note{
"dplyr" package is required.  If the argument "years" include a year
where no data file exist, then the function returns NULL and a
warning showing an invalid year
}
\examples{
fars_read_years(c(2013, 2014))
}
