% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create list of data frames}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years for which the data files in the working
directory need to be read and processed. If a corresponding CSV file
for a year in the vector of numeric years is not present in the working
directory, an error message indicating "invalid year" is generated and the
data frame for that year will be set to \code{NULL}}
}
\value{
A list of data frames; each data frame contains the month values from
the processed data files in the working directory and the corresponding year
specified in the \code{years} parameter.
}
\description{
The \code{fars_read_years()} function reads and processes data files in the
working directory for multiple years; it takes a vector of numeric years as
input and returns a list of processed data frames.
}
\section{External Functions Not Imported}{

\itemize{
\item \code{\link[dplyr:mutate]{dplyr::mutate()}} used to add a year column to the data frame.
\item \code{\link[dplyr:select]{dplyr::select()}} used to select the \code{Month} and \code{year} columns.
}
}

\examples{
\dontrun{
fars_read_years(c(2013,2014,2015))
fars_read_years(2013:2015)
}

}
\seealso{
\code{\link[=make_filename]{make_filename()}} to create a file name from a year
}
