% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Print "fars_read_years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A int vector with years contained in the file name}
}
\value{
This function returns a list with data contains MONTH and year
}
\description{
This is a simple function return data contains MONTH and year
}
\details{
This function fails when the file whose year is not in years
}
\examples{
\dontrun{
years <- 2013:2014
fars_read_years(years)
}
}
