% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data files.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of \code{integr}s, or other types coerced to intgers, represent the years of the \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data files.}
}
\value{
A \code{list} of \code{data.frame}s, each of its elements correspond to one of the inpute years.
}
\description{
This is a custom \code{readr_csv} function that reads a \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data file given their corresponding years.
}
\details{
The function returns a warning \code{invalid year: year} if the input year does not correspond to one of the files in the current working directory.
othewise construct a vecotr of file names using \code{\link{make_filename}} and reads them in a list using \code{\link{fars_read}}.
}
\examples{
\dontrun{
# invalid input, year doesn't exist
fars_read_years(c(2013, 2014, 1900))

# valid input
fars_read_years(2013:2015)
}

}
