% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS data for selected years and generates separate lists for each year with two columns each: MONTH and year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numerical vector representing years to read in the dataset(s)}
}
\value{
This function returns a list for each year containing two columns: MONTH and year
}
\description{
Reads FARS data for selected years and generates separate lists for each year with two columns each: MONTH and year.
}
\note{
Throws an error if the \code{year} is not a number or not a year found in the file(s)
}
\examples{
fars_read_years(2013:2015)
\dontrun{fars_read_years(not_a_number)}

}
