% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Produces a list of month,year from the read data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer/numeric vector of years, with each year in the format yyyy}
}
\value{
A list of data.frame objects with columns month, year based on the years passed
in the argument years and the data contained in the associated files for those years
}
\description{
This function takes in an argument of a vector of years with each
year in the format yyyy and generates a list containing month,year
for each month year combinations in the data from the read CSv file.
}
\note{
The files need to be in the current working directory
}
\examples{
\dontrun{
fars_read_years(2013:2015)
fars_read_years(c(2013, 2015))
}

}
