% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Subset month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years in numeric or integer format.}
}
\value{
Returns a list of tibbles (data frames) with rows as each accident events and
   columns of month and year. Returns NULL and a warning if the file does not exist.
}
\description{
This simple function subsets month and year from the file for the chosen year.
}
\examples{
\dontrun{fars_read_years(c(2014,2015))}

}
