% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Make a list of occurances year and month wise.}
\source{
This function uses the function make_filename() inside to get individual
output year wise.
}
\usage{
fars_read_years(years)
}
\arguments{
\item{Input}{class 'integer' value(s) of 'year(s)' as the one argument.}
}
\value{
This function returns an output that contains one-to-one row for
every single occurance in a MONTH within a year. Each row contains only MONTH and year
number for every occurance. This function prints a Tibble format of the output.
}
\description{
This function provides a list of occurances (accidents) in a year. For every occurance 
it adds a row in this list and records the corresponding MONTH and year in that
list for that particular occurance.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013,2015))
}

}
