% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function reads one or more data files at once. It tries to read in a
file for each month and year and then append the month and year columns.
If there is no applicable year, the function returns NULL and continues to
the next year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list containing the years of the files. Must provide the
years of data that are available.}
}
\value{
A list of dataframe(s). If year is not valid, returns NULL.
}
\description{
This function reads one or more data files at once. It tries to read in a
file for each month and year and then append the month and year columns.
If there is no applicable year, the function returns NULL and continues to
the next year.
}
\examples{
\dontrun {
fars_read_years(c(2014,2015))
fars_read_years(2016)
}

}
