% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads the month and year columns of multiple fars-data files and returns it as a list of dataframes.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integer values.}
}
\value{
A list of data.frames (tbl_df) containing the month and year columns.
}
\description{
Reads the month and year columns of multiple fars-data files and returns it as a list of dataframes.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
}

