% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years(): Read accidents data with specific year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An year of interest in which data to be read (2013-2015)}
}
\value{
A list of MONTH and year from the accidents dataset of given year
}
\description{
This function accepts a specific year and generates an accident filename of that year to be read.
If the file doesn't exist, warning msg will be displayed.
Otherwise, that file of accidents will be read and a list with only MONTH and year column will be returned
}
\examples{
\dontrun{
fars_read_years(2013)
}
}
