% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Selects all months in each file loaded (year)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector with all years that should be loaded}
}
\value{
This function returns a list with the column month of each file.
   if the year is invalid a warning will be generated to inform it.
}
\description{
This function creates a list of all months in the files.
}
\details{
This function uses mutate and select functions from dplyr package.
   It also calls make_filename function in order to create a string with the filename and
   fars_read function to load the files in a dataframe of all years specified in the parameter.
   In case the year required doesn't have a correspondent file it will generate a warning.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
fars_read_years(c(2014))
fars_read_years(c(2013, 2015))
}

}
