% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load data from accident files given a list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of strings or integers giving the years to retrieve}
}
\value{
a list of data_tbl with columns named MONTH and year and a row for each accident.
}
\description{
This is a function that loads data from highway accident files with names in a specific format
and returns a data_tbl with the month and year of each accident.
}
\details{
The highway accident files for the years given in \code{years} must be in the current working directory or an error is thrown.
}
\examples{
fars_read_years(c(1984,1986,1988))
fars_read_years(1984:1988)
fars_read_years(c("1984","1986","1988"))

}
