% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in FARS data for a specified range of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A year or a vector of years (in any class that can be coerced to numeric)
to specify which FARS reports are to be read in}
}
\value{
This function returns a list of FARS data where every list entry contains the data
        for one of the years specified in form of a \code{tbl_df}. The data frame has one
        line for every data point in the respective FARS data set and indicates the month
        and the year for this data point.\cr
        It produces an error warning if an invalid year is given as input.
}
\description{
The \code{fars_read_years()} function is used to read in FARS data for a specified range of years.
}
\details{
The function internally uses the \code{\link{make_filename}} function to create standardised file
        names from the \code{years} input. It then reads in CSV files from the working directory that
        correspond to this standardised file name and uses \code{dplyr} functions to reorganise the
        data.
}
\examples{
\dontrun{fars_read_years(c(2014, 2015)) #Read in FARS data for 2014 and 2015}

}
