% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads month and year from accident files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years in numeric or integer format.}
}
\value{
a list of tibbles with the same number of rows as the data in
"accident_<year>.csv.bz2" files and two columns - MONTH and
year. NULL value if year is invalid (also a warning will be raised).
}
\description{
This function returns a list of data frames with MONTH and year columns
based on data in "accident_<year>.csv.bz2 files.
The files need to be located in the working directory.
}
\examples{
\dontrun{
fars_read_years(2013:2014)
fars_read_years(list(2013, 2014))

# Results in a warning
fars_read_years(2016)
}

}
