% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files with specific years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing multiple character strings}
}
\value{
This function returns a list of dataframes with specific years.
}
\description{
This function read csv file with specific years into a list of dataframe.
If the input year doesn't exist, then a warning information "invalid year" will be printed
and return NULL
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
}

}
