% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_W4.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Import data of the years requested}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years of the data requested}
}
\value{
A list of tibbles with the month and year columns requested
}
\description{
This function accepts a vector of years of the data requested and
creates a list of tibbles with the month and year columns based
on the input years.
}
\note{
An error will occur if the file does not exist in the working directory.
}
\examples{
\dontrun{
  d <- fars_read_years(c(2013,2014,2015))
}
}
