% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple FARS dataset}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years for which FARS data has to be read}
}
\value{
list of datasets for each year having the MONTH and year column. The count of MONTH represents the fatalities in that month.
}
\description{
Reads the data for the given years. If the dataset corresponding to a year is not found,
the function creates a warning and returns null
}
\details{
Error if file for the given year not found in the FARS dataset
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
}
}
