% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return FARS data for a set of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numeric data convertible to four-digits integers
defining the years of interest. The years should belong to the interval [2013, 2015],
otherwise, an exception is trown.}
}
\value{
Returns a tibble containing FARS data for the chosen years (a row per an accident).
 The tibble returns only two columns:
\itemize {
\item \code{MONTH} {The month of an accident}
\item \code{year} {The year of an accident}}
}
\description{
This is an internal function. It returns FARS data for a set of years (defined using \code{years} parameter).
For more information about FARS data, see
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}.
}
\examples{
accidents <- fars_read_years(c(2013, 2014))
fars_read_years(2015)
}
