% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric or character vector contining the year(s) for which
the data should be read in.}
}
\value{
This function returns a list of length that equals the length(number of years)
   of the input vector. Each element of the list contains a data frame with data
   of the corresponding year.
}
\description{
A helper (not exported) function used by \code{fars_summarize_years} Function.
It reads in data for all the given years. The function selects the year and month
from every observation in the dataset.
If an invalid year is given the function generates a warning.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years("2013")
fars_read_years(c(2013, 2014))
fars_read_years(c("2013", "2014"))
}

}
