% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Generate a dataframe with months and years for a provided series of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers}
}
\value{
This function returns a dataframe with for each year the month numbers in the datafile.
}
\description{
This function reads multiple csv files for a given series of years, to provide an overview of months and years in those files.
}
\details{
If a year is provided for which no data file is found, the function returns an error message.
}
\examples{
\dontrun{
fars_read_years(2017)
fars_read_years(c(2013, 2014, 2015))
}

}
