% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Import yearly accident csv files}
\usage{
fars_read_years(years, demo = FALSE)
}
\arguments{
\item{years}{a character/numeric string, vector or list of years}

\item{demo}{logical if you want to use the function with demo data that comes with the package.
Default to FALSE}
}
\value{
a list. Each element of the list could be a tibble if the function finds a csv corresponding to
the year, or NULL if the csv file is not present.
}
\description{
This function takes one or more years as input, imports the related csv file(s)
creates a new column called year and returns MONTH and year variables as a tibble.
}
\examples{
\dontrun{acc_list <- fars_read_years(2013:2015)
acc <- fars_read_years("2015")}

}
