% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read individual files (each year)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{is an integer or list or vector representing year(s)}
}
\value{
data frames for each year-file containing month and year columns

if year is invalid (no existing file) returns warning
}
\description{
Read individual files (each year)
}
\examples{
\dontrun{
 fars_read_years (2013)
 fars_read_years (2013:2015)
 fars_read_years (c(2013,2015))
}
}
