% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return list of month and year dataframes for Fatality Analysis Reporting System}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years in integer, numeric or number character form}
}
\value{
a list of dataframe where each dataframe contains month and year
for valid year and NULL for invalid year.
}
\description{
This function takes a vector of years, create the file name using
\code{make_filename} function and then read in the data using \code{fars_read}
It uses lapply to run the loop through all the years. A warning will print
if there exists an invalid year.
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}
}
\seealso{
\code{\link{make_filename}} , \code{\link{fars_read}}
\code{lapply} , \code{tryCatch}
}
