% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Helper Function}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer or a vector of integers giving the year(s) the user is interested in.}
}
\value{
This function returns a list of tibbles. Each tibble corresponds to one year and contains
the months and year of each observation in the original data table.
}
\description{
This is a helper function used by the \code{fars_summarize_years} function to summarize data.
}
\note{
Not exported to the user.
}
