% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Define 5 functions for export : need to add export before each function}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of years}
}
\description{
\code{fars_read(filename)} read csv file with name filename
\code{make_filename(year)} make csv file of the year
\code{fars_read_years(years)} make csv file of years
\code{fars_summarize_years(years)} summarise data of years.
\code{fars_map_state(state.num, year)} plot map of the state.num
}
\examples{
\dontrun{
fars_map_state(18, 2015)
}

}
