% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{\code{fars_read_years} takes in a vector of 'years' specified to create and return a list of dataframes
with only the columns 'YEAR' and 'MONTH'. If one of the name of the file created by \code{make_filename} function does not exist,
a warning message \code{invalid year: xxxx} will be returned.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of years to read in}
}
\value{
returns a list of dataframes with colums 'MONTH' and 'YEAR', NULL when an error is encountered.
}
\description{
\code{fars_read_years} takes in a vector of 'years' specified to create and return a list of dataframes
with only the columns 'YEAR' and 'MONTH'. If one of the name of the file created by \code{make_filename} function does not exist,
a warning message \code{invalid year: xxxx} will be returned.
}
\examples{
\dontrun{
far_read_years(c(2011, 2012))
}
}
