% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get month and year variables from the NHTSA FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of the years of observations to read.}
}
\value{
Returns a list of tibbles consisting of two columns: \code{month} and
  \code{years}.
}
\description{
Reads table data from one or more Fatality Analysis Reporting System (FARS) data files
containing observations from the specified year(s) and returns month and year for each observation.

Will warn if any of the specified years is invalid.

Year value is taken from the parameter not from the data file.
}
\details{
Returns a warning and NULL value if the file with \code{year} does
  not exist.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
