% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data for a given set of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector specifying years of interest}
}
\value{
This function returns a list of data.frames containing accident data.
Each element of a list contains data for one year.
}
\description{
This function reads data for a set of years specified in the \code{years}
argument. \code{years} must be numeric or be possible to be coerced to numeric,
otherwise an error will occur.
\code{fars_read_years} is dependent on \code{dplyr} package.
}
\examples{
\dontrun{fars_read_years(c(2013,2015))}

}
