% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read a list of years' accident files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of numeric values standing for years}
}
\value{
A list of data frames with two columns in each of them, MONTH and year
}
\description{
Read in a list of different years' accident files, add year info as a column
for each data frame, and choose MONTH and year columns for futher analysis,
if the input of year is invalid, it will give a message and return no result.
}
\details{
If years is not invalid number for years, a warning will be issused.
}
