% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read csv file with given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of list of year to read in}
}
\value{
A date table including month for all entry in data, or NULL if the given year is not available
}
\description{
Read the FARS data from disk for input year and returen a list of data frames
}
\examples{
\dontrun{fars_read_years(2013)}

}
\seealso{
\link{fars_read}

\link{make_filename}
}
