% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data from multiple csv files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector specifying a selection of years in 4-digit numbers.}
}
\value{
A list of tibble objects. Each object contains the data from
    the corresponding file of that year.
}
\description{
This function imports the data from multiple files, specified by the input
years, and generates a list of objects indexed by the input value. Each
object can be either a tibble of two columns (year and MONTH representing
the time of an accident), when the the file exits, or NULL, when the data
file is unavailable. In the latter case, a warning message will be printed
to the console.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013, 2014, 2015))
}
}
