% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years Function}
\usage{
fars_read_years(years)
}
\arguments{
\item{A}{vector with all the years you want to study}
}
\value{
a data list with objects the data from all the years you are interested in.
}
\description{
This is a function that gathers the data cancerning all the years you give as input.
}
\note{
An error occur when there are no data for a year that you asked'
}
