% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{numeric vector that specifies the years to read the data}
}
\value{
This function returns a list of data frames
}
\description{
This is a function that reads FARS data for one or more years
specified by \code{years} argument and returnes a list of corresponding data frames
(one data frame by year)
}
\details{
Each year in separate data frame in the same list
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{mutate}} and \code{\link[dplyr]{select}} in the \code{dplyr} package.
}

#' @examples
\dontrun{
fars_read_years(c(2014, 2015))
fars_read_years(2013:2015)
fars_read_years(list(2013, 2014, 2015))
}
}

