% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_fars_fun.r
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years for filenames to be read into R.}
}
\value{
fars_read_years returns a list of tibbles with year and month as columns
}
\description{
This function iterates over a list of years provided by user, first making a filename to
read from the folder, trying to read the file and if successful, creating a year variable
equal to the element, and finishing by selecting the month and year. If unable to read
file, throws back an error.
}
\note{
Error is returned if file is not valid name, which is due to invalid year.
}
