% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Gives number of data points in each year given}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{A vector with years}
}
\value{
This function returns dataframe with years in column and number of rows of data as values
}
\description{
This function reads all files with years given and gives number of rows in each year
}
\details{
if there is no file with year prefixed after accident_ and suffixed with .csv.bz2, it will give error
}
\examples{
\dontrun{
fars_summarize_years("2021")
fars_summarize_years(c(2021,2022))
}

}
