% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise FARs data for years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{This is a list/vector of years in YYYY format}
}
\value{
This function returns a tibble containing the number
    of observations in the given FARs years for each month
    in the data. The rows are the months and the columns are
    the given years
}
\description{
This function reads in and then summarises the FARs data into
number of observations by month and year
}
\examples{
\dontrun{
fars_summarize_years(c("2013", "2014", "2015"))
}
}
\author{
John J Smith, \email{fakeemail123@duvman.ie}
}
\keyword{FARS}
