% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years: Returns a tibble of complete observations where
the year is included as a column. Data from multiple years is expressed
in one tibble with the fatalities summarized by month and year over the
years supplied.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of the years for the required FARS data.}
}
\value{
A tibble with the number of accidents each year summarized by month.
}
\description{
fars_summarize_years: Returns a tibble of complete observations where
the year is included as a column. Data from multiple years is expressed
in one tibble with the fatalities summarized by month and year over the
years supplied.
}
\examples{
\dontrun{plot(fars_summarize_years(2014)}
\dontrun{fars_summarize_years(yearsVector)}
}
