% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years
fars_summarize_years Read the data from the specified years and calculate how many observations were for each month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Year or range of years from which we want to extract the information}
}
\value{
A tibble with the number of observations per month for each year
}
\description{
fars_summarize_years
fars_summarize_years Read the data from the specified years and calculate how many observations were for each month
}
\examples{
\dontrun{
summary_data<-fars_summarize_years(2013:2015)
}

}
