% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize  for specific years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer o chracter string.}
}
\value{
This function returns a  object of the tbl_df class that the table with counts.
}
\description{
This function create summary table with counts of number of records for each month in a specified years.
}

