% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize records per month for a series of years provided}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers representing a series of years}
}
\value{
returns a tibble (data.frame) from a series of years (\code{years}) as input and with the number of records per month, by year
}
\description{
This function generates a table showing when accidents occur with data for month and year
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)
}
}
