% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_w2.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of years to be evaluated.}
}
\value{
Dataframe with summary
}
\description{
Returns a summary of the number of entries on each month and year
}
\note{
This functions depends on the bind_rows, group_by, summarize functions from dplyr and spread from tidyr
}
\examples{
\dontrun{fars_summarize_years(c("2013", "2014"))}

}
