% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Quick summarization of FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector containing the years of data. Data files should be in the
same directory.}
}
\value{
A summarization data frame or a \code{NULL}.
}
\description{
The function takes user specified years and summarize the in the data frame format.
}
\note{
You must include tidyr and dplyr packages.
The data files should be named "accident_X.csv.bz2" and be in the same directory.
}
\examples{
\dontrun{fars_summarize_years( c('2014') )
fars_summarize_years( c(2013,2014) )}

}
