% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Print "Fars Summarize Years"}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of one or more character string years for the desired Fars data.}
}
\value{
This function returns a single data frame with the number of the records per month
and year that has been tidied so that each year's records is shown in a column.
}
\description{
This function returns a summary view of the number of records per year and month
of the Fars data for all years passed to it.
}
\note{
Invalid years will result in a warning message with nothing returned. If working directory
has not been set, a warning message will result for valid file names.
}
\examples{
 \dontrun{
fars_summarize_years(c('2014','2015'))
fars_summarize_years(years = c('2014','2015'))
}

}
