% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a summary table with the counts of car fatalities.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector representing the years of a time period.}
}
\value{
A \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
   object with the counts of car fatalities by month and year.
}
\description{
This function creates a summary table with the counts of car fatalities reported
by the US National Highway Traffic Safety Administration for a specific period.
The data are shown seperately for each year in the period.
}
\section{Warning}{

   If a file does not exist for a specific year, the function will
   throw a warning that the information cannot be accessed.
}

\examples{
# Summary with the fatality counts in 2013.
fars_summarize_years(2013)
# Summary with the fatality counts in 2013 and 2014.
fars_summarize_years(c(2013, 2014))
}
