% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Calculate a number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list (numeric or integer) with the years of accident records.}
}
\value{
This function returns a tibble (data.frame) with number of accidents per month and year
from the indicated years of records.
}
\description{
This function accepts a vector or list of years \code{years}, where the records from US
National Highway Traffic Safety Administration's Fatality Analysis Reporting System should
be available. It calculates a number of accidents per month and year and returns a tibble
(data.frame) with months in rows and years in columns.
}
\note{
This function will give error if the files are not located in the working directory or
if the input \code{year} is non-numeric or non-integer.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)

}

