% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector or list of years}
}
\value{
Function \code{fars_summarize_year} returns a data frame with the total numbers of
fatal incidences by \code{month}, in each \code{year}.
}
\description{
Function \code{fars_summarize_years} summarizes
the number of fatal accidents in each \code{year} by \code{MONTH}.
}
\note{
If a element \code{year} in argument \code{years} is invalid,
function \code{fars_read_years} returns \code{NULL}
and be reported as error in the invalid \code{year}.
If \code{year} is not integer, it would be transfered to \code{\link{as.integer}}.
}
\examples{
fars_summarize_years(c(2013,2014))
fars_summarize_years(2013)

}
