% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise grouped accident data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vector containing four-digit numeric values.}
}
\value{
A tibble
}
\description{
\code{fars_summarize_years} takes a numeric vector of four-digit name of years and
 creates a single tibble containing data from various years.
}
\details{
\code{fars_summarize_years} takes a numeric vector of four-digit name of
 years. It calls to \code{fars_read_years} to create a list of tibbles containing data
 for the corresponding years. After that it calls to \code{\link[dplyr:bind_rows]}
 to attach row-wise the elements of the list in order to create one single tibble.
 It subsequently groupes the data base on \code{year} and \code{MONTH} and counts the
 number of observations in group by means of \code{\link[dplyr:group_by]} and
 \code{\link[dplyr"summarise]}. The summarised data will be reshaped using
 \code{\link[dplyr:spread]} in order to provide the number of observation and their
 respective years.
}
\examples{
 fars_summarize_years(c(2013, 2014, 2015))

}
