% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize grouped accident data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vector containing four-digit numeric values.}
}
\value{
A tibble
}
\description{
This function takes a numeric vector of four-digit name of years and
 creates a single tibble containing data from various years.
}
\details{
This function takes a numeric vector of four-digit name of
 years. It calls to \code{fars_read_years} to create a list of tibbles containing data
 for the corresponding years. After that it calls to \code{\link[dplyr]{bind_rows}}
 to attach row-wise the elements of the list in order to create one single tibble.
 It subsequently groups the data base on year and month and summarizes the
 number of fatalities in every month per year.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
}

}
