% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function 'fars_summarize_years'}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years (a vector consisting of numbers with 4 digits)}
}
\value{
This function returns a key-value pair across multiple columns
   (with rows per month and columns per year)
}
\description{
The function returns a data frame with the number of accidents per month and year.
   The dimensions are the months per row and years per column.
}
\examples{
yyyys <- c(2014, 2015)
fars_summarize_years(years = yyyys)

}
