% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize record counts by year}
\usage{
fars_summarize_years(years)
}
\value{
This function returns a tibble with summaries of record counts by month for each Year of interest
}
\description{
Summarize record counts by year
}
\examples{
fars<-fars_summarize_years(c(2012,2013,2014))

}
