% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize record counts by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integer years, if data for any years don't exist, no data will be returned}
}
\value{
This function returns a tibble with summaries of record counts by month for each Year of interest
}
\description{
Summarize record counts by year
}
\examples{
\dontrun{fars<-fars_summarize_years(c(2012,2013,2014))}

}
