% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize number per month in a specified year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The input year for which to summarize the data}
}
\value{
This function returns a summary of the data grouped by months in
   a given year.
}
\description{
This function summarizes "FARS" data per month in a given year, that
is specifid in the input
}
\examples{
\dontrun{fars_summarize_years(2013)}

}
