% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Counts number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Looks for a vector or list of years (numeric or integer) in the data
files.}
}
\value{
Returns a pivot tibble (data frame) with months in rows and selected
years in columns containing the number of accidents. Returns a warning for
every input year that does not exist in the datasets. Returns an error medssage if
non-numeric or non-integer input is presented.
}
\description{
Calculates the number of accidents based on the list of years by months.
Accepts years as a list or a vector.
}
\examples{
\dontrun{
fars_summarize_years(2013:2016)
}

}
