% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the FARS Dataset}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list with numerical values depicting the years of interest in FARS dataset}
}
\value{
The function returns a wide tibble object with the result of total number of events in year and month
}
\description{
This function returns a tibble dataset of multi-years labeled FARS dataset
}
\examples{
\dontrun {years <- c(2006:2010)}
\dontrun {fars_summarize_years(years)}
}
