% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a summary of number of fars records per month for each year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years:}{A vector of integer years}
}
\value{
A data frame of the counts of number of fars records per year.
Months are numbered between 1 and 12 (both ends included) and the counts
are present under the years column. Invalid year entries are left out of the summary.
}
\description{
Creates a summary of number of fars records per month for each year.
}
\examples{
> ret <- fars_summarize_years(c(2013, 2015, 2021))
Warning message:
  In value[[3L]](cond) : invalid year: 2021
> ret
# A tibble: 12 x 3
MONTH `2013` `2015`
* <int>  <int>  <int>
  1     1   2230   2368
2     2   1952   1968
3     3   2356   2385
4     4   2300   2430
5     5   2532   2847
6     6   2692   2765
7     7   2660   2998
8     8   2899   3016
9     9   2741   2865
10    10   2768   3019
11    11   2615   2724
12    12   2457   2781

}
