% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years (given as strings)}
}
\value{
A dataframe with number of accidents by years summarized by month
}
\description{
This function summarizes  accident data, by year and month.
}
\examples{
\dontrun{
plot(fars_summarize_years(2015))
fars_summarize_years(c(2015, 2014))
}

}
\seealso{
\link{fars_read_years}
}
