% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS incidents by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years for which to process the FARS data.}
}
\value{
A tibble with rows for MONTH and each year with the total number of accidents. Returns a warning for every
    year for which no data exists.
}
\description{
Summarizes the number of incidents in a year by month
}
\examples{
\dontrun{
fars_summarize_years(2013:2014)
}
}
