% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Table of Yearly by Monthly Accident Data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{\code{integers}}
}
\value{
Creates a \code{tibble} of where columns is years and where rows represents months of the data
}
\description{
Summarizes observations by months for the user-defined years in a table
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}
