% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count the number of accidents per month}
\usage{
fars_summarize_years(years, path)
}
\arguments{
\item{years}{a list of values corresponding to year}

\item{path}{directory containing the Fars data set to "data" by default}
}
\value{
a data frame table with columns
\itemize{
\item \code{month} : month number
\item \code{year} : four digit year
\item \code{n} : the number of corresponding rows in the dataset interpreted as the number of accidents
}
}
\description{
Using a list of years as input, this function reads the collect the available data in the
Fatality Analysis Reporting System dataset. It then counts the number of accident per month.
}
\details{
Uses \code{fars_read_years} then transforms the list into a data.frame.
The data is aggregated by the MONTH and year columns and summerize with the count \code{n()} function.
}
\note{
\itemize{
\item only years present in the dataset will be represented
\item this function require to run from the directory contianing the datasets
}
}
\examples{
\dontrun{
# need to be in the same directory as the data
setwd("data")
# with one year
fars_summarize_years(2013)
# with multiple years
fars_summarize_years(2013:2015)
# with characters
fars_summarize_years(c("2013","2015"))
}

}
\seealso{
fars_read_years
}
