% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years - Summarize the number of accidents per month by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years (strings)}
}
\value{
A dataframe summary table that contains the number of accidents grouped
   by year and month of the given list of years
}
\description{
This function summarizes yearly accidents data, by year and counts the number
of reports per month
}
\examples{
\dontrun{
plot(fars_summarize_years(2013))
fars_summarize_years("2014")
fars_summarize_years(c(2013,2014))
fars_summarize_years(2013:2015)
fars_summarize_years(list(2013, 2014, 2015))
}

}
\seealso{
\link{fars_read_years}
}
