% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data set by year and each month in year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character or integer vector of years}
}
\value{
This function returns a dataframe.
}
\description{
fars_summarize_years is used to create a summary of the number of data points per
year and each month in the year.
}
\details{
The function creates a summary dataframe of number of data points in the FARS data set
summarized per year and each month
}
\note{
Error message:If an incorrect year integer or string is entered then an error is thrown
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))
}
}

