% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function \code{fars_summarize_years}}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{numeric the year}
}
\value{
A data frame with two columns, representing the months and the number of traffic deaths for each month
}
\description{
This functions calculates the number of traffic deaths for each month during a given year
Will return with a warning if data for the specified year not can be found
}
\examples{
fars_summarize_years(2015)

}
