% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data for multiple years}
\usage{
fars_summarize_years(years)
}
\value{
A dataframe with the number of fatal injuries suffered in motor vehicle traffic
  crashes by month with each year in its own column. If no valid years are found, the function
  will give an error.
}
\description{
This function can be used to get a data frame of FARS data for
  multiple years. Summarizes the number of fatal injuries suffered in motor vehicle traffic
  crashes by month with each year in its own column.
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{group_by}} and \code{\link[dplyr]{summarize}} in the
  \code{dplyr} package.
  \item \code{\link[tidyr]{spread}} in the \code{tidyr} package.
}
}

\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years(c(2013:2015))
fars_summarize_years(list(2013, 2014, 2015))
}

}
