% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS Summarize Years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character vector of years which will be cycled through
to form the summarized data set}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function that summarises the list of tables that result
from running the function fars_read_years
}

