% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of years}
}
\description{
\code{fars_summarize_years} returns a data frame with number of total accident under
each month in different years
}
\examples{
fars_summarize_years(2013:2015)

}
\seealso{
\code{\link{fars_read_years}}
}
