% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{summarize by fars year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vectors for years}
}
\value{
a wide tibble. one column for year and all other columns are month. Values are the count of oberservations by month in a year row
}
\description{
Count of number of observartion by month in the given years
}
\examples{
\dontrun{fars_summarize_years(2013:2015)}

}
