% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years - read files with all the years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{List of years whcih the files will be read.}
}
\value{
a \code{\link{dplyr}} \code{\link{tbl_df}} which summarize
   number of accident in each the month in each year specified.
}
\description{
This function takes a list of years of the accidents,
read all the coorisponding files and summarize how many
accients in each month in each year specified.
}
\examples{
\dontrun{recent_accidents_report <- fars_summarize_years("2015")}
\dontrun{fars_summarize_years(c(2013,2014))}

}
