% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Given a list or vector of years, return the number of observations in the FARS data for all the given years and
the months in these years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector of values that can be coerced to integers and that each represents a year.}
}
\value{
A tibble containing the number of observations in the FARS data for the given \code{years} at a month granularity.
}
\description{
Given a list or vector of years, return the number of observations in the FARS data for all the given years and
the months in these years.
}
\note{
The elements in \code{years} must be castable to integers.
}
\examples{
\dontrun{
fars_summarize_years(list(2014))
fars_summarize_years(list(2013,2014,2015))
}
}
