% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Monthly Accident Number for each year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Collection of years as vector or list where each element is in integer format or can be converted with as.integer}
}
\value{
This function returns a data frame with dimension (12,length(years)).
}
\description{
For a given collection of years this function reads in the accident data and then returning a summary of the data
which contains one row for each calendar month containing the number of accidents in that month for different years
in seperate columns
}
\examples{
fars_summarize_years(c(1991,2006))
fars_summarize_years(list(1991, "2006"))
}
