% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS Year Summarize}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character vector, numeric vector, or integer vector containing
the years for which data will be read and summarized.}
}
\value{
A data frame summarizing total number of fatalities by year and month.
}
\description{
This function takes a vector of years, and returns a single data frame that
 summarizes each year by total fatalities in each month.
It uses functions from the 'magrittr' and 'dplyr' package, as well as the
 function 'fars_summarize_years' defined earlier.
}
\note{
Inputting years for which there is no data, or years that are improperly
 formatted, will result in errors.
}
\examples{

\dontrun{fars_summarize_years(c(2013,2014,2015))}
\dontrun{fars_summarize_years(c("2013","2014"))}

}
