% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read multiple years of data and summarize.
This function will take a vector of years and return a summary of the data by month and year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
If the years are valid, the function will return a summary of the total events by month and year. Otherwise the function will return null.
}
\description{
Read multiple years of data and summarize.
This function will take a vector of years and return a summary of the data by month and year.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014,2015))
example_years <- 2013:2015
fars_summarize_years(example_years)
}
}
