% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise the data for the list of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{List or vector of the years to be extracted from the datafile.}
}
\value{
This function returns a data frame. The data frame has 12 rows, each month has a row.
The data frame has one column for the month plus each year from the  \code{years} argument
has its own column.  specific element in a row for a given year contains the number of
accidents from the sample data.
}
\description{
This function create a summary for some years from the input files.
The function uses \code{fars_read_years()} to collect the data.
The summarized data is a non-tidy table of the number of accidents for a give month and year.
}
\examples{
\dontrun{
fars_summarize_years(years=c(2014,2015))
}

}
