% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of accident counts by years and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Desired range of years as numeric input.}
}
\value{
Returns a summary dataframe by year (column) and month (row) with a summary numeric count of accidents.
}
\description{
This function creates a simple tabular summary of accident counts by year and month.
}
\details{
Uses the dplyr function to create a file of years input then summarized and formatted by year, month and accident count. Tidyr is used to format using spread.

Uses the function far_read_years
}
