% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function summarizes fars data, group by year and month, and finally spread the summaries over year and count}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years for which data will be loaded}
}
\value{
This function returns summaries data as a dataframe
}
\description{
This function summarizes fars data, group by year and month, and finally spread the summaries over year and count
}
\details{
This function raise errors when external library is not installed or invoked using library() function
}
\examples{
\dontrun{
summary <- fars_summarize_years(c(2013,2014))
}

}
