% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes multiple files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of numbers or character string with numbers}
}
\value{
a table containing the number of casualties with months as rows and years as columns
}
\description{
From given years in numbers, this function makes one dataframe containing data from all years.
From this data frame a table is made from the number of casualties with months as rows and years as columns
}
\examples{
fars_summarize_years(c(2013,2014))

}
