% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise FARS data for given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of years}
}
\value{
data.frame with summarized data for every month and year
}
\description{
Summarise FARS data for given years
}
