% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generate a summary of the available data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years for which data is desired.}
}
\value{
A single tibble summarizing the available data for each of the months in the provided vector of years
}
\description{
Generate a summary of the available data.
}
\section{Errors}{

This function will throw an error if an unavailable or invalid year is encountered.
}

\examples{
#Basic usage:
fars_summarize_years(years = c(2013, 2014, 2015))
}
