% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the data of each year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of the name of multiple years}
}
\value{
a tibble of the summarized data
}
\description{
Summarize the number of accident in each year
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}

