% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes your data per year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{numerical variable - can be a list}
}
\value{
a tibble where aach column represents the year you specified with the number of observations per
specific month.
}
\description{
This function opens your files regarding the specific year and also does summarization. It
will create one file using all files you have given the function  by simply binding the rows together.
Then it groups the data by year and month and counts observations per group (month of a year).
In the end it will give you the summarized data. Each column will represent the year you specified with the number of observations per
specific month.
}
\details{
It uses function 'spread' from tidyr package to describe the data per year.  If you submit a non-existing year, the same errors as for fars_read_years function
will appear.
}
\examples{

fars_summarize_years(list(2013))

}
