% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{print("fars_summarize_years")
This function returns a tibble with the monthly summary of the data from the
US National Highway Traffic Safety Administration's Fatality Analysis
Reporting System (FARS) for each of the years requested as a vector with an
argument passed onto the function, which can be done using the \code{years} argument.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the year for which the summary table will be constructed.}
}
\value{
This function returns a tibble containing a monthly summary for each
   of the requested years of the FARS database.
}
\description{
print("fars_summarize_years")
This function returns a tibble with the monthly summary of the data from the
US National Highway Traffic Safety Administration's Fatality Analysis
Reporting System (FARS) for each of the years requested as a vector with an
argument passed onto the function, which can be done using the \code{years} argument.
}
\examples{
\dontrun{
fars_summarize_years(2015)
fars_summarize_years(years = c(2013:2015))
}

}
