% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize multiple years of motor vehicle crash data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A string or integer vector giving the years of motor vehicle crash files to read in.}
}
\value{
This function returns a data frame containing the number of motor vehicle crashes for each year and month.
	An error message occurs if the there is no data file for a corresponding input year.
}
\description{
This function (fars_summarize_years) accepts a vector of years or a single year as input.
Motor vehicle crash data for each year in the input vector is read using the fars_read_years function from this package (\code{\link{fars_read_years}}).
The list of dataframes are then merged using the bind_rows function from the dplyr package (\code{\link[dplyr]{bind_rows}}).
The resulting dataframe is grouped by the year and MONTH columns using the group_by function from the dplyr package (\code{\link[dplyr]{group_by}}).
The total number of records relating to each group is then calculated using the summarize function from the dplyr package (\code{\link[dplyr]{summarize}}).
The results are represented with months displayed along the rows and years along the columns by using the spread function from the tidyr package (\code{\link[tidyr]{spread}}).
An error message occurs if the there is no data file for a corresponding input year.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013, 2014))
fars_summarize_years("2013")
fars_summarize_years(c("2013", "2014"))
fars_summarize_years(years=2013)
}

}
