% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data for years provided}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A \code{numeric} vector of years for which data is to be summarized}
}
\value{
Summarized data in a data frame table(\code{tbl_df})
}
\description{
This function summarizes the list of data frames obtained using (\code{fars_read_years}).
The information about the number of cases per year every month is returned.
}
\examples{
\dontrun{
fars_summarize_years(2010:2012)
}

}
