% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years - summarize statistical data about accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{- vector of years}
}
\value{
data frame with summary about accidents in the given years: quantity of accidents per month for each year. If data for any of given years do not exist
function stops working and return an error message.
}
\description{
Function takes given vector of years and summarize statistical data about accidents
}
\examples{
fars_summarize_years(c(2013, 2014, 2015))

}
