% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_original.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of accidents per months/year(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{One or more years, numeric integer value specified in form of a list.}
}
\value{
This function returns a summary of the number of accidents by month and for each specified year.
   It will return a tibble object that have all these classes: 'tbl_df', 'tbl' and 'data.frame'.
   One row for each month, the month number in the first column (MONTH) and for each specified year a column with
   named from the year (`yyyy`) containing for each month the number of accidents.
   If one or more specified years is not valid, the function return an error and specify the invalid year.
}
\description{
This function returns a summary of the number of accidents by month and for each specified year.
}
\examples{
\dontrun{
summarize_years<-fars_summarize_years(years=2013:2015)
}

}
