% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Title Function for summarizing the data by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector with list of years}
}
\value{
A dataframe with the total monthly number of occurrence in each year
}
\description{
Title Function for summarizing the data by years
}
\examples{
fars_summarize_years(c(2013, 2014, 2015, 2018))
}
