% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector with desired years in format YYYY}
}
\value{
A data frame with the amount of accidents for year by month
}
\description{
Extract the total amount of accidents by year group by months
}
\details{
Print "fars_summarize_years"
}
\examples{
fars_read_years(c(2013,2014,2015))
return a data frame like this:
   MONTH `2013` `2014` `2015`
*  <int>  <int>  <int>  <int>
1      1   2230   2168   2368
2      2   1952   1893   1968
3      3   2356   2245   2385
4      4   2300   2308   2430
5      5   2532   2596   2847
6      6   2692   2583   2765
7      7   2660   2696   2998
8      8   2899   2800   3016
9      9   2741   2618   2865
10    10   2768   2831   3019
11    11   2615   2714   2724
12    12   2457   2604   2781

}

