% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector with desired years in format YYYY}
}
\value{
A data frame with the amount of accidents for year by month
}
\description{
Extract the total amount of accidents by year group by months
}
\details{
Print "fars_summarize_years"
}

