% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{summarise of contents of one or more fars Files.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{One or more years as integer or string}
}
\value{
This function returns a data frame containing summary information
}
\description{
\code{fars_summarize_years} create a summary of FARS years and
  months
}
\details{
This function takes an input years, and produces a summary table that
contains the number of observations for each month/year
}
\note{
If the file for a specified year, or permission to it, does not exist, an error is produced.
}
\examples{
\dontrun{
fars_summarize_years(years = 2013)
}

\dontrun{
fars_summarize_years(years = 1066)
}

}
