% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary one or more years' dataset}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years which you need to load.}
}
\value{
This function returns a dataframe of which the column is the year you chose, the row is month 1-12
, and the value means the event happened.
}
\description{
This function read one or more years' data (only have 2013-2015, other values woule result in an error) and
return a summary of events hapened each month in the year correspondingly
}
\examples{
\dontrun{s.df=fars_summarize_years(c(2013,2014))}
\dontrun{head(s.df)}

}
