% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count the total number of months for the year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years you are interested}
}
\value{
A data frame containing the number of observations that are in certain month and year
}
\description{
Returns a data frame containing the total number of observations in each month for each year. It counted all observations in the data.
}
\details{
If you entered a year that is not valid, there will be a warning.
}
\examples{
fars_summarize_years(2013:2014)
}
