% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes accident data for a set of years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector of values representing years.}
}
\value{
A tibble with the number of accidents per year per month.
}
\description{
Summarizes and tidies the data returned by \code{fars_read_years} by
counting the accident numbers per year and month.
}
\examples{
fars_summarize_years(c(2013, 2015))

}
