% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years function}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of numbers specifying the years of interest}
}
\value{
this function returns a summary table of total number of observations by year and month
}
\description{
this is a function that combines data of different years and then
creates a summary table of total number of observations by year and month
}
\examples{
\dontrun{fars_summarize_years(2013)}

}
