% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of years' FARS files to open.}
}
\value{
A data.frame with number of accidents by years summarized by month.
}
\description{
This function summarizes yearly accidents data, by month.
}
\examples{
fars_summarize_years(years = c(2014, 2015))

}
