% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data files by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
The function returns an object of classes "tbl_df", "tbl", "data.frame".
The object is a data frame with twelve rows (one for each month) and n+1 variables.
The first variable is the MONTH variable, and it is followed by n year variables,
where n is the number of different years occurring in the tables which were read.
}
\description{
The function reads data by years, binds them by rows, groups by years, and months,
and summarizes for each month of each year.
}
\details{
The function will give an error fars_read_years() would give an error.
}
\examples{
setwd("./data")
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))
setwd("../")
}
