% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function to reads multiple files based on years parameter and sumarizes data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Valid years as a vector (atomic or list) or an expression object.}
}
\value{
File data is read and sumarized into a dataframe object by MONTH and year.
}
\description{
This function takes the years parameter.
It reads multiple files and returns dataframe object.
It outputs a warning message if it cannot find a file for a year in the years parameter. It returns NULL.
It uses function: fars_read_years
}
\examples{
\dontrun{
fars_summarize_years(list(2013, 2014, 2015))
fars_summarize_years(year=list(2013, 2014, 2015))
fars_summarize_years(list("2013", "2014", "2015"))
fars_summarize_years(year=list("2013", "2014", "2015"))
}

}

