% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count the number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An character vector, integers or a list containing years}
}
\value{
a spread data frame and returns \code{NULL} if the years can not be found in the files 
in the directory or if the directory is not properly set
}
\description{
This function load data from data from
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
binds the yearly data together, group by year and month, counts th number of accidents and 
spread the results with month in rows and a year per column.
}
\examples{
\dontrun{
head(fars_summarize_years(2013:2015))
}
}

