% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS Summary}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The year(s) of FARS data to examine.}
}
\value{
This function returns a file name with the year, as defined by the
user, appended.
}
\description{
This is a function that summarizes FARS data by the month and year. The user
defines the year(s) of FARS data to be summarized.
}
\examples{

\dontrun{
fars_summarize_years(years = c(2013, 2014, 2015))
fars_summarize_years(years = 2015)
}

}
