% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of intergers, representing the years to summarise}
}
\value{
This function returns a dataset containing the year, month and number of observations.
}
\description{
This takes in a vector of \code{years} and reads the data for them. Then this function
creates a summary of the number of observations in each year and month.
}
\examples{

\dontrun{
fars_summarize_years(c(1991,1992,1993))
fars_summarize_years(2001)
fars_summarize_years(c(1991,2002,2013))
fars_summarize_years("ababab")
}

}
