% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of number of observation per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector}
}
\value{
Returns a tibble summary table
}
\description{
Summary of number of observation per month and year
}
\details{
This function counts the number of observation per month and year
}
\examples{
\dontrun{fars_summarize_years(2013:2015)}
}
\author{
James Bertschy \email{james.bertschy@gmail.ch}
}
