% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function reads a list of years and summarizes them.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer}
}
\value{
This function returs a list of data.frames with the data summarized
}
\description{
This function reads a list of years and summarizes them.
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014))}
}
