% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Counts the number of accidents per month in a given range of years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The range of years to analyze.}
}
\value{
A data frame with the number of accidents per month in a given range of years.
}
\description{
Returns a data frame that contains months as rows and the number of accidents years as columns.
}
\examples{
fars_summarize_years(1992:1999)
fars_summarize_years(years = 1992:1996)

}
