% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{counts number of accidents records for the month of the specified year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer year, a list of integer years and a vector of integer}
}
\value{
Returns a pivot tibble (data frame) with months in rows and selected years in columns containing the number of accidents

a warning for every input year that does not exist in the datasets

an error (no results) if a non numeric or a non integer input is passed

a data frame with the first column of the returned data frame is the month and the rest of the columns of the monthly counts for the passed year(s)
}
\description{
counts number of accidents records for the month of the specified year
}
\details{
use the function far_read_years
}
\note{
this function returns the data frame of monthly record counts for the specified year(s)

the first column of the returned data frame is the month

the second column is the record counts of the corresponding month for the passed year(s)
}
