% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize a selection of data from FARS files.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of character strings representing years,
four-digit format (i.e. 2013).}
}
\value{
This function returns a data frame.
}
\description{
\code{fars_summarize_years} reads in a vector of character strings
representing years. It returns a data frame with the number of accidents
by year and month, with a column for every year in \code{years} and
a row for every month.
}
\details{
This function will raise an error if there is no file
that matches the year in \code{years}.
}
\examples{
years <- c('2013', '2014')
accident_summary <- fars_summarize_years(years)

}
