% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Return a summary of accidents by month in the specified year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer or a string corresponding to a specific year}
}
\value{
The fars_summarize_years function returns the summary of accidents as a tbl_df object
}
\description{
The fars_summarize_years function takes as input a year in the form of an integer or string.
It then summarizes accident data from the specified year.
}
\details{
The fars_summarize_years function returns an error message if years is invalid.
}
\examples{
\dontrun{
if (requireNamespace("dplyr", quietly = TRUE)){
  if (requireNamespace("tidyr", quietly = TRUE)){
      fars_summarize_years(2015)
    }
  }
}

}
