% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Display a count of the number of fatalities per month and year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years for the files you're interested in.}
}
\value{
This function returns a tibble.
}
\description{
This function counts the number of traffic fatalities per month/year for the
vector of years specified in argument \code{years}.
}
\details{
This function makes use of the \code{fars_read_years()} function, which in
turn makes use of \code{fars_read()} which uses \code{make_filename()}. First,
the list of tibbles returned by \code{fars_read_years()} is unioned and then
grouped by year and month. Then \code{dplyr::summarize()} is used to count
the number of fatalities in each year/month group. Finally, \code{tidyr::spread}
pivots the counts so that they're visually appealing for printing.
}
\examples{
fars_summarize_years(c(2013, 2014, 2015))

}
