% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{integer a vector}
}
\value{
a data frame
}
\description{
Summarizes accidents over years and months.
}
\examples{
if(interactive()){
   fsum <- fars_summarize_years(2013:2015)
}

}
