% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize datasets}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A integer representing a specific year}
}
\value{
A tibble that contains the count of measures per month and year
}
\description{
This functions summarize the dataset(s) by years and month.
}
\examples{
\dontrun{
x <- c(2014, 2015)
fars_summarize_years(x)

fars_summarize_years(2013)
}
}
