% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of fatal accidents per month in FARS data}
\usage{
fars_summarize_years(years = c(2013, 2014))
}
\arguments{
\item{years}{(integer vector)}
}
\value{
Data frame with columns month, and the number of fatal accidents for the specified years.
}
\description{
From specified years, return the number of fatal accidents per month
as included in the (FARS) data.
Only data for the years 2013, 2014 or 2015 is available.
}
\details{
Throws an error if data for specified year does not exist.
}
\examples{
\dontrun{fars_read_years(c("2013","2014"))}
}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
