% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A data frame with one column for all months and one column for each
year with the number of accidents per month and year. If type of \code{years}
is not correct or if its elements are invalid, an error message is generated
and nothing is returned.
}
\description{
The function \code{fars_summarize_years} reads data from the US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System
and counts accidents per month and year.
}
\examples{
setwd(system.file("extdata", package = "fars"))
fars_summarize_years(2013:2015)

}
