% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data counts by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of vectors with characters, integers or numerics representing the year}
}
\value{
A data.frame summarizing the observations in the fars data with each month as a row and each year as a column
}
\description{
Inputs a list or vector of years and outputs a data frame summarizing the counts of data for each year and month
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014,2015))

}
