% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This is a function that generates a monthly sum for each year for the data. It uses \code{\link{fars_read_years}} function internally.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years in integer format to read. It might stop if the argument is not in integer format.}
}
\value{
A data.frame(tibble) with monthly total for each year.
}
\description{
This is a function that generates a monthly sum for each year for the data. It uses \code{\link{fars_read_years}} function internally.
}
\examples{
\dontrun{
years = seq(2013, 2015)
fars_summarize_years(years)
fars_summarize_years(years=years)
}

}
