% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function \code{fars_summarize_years} provides number of accidents for each
month and year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Four-digit year, as integer or as a string (i.e. 2001, "2001").}
}
\value{
Data frame with columns:
\describe{
   \item{MONTH}{a value between 1 and 12, 1 for January, 12 for December}
   \item{year}{four-digit value for a year, column contains numbers of
               accidents for each month for the given year}
}

If the requested year input is not present in available data, the function stops
and return an error:
"file 'filename_YEAR.csv.bz2' does not exist".
}
\description{
Function \code{fars_summarize_years} provides number of accidents for each
month and year.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years("2013")
}
}
