% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count months and years}
\usage{
fars_summarize_years(years)
}
\value{
This function returns a tibble with the months in the first column
        and the counts for the different years in the remaining columns.
}
\description{
This function reads csv files for given \code{years}, extracts the months, and counts them for each year.
}
\note{
This functions throws a warning if a file for the given year does not exist.
}
\examples{
\dontrun{
setwd(system.file("extdata", package = "fars"))
fars_summarize_years(2013:2016)
}

}
