% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_pack.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{-   Years as a 4 digit number.
The year  can be provided as single year, as a vector or list of years}
}
\value{
Dataframe containing:  Month, Year1,[Year2]..[YearN]
}
\description{
fars_summarize_years takes the table(s) of years and months, created by the fars_read_years function
and  binds these dataframes together, summarizing the data to show the
frequency of accidents for each month, with seperate columns for each year specified.
}
\note{
If the underlying CSV file is not stored in the current working directory the process
will call an warning :
      In value[[3L]](cond) : invalid year: 2013
Requires the use of the pipe function from the magrittr library
Code needs to be updated to unclude a directory option as an input parameter
}
\examples{
   fars_summarize_years(2013)
   fars_summarize_years(c(2013,2014,2015))
   fars_summarize_years(c("2013","2014","2015"))


}
\references{
The US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census, provides the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
You can download the FARS data at \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\seealso{
\link{fars_read_years}
 \link{fars_map_state}
 \link{fars_read_years}
}
