% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{It starts with having one row per observation
and columns for Month and years, what this
function does is to count how many observations
happen by Month and Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{input numeric vector}
}
\value{
tibble
}
\description{
It starts with having one row per observation
and columns for Month and years, what this
function does is to count how many observations
happen by Month and Year
}
\examples{
fars_summarize_years(c(2014,2015))
}
