% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with a list of years to summarize by.}
}
\value{
A data.frame with number of accidents by years summarized by month,
or an error in the case that \code{fars_read_years} function return \code{NULL}.
}
\description{
This function summarizes yearly accidents data, by month
}
\examples{
\dontrun{
plot(fars_summarize_years(2014))
fars_summarize_years(c(2014, 2013))
}

}
\seealso{
\link{fars_read_years}
}
