% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with the years value}
}
\description{
This function summarize the years, grouping the rows by year and month
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014)
fars_summarize_years(c(2013,2014,2015))
}

}
