% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes yearly fars data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector. The years of the traffic data.}
}
\value{
A tibble withe fars data, each year represented as a column.
}
\description{
\code{fars_summarize_years} loads-in fars data file for the specified years
and joins them to one tibble.
}
\examples{
\dontrun{
fars_summarize_years(2014)
fars_summarize_years(c(2013, 2014, 2015))
}
}
