% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fras_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector with years of interest}
}
\value{
a tibble with the amount of accidents per month in each year
}
\description{
Creates a  tibble with the amount of accidents per month in each year
}
\examples{
fars_summarize_years(c(2013, 2014))

}
