% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count accidents for month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of four digit years either as numbers or as character strings.
The years must correspond to one file previously downloaded in the current
working directory, otherwise the function returns NULL}
}
\value{
data.frame with the total number of accidents for every month and
 year given in \code{years}
}
\description{
Comptutes the total number of accidens from FARS data for given years
}
\examples{
fars_summarize_years(c("2013", "2014"))
fars_summarise_years(2015)
fars_summarise_years(2012)

}

