% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Fatality counts}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer, vector, or list of year(s)}
}
\value{
Returns a tibble of counts by month for each of the seleted years
}
\description{
This function will count the nunber of motor vehicle fatal injuries for each month of each year
}
\details{
The yearly data is first aggregated into one list by using \link{fars_read_years}. Next, this
function row binds the list elements together into a tibble. Finally, dply and tidyr functions are used
to summarize and spread the data in order to generate the final output.
}
\examples{
fars_summarize_years(2013:2015)

}
