% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data for given years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric or character vector that can be convereged to numeric vector,
 meaning years to be filter.}
}
\value{
This function returns a table with counts.
 Returned object has following classes: tbl_df, tbl, data.frame.
}
\description{
This is a simple function that creates summary table with counts of number of records
 per each month in a specified years.
}
\examples{
fars_summarize_years(c(2014,2015))
fars_summarize_years(c("2014"))
}

