% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{year as a number, or a list of numbers,
in form of integer of character}

\item{MONTH}{selected month}

\item{n}{observed numbers}
}
\value{
This function returns a list in which each item is correspond to
a year, each item is a dataframe containing two columns: Month and year
}
\description{
This function gets a vector of years/year and returns a summary of
quantity of accidents, columns being the years and rows being months
}
\examples{
\dontrun{fars_summarize_years("2013")}
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years(c(2013,2014,2015))}

}
