% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Obtener un par clave-valor}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector entero que indica los a침os}
}
\value{
Esta funci칩n regresa un par clave-valor, en este caso el a침o y las agrupaciones
}
\description{
Esta funcion permite obtener una par clave-valor entre agrupaciones
}
\details{
Se debe ingresar correctamente el argumento, sinoo se obtendr치 una mensaje de error
}
\examples{
\dontrun{
fars_summarize_years(2013:2016)
}

}
