% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize monthly data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of integers representing years}
}
\value{
This function returns a dataframe with the number of accident per months for
every given year. Each year is represented as a column in the dataframe.
}
\description{
This function can be used to obtain the number of accidents per month for several
years, selected via the \code{years} parameter.
}
\examples{
\dontrun{
fars_summarize_years(list(2014,2015))
}

}
