% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of observations by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{numerical input to indicate the years of the required data}
}
\value{
Output the data frame with monthly number of accidents (row) and selected years
(column).
}
\description{
Read the multiple data files with different years from Fatality Analysis Reporting
System and summarise the number of observations by month and year.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}
\seealso{
\code{\link{fars_read_years}}
}
