% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Print "fars_summarize_years"}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector that contains POSIXlt date time format}
}
\value{
This function returns a data frame with the number of traffic crashes
occurred on each input year and month.
}
\description{
This function provides a summary of number of traffic crashes
occurred by months and years.
}
\examples{
\dontrun{
fars_summarize_years(c(2013:2015))
}

}
