% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data.}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System \url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years in which a separate data frame will be created for each year.}
}
\value{
The function will return a dataframe summarizing the number of observations for a month and year.
}
\description{
This function will create a dataframe with n columns and 12 rows.  The columns represent the years input
as a parameter and the rows represent the 12 months of the year.  The values within the data frame
represent the total number of observations found for each month/year combination.
}
\note{
The function directly calls the fars_read_years() function, and indirectly calls the make_filename() and fars_read() functions
so the same notes listed for those functions apply here.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))
}
\references{
This function directly calls the fars_read_years().
}
