% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS Data for several years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of numeric year values for which to load data.}
}
\value{
This function returns a summary table with rows for each \code{MONTH}
   and columns for each \code{year} providing count of fatal accidents for each
   \code{(MONTH, year)} pair.
}
\description{
This function provides a summary table of fatal accident count by month from
FARS data for the requested years.
}
\details{
FARS is the US National Traffic Safety Commission's
\href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System}.
}
\note{
This function checks that the files for the requested years exist in the working directory,
and will generate warnings for each of those files that does not exist.
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014))}
\dontrun{fars_summarize_years(list(2013, 2014))}

}
\seealso{
\code{\link{fars_read_years}}
}

