% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years (numeric or integer) that will be
searched in the data}
}
\value{
Returns a pivot tibble (data frame) with months in rows and selected
years in columns containing the number of accidents. Returns a warning for
every input year that does not exist in the datasets. Returns an error (no
results) if a different than numeric or integer input is presented.
}
\description{
Count the number of observations by month in the given years
}
\examples{
\dontrun{fars_summarize_years(2013:2015)}

}
