% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{generates a tibble that tallies the number of accidents
per year, segmented by month.}
\usage{
fars_summarize_years(years)
}
\value{
Returns a tibble dataframe with the months sorted
vertically by row and the years organized horizontally as
column headers. This function is also designed to return a
warning if the year is not available as source data (from
fars_read_years()) and an error if the input argument is not
numeric.
}
\description{
This function accepts a list of qualifying years and
will create a new tibble dataframe that
separates the twelve months into rows, and the years
into columns. The function then tallies the number of
accidents in each month.
}
\examples{
\dontrun{
fars_summarize_years(2014)
}


}
