% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a table of number of observations for each month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a single year or a vector containing multiple years}
}
\value{
This function returns a table of number of observations for each month and year
}
\description{
This function uses the list created by \code{fars_read_years}, calculates the number of observations for each month and year pair and generates a tabular output of this data.
}
\examples{
fars_summarize_years(c(2010:2013))

}
