% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Fatality Analysis Reporting System data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{(\emph{character} or (\emph{integer}) ) A vector containing the desired years. 
Both \emph{numeric} and \emph{character} vectors are allowed for \code{years}.
\code{years} is coerced to integer. If \code{years} cannot be coerced to integer,
\code{fars_read_years} will result in an error as filenames cannot be created
properly.}
}
\value{
Summarized data is returned as tibble (see \code{\link[dplyr]{tbl_df}})
where for each month (rows) the number of incidents are given for each year (columns).
}
\description{
Summarizes the Fatality Analysis Reporting System (FARS) incidents per month
and year for given \code{years}.
}
\details{
FARS data is read using \code{\link{fars_read_years}}. FARS incidents
are grouped by year and month and then counted. \cr
Files must exist in current working directory.
}
\examples{
\dontrun{
fars_summarize_years(c(2014, 2015))
smry <- fars_summarize_years(list(2014, "2015"))
}

}
\references{
US National Highway Traffic Safety Administration \cr
(\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\seealso{
\code{\link{fars_read_years}}
}
