% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Make summary of accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{One or several years in a vector of character string
or numeric values.}
}
\value{
Lists of accidents by month and year.
}
\description{
This function summarizes number of US traffic accidents
by year and month.
}
\note{
Years from 2013 to 2015 are available, function throws an error if
year(s) outside 2013-2015 is selected.
}
\examples{
\dontrun{
fars_summarize_years(2014)
fars_summarize_years(c(2013, 2014, 2015))
}
}
\references{
US National Highway Traffic Safety Administration
}
\seealso{
\code{\link{fars_map_state}} for map plots.
}
