% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Get the number of observations per month from a FARS file}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{(numeric) A vector of years}
}
\value{
A tbl_df with column MONTH and columns corresponding to years with
the number of observations per month per year.
}
\description{
A FARS file with standard naming is expected to be found in the working
directory.
}

