% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS summarize years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector containing the years of interests as \code{integer} or
\code{character}.}
}
\value{
A data frame table (see \code{\link[dplyr]{tbl_df}} in the \code{dplyr} package).
}
\description{
Summarizes the number of fatal injuries suffered in motor vehicle traffic
crashes by month for each provided year.
}
\details{
This function performs the steps below:
\enumerate{
  \item Read the FARS data files in R as a data frame tables for
  provided years (using \code{\link{fars_read_years}}).
  \item Bind all of the data frame tables by row into one data frame table (see \code{\link[dplyr]{bind_cols}}).
  \item Summarize the number of incidents by \code{year} and \code{MONTH}.
}
}
\section{Depends on}{

\enumerate{
  \item \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{group_by}} and \code{\link[dplyr]{summarize}} in the
  \code{dplyr} package.
  \item \code{\link[tidyr]{spread}} in the \code{tidyr} package.
}
}

\examples{
\dontrun{
fars_summarize_years(c(2013, 2014))
fars_summarize_years(c("2013", "2014"))
}

}
