% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector containing the years to search}
}
\value{
This function returns a tibble containing the summarizes accident count per year and month.
The function will return an error if the year value is not numeric or integer
}
\description{
This function calculates the number of accidents per year on a monthly basis. The years are passed as a vector or a list
}
\examples{
\dontrun{fars_summarize_years(2013:2016)}

}
