% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise data for multiple years of Fatality Analysis Reporting System data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of years. Uses a YYYY specification.}
}
\value{
A tibble showing the number of fatalities by month and year.
}
\description{
This function returns a summary of the number of fatalities in motor vehicle
crashes for multiple years. The summary is done by year and month.
}
\note{
The csv files with the data need to be downloaded previously from the
  US National Highway Traffic Safety Administration's Fatality Analysis
  Reporting System website
  (\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\examples{
\dontrun{fars_summarize_years(2010)}
\dontrun{fars_summarize_years(2012:2015)}
}
