% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS data Summary Years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector containing a list of years to be sorted and summarized}
}
\value{
a data frame summarizing the number of fatal accidents each month sorted by year
}
\description{
summarizes FARS data for multiple years.
}
\examples{
fars_summarize_years(c(2013,2014,2015))
plot(fars_summarize_years(c(2014,2015)))

}
\seealso{
\code{\link{fars_read_years}}
}
