% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Show a summary of accidents during given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer}
}
\value{
a tibble showing summary of number of accidents by month
each year
}
\description{
Show a summary of accidents during given years
}
\note{
"dplyr" package is required
}
