% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count records per month for each year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years for which the data files in the working
directory need to be read and processed. If a corresponding CSV file
for a year in the vector of numeric years is not present in the working
directory, an error message indicating "invalid year" is generated and the
data frame for that year will be set to \code{NULL}}
}
\value{
A data frame with a \code{MONTH} column and a column of record counts
for each year in the \code{years} parameter.
}
\description{
The \code{fars_summarize_years()} function summarizes the data obtained from
multiple years by counting the number of records for each month in each year.
The function takes a vector of years as input and returns a data frame with
summarized information.
}
\section{External Functions Not Imported}{

\itemize{
\item \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} used to combine all data frames into one data frame.
\item \code{\link[dplyr:group_by]{dplyr::group_by()}} used to group rows of data frame by \code{year} and \code{MONTH}.
\item \code{\link[dplyr:summarise]{dplyr::summarize()}} used to calculate a summary statistic for each group.
\item \code{\link[tidyr:spread]{tidyr::spread()}} used to convert data frame from long to wide format.
}
}

\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years(2013:2015)
}

}
\seealso{
\code{\link[=fars_read_years]{fars_read_years()}} to create a list of data frames
}
