% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Print "fars_summarize_years"}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A int vector with years contained in the file name}
}
\value{
This function returns a wide table that contain number observation by month and year
}
\description{
This is a simple function summarize observations number by month and year
}
\details{
If years does not contain the year in the file name the function fails
}
\examples{
\dontrun{
years <- 2013:2015
fars_summarize_years(years)
}
}
