% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generate a summary of fatalities per month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector, or list, of years.  The elements of the list/vector
must be coerceible to integers e.g. character string "2012".}
}
\value{
This function returns a summary, \code{data.frame} (\code{tbl_df}) listing
   the number of fatalities per month for each given year.
}
\description{
This function generates a summary of fatalities per month, for the
given vector/list of years.  \code{fars_summarize_years} takes a
single argument; \code{years}; a vector/list of years that the summary
will cover.
}
\examples{
fars_summarize_years(c(2012, 2013, 2014))
fars_summarize_years(c("2012", "2013", "2014"))

}
