% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{fars_read_years}
}
\value{
A \code{tibble} of number of accidents with columns represent the years and rows of the months.
}
\description{
This function returns a summary of the number of accidents per month for each year of the \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data
}
\details{
The function returns a numerical summary of the number of accidents per month for each year provided in the input.
}
\examples{
\dontrun{
years <- 2013:2015
fars_summarize_years(years)
}

}
