% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function summarizes the number of incidents by month and year in FARS data and returns a dataframe.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numerical vector representing years to read in the dataset(s)}
}
\value{
This function returns a dataframe summarizing the number of FARS incidents by month and year.
The datframe includes a column for MONTH and a column for each year summarized.
}
\description{
This function summarizes the number of incidents by month and year in FARS data and returns a dataframe.
}
\note{
the function will throw an error if the \code{year} supplied is not found in the dataset(s) or is not a number
}
\examples{
\dontrun{fars_summarize_years(2013:2015)}
\dontrun{fars_summarize_years(not_a_number)}

}
