% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes number of accidents by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer/numeric vector of years, with each year in the format yyyy}
}
\value{
A data.frame of columns <month>, <year1>, <year2>, <year3>, ... with each
year column having the number of accidents for each month,year combination
}
\description{
This function takes in an argument of a vector of years with each
year in the format yyyy and generates a summary of accidents by month
and each year passed in the argument years.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
fars_summarize_years(c(2013, 2015))
}

}
