% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years in numeric or integer format.}
}
\value{
This function returns a data.frame with months as rows
   and chosen year as column with counts of accident events per month and respective year(s)
}
\description{
This simple function summarises the accident events by month and year in a chosen year(s)
}
\examples{
\dontrun{
fars_summarize_years(2015)
}
}
