% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Make a list of cumulative number of occurances (accidents) in a MONTH of the year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{Input}{class 'integer' value(s) of year(s) as the one argument.}
}
\value{
This function returns a data frame that contains total number of occurances
(accidents) in a MONTH for the input year.
}
\description{
This function calculates the total number of accidents happened in a MONTH within a
year. It yields a data frame and prints a Tibble format of the output.
}
\examples{
\dontrun{
fars_summarize_years(2015)
fars_summarize_years(c(2013,2015))
}

}
