% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Add together two numbers}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integer values.}
}
\value{
A data.frame giving the counts of the present elements for each month by year.
}
\description{
Add together two numbers
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014))}
}

