% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years(): Summarize monthly accidents for given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An year of interest in which data to be read (2013-2015)}
}
\value{
A data table summarizing the number of accidents per month
}
\description{
This function accepts a specific year and reads the accident data of that year.
Afterwards, summary of accidents per month of given year will be displayed
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014))
}
}
