% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize how many accidents happend in each month by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector with all years that should be loaded}
}
\value{
This function returns a dataframe summarizing the numer of accidents in each month of each year.
   if the year required is invalid a warning will be generated to inform it.
}
\description{
This function summarizes the number of occurrencies in each month by year.
}
\details{
This function uses bind_rows, group_by and summarize functions from dplyr package
   and spread function from tidyr package.
   It also calls fars_read_years function in order to select only month column from each year.
   In case the year required doesn't have a correspondent file it will generate a warning.
}
\examples{
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years(c(2014))
fars_summarize_years(c(2013, 2015))

}
