% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Load data from accident files and summarize by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of strings or integers giving the years to retrieve}
}
\value{
a data_tbl with a row for each month of a year and a column with the number of accidents
}
\description{
This is a function that loads data for a list of years and summarizes the data
by year and month and returns the data in a tidy data format.
}
\details{
If a accident file with standard naming convention is not in currect working, a "invalid year" warning is generated.
}
\examples{
\dontrun{
fars_summarize_years(c(1999,2000,2001)) 
fars_summarize_years(c("1999","2000","2001")) 
}

}
