% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read in and summarise FARS data for a specified range of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A year or a vector of years (in any class that can be coerced to numeric)
to specify which FARS reports are to be read in}
}
\value{
This function returns a \code{tbl_df} object with one column for every year analysed and
       twelve rows, one for each month. The columns contain the sum of accidents reported to
       FARS in each month of the given year.
}
\description{
The \code{fars_summarize_years} function is used to read in FARS data from a specified range of years
and return the monthly number of accident registered in FARS.
}
\details{
The function internally uses the \code{\link{fars_read_years}} function to read in and
        extract the data and then uses \code{dplyr} and \code{tidyr} functions to summarise and
        reformat the data.\n
        Like \code{\link{fars_read_years}} it reads the data from your working directory
        and it will produce a warning if an invalid year is given as input. Please be aware that
        the standard naming convention for FARS data has to be used (cf. \code{\link{make_filename}}).
}
\examples{
\dontrun{fars_summarize_years(2013:2015) #Summarize FARS data for the years 2013, 2014 and 2015}

}
