% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise number of records per month contained in fars file(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector of numerics, integers or numeric/integer strings to read "accident_<year>.csv.bz2"
files from the working directory.}
}
\value{
A tibble contain number of records month by year
}
\description{
Takes a list or vector of years and returns a tibble with number of records month by year for
corresponding fars csv files in the working directory.

If file does not exist for corresponding year a warning will be generated.
}
\examples{
\dontrun{
fars_summarise_years(2018)
fars_summarise_years(2018L)
fars_summarise_years('2018')
fars_summarise_years(c(2017, 2018))
fars_summarise_years(list(2017, 2018))
}

}
