% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize statistical results of FARS}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector containing multiple character strings}
}
\value{
A dataframe of fatal injuries in given years
}
\description{
This function gives a summary of fata injuries given years
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
}

}
