% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_W4.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarising data of the years requested}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years of the data requested}
}
\value{
A tibble summarising the number of accidents by month and year
}
\description{
This function accepts a vector of years of the data requested and
summaries the number of accidents by month and year.
}
\note{
An error will occur if the file does not exist in the working directory.
}
\examples{
d <- fars_summarize_years(c(2013,2014,2015))

}
