% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Fatalities by month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years for which fatalities count is calculated}
}
\value{
The count of fatalities by year and month
}
\description{
Summary of fatalities by year and month
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014))
}

}
