% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize dates within FARS datsets}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integers, or objects that can be coerced into
integers, representing specific years of FARS data.}
}
\value{
a datframe with the first column MONTH with integers 1 through 12,
  and a columm for each specified year. A worning is diplayed if a year does
  not return a valid source file.
}
\description{
Given a vector of years, this function returns a table for numbers of
observations within each month/year in FARS datasets.
}
\examples{
my_file <- system.file("extdata", "fars_data.zip", package = "fars")
unzip(my_file, junkpaths = TRUE)
fars_summarize_years(2013:2015)

}
\seealso{
\code{\link{fars_read_years}}
}
