% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Return number of accidents aggregated by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of numeric data convertible to four-digits integers
defining the years of interest. The years should belong to the interval [2013, 2015],
otherwise, an exception is trown.}
}
\value{
Returns a tibble containing aggregated number of accidents.
 The tibble has the following columns:
\describe{
 \item{MONTH}{The month of a year, \code{[1:12]} }
 \item{<year_1>}{The first year from the \code{years} parameter}
 \item{...}{}
 \item{<year_N>}{The last year from the \code{years} parameter}}
}
\description{
The function processes FARS data to aggregate number of accidents by month and year.
For more information about FARS data, see
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}.
}
\examples{
fars_summarize_years(c(2013, 2014))
summary2015 <- fars_summarize_years(2015)

}
