% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of acccidents per year and per month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric or character vector contining the year(s) for which
the data should be summarized.}
}
\value{
This function returns a data frame with number of accidents
   summarized per year and month.
}
\description{
This is a function that summarises the number of accidents per year and  month.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years("2013")
fars_summarize_years(c(2013, 2014))
fars_summarize_years(c("2013", "2014"))
}

}
