% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data per month for a series a provided years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers}
}
\value{
This function returns a dataframe with the number of accidents by month and year.
}
\description{
This function generates an overview of the number of accidents per month in the provided year(s).
}
\examples{
\dontrun{
fars_summarize_years(2017)
fars_summarize_years(c(2013, 2014, 2015))
}

}
