% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of the number of observations per month and year in the imported file/s}
\usage{
fars_summarize_years(years, demo = FALSE)
}
\arguments{
\item{years}{a character or numeric value, a vector or a list of years}

\item{demo}{logical if you want to use the function with demo data that comes with the package.
Default to FALSE}
}
\value{
a tibble showing the number of observations by month for the imported files.
 Each file read is one column of the tibble.
}
\description{
This function counts the number of observations by months for the specified file/s.
Files should be in working directory.
}
\examples{
\dontrun{acc_summ <- fars_summarize_years(2013:2015)
acc_2015 <- fars_summarize_years("2015")}

#'@include make_filename.r fars_read.r fars_read_years.r
}
