% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize year file}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{is an integer or list or vector representing year(s)}
}
\value{
data frame containing a row for each month and columns with counts for each year
}
\description{
Summarize year file
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)
fars_summarize_years(c(2013,2015))
}
}
