% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Return monthly total accidents for given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
a tibble summarizes total accidents per month for given years
}
\description{
This function takes a vectors of year and feeds it into the \code{\link{fars_read_years}}
function. It then uses dplyr and tidyr packages to group total number of accident
by months and spread it by the years.
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014, 2015))}
}
\seealso{
\code{\link{fars_read_years}}
}
