% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary Statistics}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer or a vector of integers giving the year(s) the user is interested in.}
}
\value{
This function returns a tibble containing the number of observations per month per year.
A warning tells you if the requested year is invalid.
}
\description{
Number of observations per month per year.
}
