% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{\code{fars_summarize_years} combines all the dataframes obtained from the \code{fars_read_years} function into one single dataframe,
groups the data into each specific 'year' and 'month', summarises by counting the number of data points for each group, and then transforms the
table from the long to wide format according to the 'year' and number of counts for each group variables.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years to read in}
}
\value{
a dataframe of summarised data of wide format
}
\description{
\code{fars_summarize_years} combines all the dataframes obtained from the \code{fars_read_years} function into one single dataframe,
groups the data into each specific 'year' and 'month', summarises by counting the number of data points for each group, and then transforms the
table from the long to wide format according to the 'year' and number of counts for each group variables.
}
\examples{
\dontrun{
fars_summarize_years(c(2011, 2012))
}
}
