% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{NHTSA FARS summary: number of fatal injuries per month for the specified years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of the years of observations to read.}
}
\value{
A tibble.
}
\description{
Reads Fatality Analysis Reporting System (FARS) data files and builds summary table containing
number of fatal injuries suffered in motor vehicle traffic crashes for each month of the specified years.
}
\details{
Returns an error if none of the elements in \code{years} is valid.
  If only some of the elements are invalid returns the tibble and throws
  warnings for every invalid year.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}
}
