% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create summary of number of obserwations}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector specifying years of interest}
}
\value{
This function returns a table (tibble) with the number of observations
in each month and each given year
}
\description{
This function creates a summary of number of accidents in each monts and each
year specified in \code{years} argument.
\code{fars_summarize_years} is dependent on \code{dplyr} and \code{tidyr} packages.
}
\examples{
\dontrun{fars_summarize_years(c(2013,2015))}

}
