% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Sum of accidents by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of numeric values which are years}
}
\value{
A data frame with counts of accidents for each month each year
}
\description{
Take in a list of years and summarize how many accidents happened in each month
each year, returns a data frame with years as columns, rows as months, each cell
is the number of accident happened in that month (row) of that year (column).
}
