% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize entry number of each month by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of list of years to summarize}
}
\value{
a wide type data frame containing the number of accidents for each month
}
\description{
generate summary
}
\examples{
fars_summarize_years(c(2013,2014))

}
\seealso{
\link{fars_read_years}

\link{fars_read}

\link{make_filename}
}
