% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count number of accidents happend in each month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of four-digit numbers representing years.}
}
\value{
A tibble of monthly accident counts.
}
\description{
This function imports and combines data from multiple files and counts the
number of accidents happend in each month and year. The first column of the
returned tibble represents the month index, whereas other columns represent
the monthly accident counts in certain years.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013, 2014, 2015))
}

}
