% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_fars_fun.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years that were read into R for summarization.}
}
\value{
fars_summarize_years returns a tibble of number of accidents for each year with month as column
}
\description{
This function takes an argument that is a year, reads the data in if it exists,
keeping month and year, and giving the number of observations grouped by month and year.
}
