% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution.R
\name{get_distribution}
\alias{get_distribution}
\title{Generic function to extract distribution from fars_density object}
\usage{
get_distribution(object, ...)
}
\arguments{
\item{object}{An object from which to extract the distribution.}

\item{...}{Additional arguments.}
}
\value{
A matrix of random draws from the fitted skew-t distribution.
}
\description{
Generic function to extract distribution from fars_density object
}
