% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_methods.R
\name{get_ellipsoids}
\alias{get_ellipsoids}
\title{Generic Function to Extract Ellipsoids}
\usage{
get_ellipsoids(x, ...)
}
\arguments{
\item{x}{An object from which to extract the ellipsoids}

\item{...}{Additional arguments.}
}
\value{
A list of matrices defining the ellipsoids at each time.
}
\description{
Generic Function to Extract Ellipsoids
}
