% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factors.R
\name{get_factors}
\alias{get_factors}
\title{Generic Function to Extract Estimated Factors}
\usage{
get_factors(x, ...)
}
\arguments{
\item{x}{An object from which to extract the estimated factors.}

\item{...}{Additional arguments.}
}
\value{
A matrix of estimated factors.
}
\description{
Generic Function to Extract Estimated Factors
}
