% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loadings.R
\name{get_loadings}
\alias{get_loadings}
\title{Generic function to extract factor loadings}
\usage{
get_loadings(object, ...)
}
\arguments{
\item{object}{An object from which to extract the factor loadings.}

\item{...}{Additional arguments.}
}
\value{
A matrix of factor loadings.
}
\description{
Generic function to extract factor loadings
}
