% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mldfm_list.R
\name{get_mldfm_list}
\alias{get_mldfm_list}
\title{Generic Function to Extract List of MLDFMs}
\usage{
get_mldfm_list(x, ...)
}
\arguments{
\item{x}{An object from which to extract the MLDFMs.}

\item{...}{Additional arguments.}
}
\value{
A list of \code{mldfm} objects
}
\description{
Generic Function to Extract List of MLDFMs
}
