% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Generic function to extract quantiles}
\usage{
get_quantiles(object, ...)
}
\arguments{
\item{object}{An object from which to extract the quantiles.}

\item{...}{Additional arguments.}
}
\value{
A matrix of quantiles.
}
\description{
Generic function to extract quantiles
}
