% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Generic Function to Extract Residuals}
\usage{
get_residuals(x, ...)
}
\arguments{
\item{x}{An object from which to extract the residuals.}

\item{...}{Additional arguments (ignored).}
}
\value{
A matrix of residuals.
}
\description{
Generic Function to Extract Residuals
}
