% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stressed_factors.R
\name{get_stressed_factors}
\alias{get_stressed_factors}
\title{Generic function to extract stressed factors}
\usage{
get_stressed_factors(object, ...)
}
\arguments{
\item{object}{An object from which to extract the stressed factors.}

\item{...}{Additional arguments.}
}
\value{
A matrix of stressed factors.
}
\description{
Generic function to extract stressed factors
}
