% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_density.R
\name{l_density}
\alias{l_density}
\title{Compute Skew-t Densities from Forecasted Quantiles  (Linear Optimization)}
\usage{
l_density(
  quantiles,
  levels = c(0.05, 0.25, 0.5, 0.75, 0.95),
  est_points = 512,
  random_samples = 5000,
  support = c(-10, 10),
  seed = NULL
)
}
\description{
Compute Skew-t Densities from Forecasted Quantiles  (Linear Optimization)
}
\keyword{internal}
