% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or number for year}
}
\value{
This function returns name o file in bz2 format
}
\description{
This function creates filename based on year
}
\details{
Function inserts year given and creates "accident_year.csv.bz2". year will be replaced from input supplied
}
\examples{
make_filename("2021")
make_filename(2021)

}
