% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Make FARS file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{This is the year in YYYY format of the FARS
data.}
}
\value{
This function returns a string representing the
    FARS data filename for a given yeare
}
\description{
This function generates a file name for FARs data.
}
\examples{
make_filename("2013")
make_filename("2014")
make_filename("2015")

}
\author{
John J Smith, \email{fakeemail123@duvman.ie}
}
\keyword{FARS}
