% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year you would like to coerce into a filename.}
}
\value{
Returns the name of a given file
}
\description{
This is a function that will  create a filename using user-inputted year.
}
\examples{
make_filename(2000)

}
