% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename: Return a character vector containing a formatted combination
of text and variables from an inputted year. The corresponding file must be
available in the working directory as this helper function does not check if
it is present.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number or a string}
}
\value{
character vector of filename formatted as 'accident_year.csv'.
}
\description{
make_filename: Return a character vector containing a formatted combination
of text and variables from an inputted year. The corresponding file must be
available in the working directory as this helper function does not check if
it is present.
}
\examples{
make_filename(2013)
\dontrun{newFile <- make_filename('2021')}
}
