% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{make_filename
make_filename create the filename based on the year from which we want to extract the information}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer that indicates the year from which we can to extract the information}
}
\value{
The filename string for that specific year
}
\description{
make_filename
make_filename create the filename based on the year from which we want to extract the information
}
\examples{
\dontrun{
make_filename<-make_filename(2015)
}

}
