% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Define filename with specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or character string that represent the year of study.}
}
\value{
This function returns a specific character string.
}
\description{
This function allows define the year to analyze.
}
\examples{
make_filename("2013")
make_filename(2013)

}

