% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a new filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer representing year of the accident data, raw data inputting is perceived as a string}
}
\value{
Returns a string in format 'accident_%d.csv.bz2' with %d represting the imported year
}
\description{
This function makes a string representing a filename based upon the year entered
}
