% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{make_filename}
\alias{make_filename}
\title{File name generator for FARS data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer of string of the year of the downloaded dataset.
The year will be used in string \code{accident_X.csv.bz2} where
\code{X} will be replaced by the given integer/}
}
\value{
returns a name of the desired dataset
}
\description{
The function helps generating a proper name for a FARS data set just by
giving the year number.
}
\examples{
make_filename("2024")
make_filename(2024)

}
