% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Print "Fars Make FileName"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string representing 4 character length year value.}
}
\value{
This function returns the filename for the year provided to it.
}
\description{
This function creates the file name for a given year by inserting the
year into the file name pattern like accident_\code{2015}.csv.bz2.
}
\note{
An invalid year will result in invalid file name.
}
\examples{
make_filename('2015')
make_filename(year = '2015')

}
