% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename by the year of interest}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value of the year. It should correspond to the available yearly records
from US National Highway Traffic Safety Administration's Fatality Analysis Reporting System.}
}
\value{
This function returns a character sting of the filename as accident_<YEAR>.csv.bz2,
where <YEAR> is substituted by the input argument.
}
\description{
This function created a filename as accident_<YEAR>.csv.bz2, where <YEAR> is substituted
by the function input \code{year}.
}
\note{
This function will give error if the input \code{year} is non-numeric or non-integer.
}
\examples{
make_filename(2013)

}

