% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a year corresponding to the data}
}
\value{
Function \code{make_filename} returns a character vector containing the value of \code{year}
in the format of \code{filename}.
}
\description{
Function \code{make_filename} takes an \code{year} and create a filename
in the format of \code{"accident_year.csv.bz2"}.
}
\note{
If \code{year} is not integer, it would be transfered to \code{\link{as.integer}}.
}
\examples{
make_filename(2016)
make_filename(2017)

}
