% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A four-digit number indicating the year in which the data collected.}
}
\value{
This function returns a character vector containing a formatted combination
 of file name template in text and year variable.
}
\description{
This function takes a four-digit number as its argument and returns a character
vector containing a formatted combination of filename text template and year variable.
}
\details{
This function takes a four-digit number as its \code{year}
 argument indicating the year in which the data has been collected.
 Then it uses \code{\link[base]{as.integer}} to coerce it into integer type.
 After that the function calls \code{\link[base]{sprintf}} to combine the
 format string for filename with the year variable. The length of the output
 will be equal to the longest input.
}
\note{
Because formats \code{d} can only be used for integer values or sometimes
 for logical values, the function use \code{\link[base]{as.integer}} to turn the
 entry for \code{year} argument into integer value.
}
\examples{
\dontrun{
make_filename(2013)
}

}
