% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{function "make_filename"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year (number with 4 digits)}
}
\value{
This function returns a filename as a character vector of length
   that of the longest input.
}
\description{
The function returns a filename for the given year
}
\examples{
yyyy <- 2015
make_filename <- function( year=yyyy )

}
\references{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
