% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename for FARS data in a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer of year}
}
\value{
This function returns a filename to be used with fars_read
}
\description{
Make filename for FARS data in a given year
}
\examples{
\dontrun file<-make_filename(2014); fars_2014<-fars_read(file)

}
