% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename based on the input year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year which you want in your filename.}
}
\value{
This function creates a filename that shows the specified year in the filename.
}
\description{
This is a function that returns a filename in the format "accident_year.csv.bz2",
where you can specify the year.
}
\examples{
\dontrun{make_filename(2013)}

}
