% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer: year of observation.}
}
\value{
Returns a character string as "accident_<year>.csv.bz2"
to be used as a filename
}
\description{
Creates a filename for a .csv.bz2 file for the \code{year}
argument with the structue "accident_<year>.csv.bz2". Requires an
integer as input and ends with an error message, otherwise.
}
\examples{
\dontrun{
makefilename(2015)
}
}
