% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{File Naming: Rename the Fars dataset}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numeric value of the year}
}
\value{
"accident_2006.csv.bz2" a formatted character vector of a dataset
}
\description{
This function returns a character vector containing a formatted
}
\examples{
\dontrun{make_filename(2006)}
}
