% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make data filename corresponding to a given yean}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or an integer with the input \code{year}}
}
\value{
A string with the .csv filename for a given
  year, and the file path within the package.
}
\description{
This function makes a data filename related to the given \code{year}
Warning : The function does not check if the file is available.
}
\examples{
\dontrun{
make_filename(2013)
}

}
\seealso{
\link{fars_read}
}
