% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{number indicating the year when the dataset was captured}
}
\value{
Returns a character vector containing the name of the FARS CSV file in the format "accident_<year>.csv.bz2".
    Returns "accident_NA.csv.bz2" and a warning if the year argument is a string.
}
\description{
Create a filename for a Fatality Analysis Reporting System (FARS) CSV file
}
\examples{
filename <- make_filename(2017)
bad_filename <- make_filename("a")
}
