% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{User-defined filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An \code{integer}}
}
\value{
Creates a filename \code{string} with year given by the user
}
\description{
Functions creates a user-defined filename by appending the year to the filename
}
\examples{
\dontrun{
make_filename(2015)
}

}
