% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Build accident files names}
\usage{
make_filename(year,path)
}
\arguments{
\item{year}{year selected for the dataset, can be numeric, integer or charactor}

\item{path}{directory containing the Fars data set to "data" by default}
}
\value{
character string corresponding to a formatted file name
}
\description{
Create a charactor string from the year matching the Fatality Analysis Reporting System naming convention
}
\note{
\itemize{
\item user should enter a valid 4 digit value either as numeric or character
\item using a boolean for year would produce "accident_0.csv.bz2" or "accident_1.csv.bz2"
\item using alphabetical character would produce an error
}
}
\examples{
filename <- make_filename(2013)
filename <- make_filename(2013L)
filename <- make_filename("2013")
filepath <- file.path("data",make_filename(2013))

}
