% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename - Make a compressed bz2 filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or an integer with the input \code{year}}
}
\value{
This function returns the data filename as a string for a given year
}
\description{
Make a (.csv) data filename for a given \code{year}
Note: This function does not check if the file is available.
}
\examples{
\dontrun{
make_filename(2013)
make_filename("2014")
make_filename(year = 2015)
}
}
\seealso{
\link{fars_read}
}
