% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create FARS data file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character or integer variable of the year}
}
\value{
This function returns a filename in the form of a character string.
}
\description{
make_filename is used to create a FARS (Fatality Analysis Reporting System) character string filename.
}
\details{
The function creates a character string with a filename of the year say 2006
in the format of accident_2006.csv.bz2
}
\note{
Error message: If the wrong file name is used then an error will be thrown
}
\examples{
\dontrun{
make_filename("2016")
make_filename(2016)
}

}

