% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function \code{make_filename}}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric representing the year}
}
\value{
A string representing a file name
}
\description{
This function will make a file name for a given year
}
\note{
This is an internal function not to be used from outside the package
}
