% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make the Filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string indicating the year that the data file generated.}
}
\value{
This function returns a character string representing the filename.
  Will return a filename with NA if the year parameter cannot be coerced to an integer.
}
\description{
This function generates the filename for a FARS datafile and takes a year as input and produces a valid FARS filename.
    It is an internal function used by other functions in this package.
}
\examples{
\dontrun{
make_filename(2014)
make_filename("2014")
}

}
