% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year that you wish to include in your filename.}
}
\value{
This function returns a string filename including the year
input converted in a numeric value.
}
\description{
This is a simple function that creates a bz2 filename using the year
input and the "accident". The function will attempt to convert the input
to a numeric, returning NA where this is not possible.
}
\examples{
make_filename("2015")
make_filename(2015)
make_filename(year = 2015)

}

