% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make fars file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value of year}
}
\value{
a file name for accident data at different years. No specific error checking was enclosed.

A character vector containing character strings with the input \code{year}
}
\description{
\code{make_filename} returns a character string with the format of "accident_\code{year}.csv.bz2" as file name
}
\examples{
make_filename(2013)

}
\seealso{
\code{\link{fars_read}}
}
