% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a 4-digt year to format filename}
}
\value{
a string with custom year filename
}
\description{
This function makes it a custom file name based on year given
}
