% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename - create a file name for the compressed file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of the accident, need to be able to converted to an integer}
}
\value{
This function returns the accident file name in compressed format (surffix .csv.bz2) as a string.
}
\description{
This function takes the year of the accident to generate a
filename for the compressed file with the year on it.
}
\examples{
my_accident <- make_filename(2014)
make_filename("2013")

}
