% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Given a \code{year} as input return the FARS filename containing data for that year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A value that can be coerced to an integer and that represents a year.}
}
\value{
A character vector that is the filename of the FARS file containing data for the \code{year} given as input.
}
\description{
Given a \code{year} as input return the FARS filename containing data for that year.
}
\note{
The input variable \code{year} must be castable to an integer or it will be coerced to NA.
}
\examples{
\dontrun{
make_filename(2015)
make_filename("2014")
}
}
