% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate filename for given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which to generate the filename as integer or type that can be meaningfullz converted with as.integer}
}
\value{
This function returns the filename in the format "accident_YEAR.csv.bz2" with YEAR replaced with the given year
}
\description{
This function determines the filename including extensions of the accident data
of the given year
}
\examples{
make_filename(1991)
}
