% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{FARS File Name Creator}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer, numeric value, or numeric string, that represents a
year for which a file name will be built. This value should be a whole
number.}
}
\value{
A character string that defines the file name of the inputted year's
 data.
}
\description{
A simple function that takes a year as a parameter and returns a
 file name that includes the year.
}
\note{
Non-whole numbers will be truncated. Inputting character strings
 will result in errors.
}
\examples{
\dontrun{make_filename(2013)}
\dontrun{make_filename("2014")}

}
