% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a specific filename for the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The integer value of a year. The \code{year} should be a positive decimal.
In order to use with the fars_data.zip dataset, the year should be in the range
2013..2015}
}
\value{
This function returns string, containing a file name of one datafile in the
fars_data.zip dataset.
}
\description{
This function creates a valid filename based on the \code{year} argument.
The filename is useful for processing the fars_data.zip dataset.
The data of a given year is in a separate file in fars_data.zip .
}
\examples{
make_filename(2014)

}
