% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creating a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of interest which is numeric value.}
}
\value{
Returns a character vector file name based on the selected year and a bz2 extendsion.
}
\description{
This function creates a data file name based on the selected year.
}
\details{
The function uses sprintf is a wrapper for the system sprintf C-library function. Attempts are made to check that the mode of the values passed match the format supplied, and R's special values (NA, Inf, -Inf and NaN) are handled correctly.
}
\note{
The year will be formatted as an integer.
}
