% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file name from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number denoting the year of the data being processed}
}
\value{
prints a file name in the format accident_%d.csv.bz where %d means the year.
}
\description{
Create file name from year

This function takes a year and prints the file name with the year at the end.
The user may supply a string instead, which needs to be error handled.
}
\examples{
make_filename(2017)
}
