% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Generate file name from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A four digit valid year}
}
\value{
This function returns a string, filename with year.
}
\description{
Generate file name from year
}
\details{
this function raise errors when year is passed as non-numeric value
}
\examples{
filename <- make_filename(2013)

}
