% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename for specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Four digits}
}
\value{
filename for specific year
}
\description{
Takes year and returns filename
}
\examples{

\dontrun{
make_filename("2015")
}
}
