% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create uniform file directory names by year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{4-digit number representing the year.}
}
\value{
Character string of the form 'accident_[year]_csv.bz2'
}
\description{
This function creates standardized file names based on year.
}
\section{Errors}{

This function will throw an error if input value cannot be converted to an integer year.
}

