% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numeric the name of the year}
}
\value{
a character which is the file name
}
\description{
Convert the year into a the name of the file which which contain 
the accident information in this year
}
\examples{
make_filename(2015)

}

