% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a name for your file}
\usage{
make_filename(year)
}
\arguments{
\item{The}{only argument is the year (can be either in character or integer form, but it has to be a number. See examples below.)}
}
\value{
this function will return a string of your new filename with specified year
}
\description{
This function gives a name to your compressed csv file. The form of the name is
accident_[wantedYear].csv.bz2.
}
\details{
this function uses 'sprintf' function from base package.
}
\examples{

make_filename(2015)
make_filename("2017")

}
