% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{print("make_filename")
This is a function that constructs the filename corresponding to the annual
dataset from the US National Highway Traffic Safety Administration's Fatality
Analysis Reporting System (FARS), provided a year. The year is passed onto
the function as an argument, which can be done using the \code{year} argument.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the 4-digit year for which the filename will be constructed.}
}
\value{
This function returns a character string containing the filename
   corresponding to the provided year data from the FARS database.
}
\description{
print("make_filename")
This is a function that constructs the filename corresponding to the annual
dataset from the US National Highway Traffic Safety Administration's Fatality
Analysis Reporting System (FARS), provided a year. The year is passed onto
the function as an argument, which can be done using the \code{year} argument.
}
\examples{
\dontrun{
make_filename(2015)
make_filename(year = 2015)
}
}
