% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename using the FARS file naming convention}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric value indicating the year}
}
\value{
a character vector representing a file name
}
\description{
Take a year and build a character vector representing the file name.
}
\examples{
make_filename(2014)
}
