% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Building a FARS filename for a given year or set of years}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year or years for which filenames should be built. May be
of class numeric or character.}
}
\value{
The function returns the FARS filename(s) built from \code{year}.
}
\description{
This function takes one argument, \code{year}, and generates a string with
the FARS filename corresponding to that year.
}
\examples{
make_filename(c(2014, 2015))
make_filename("2014")

}
