% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make data file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or an integer with the input \code{year}}
}
\value{
This function returns a string with the data file name for a given
  year
}
\description{
Make .csv data file name related to the given \code{year}
The function does not check if the file is available.
}
\examples{
make_filename(2013)
# "accident_2013.csv.bz2"
}
\seealso{
\link{fars_read}
}
