% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename from the years provided}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A vector or list indicating the years for which filenames
are to be returned. Coerced to \code{integer}}
}
\value{
Returns a \code{character} vector in a format "accident_<year>.csv.bz2" that
can be used as file names
}
\description{
The function creates a filename for a .csv.bz2 file based on the \code{year}
argument in a form "accident_<year>.csv.bz2". The input is coerced to an integer.
If coercion is not possible, it shows a warning message
}
\details{
Works for all input that can be coerced to integers. \code{NA}s are
introduced with a warning if input can't be coerced. See examples for usage.
}
\examples{
\dontrun{
make_filename(2006)
make_filename(2007:2010)

year_list <- list(2000, "2001", '2002')
make_filename(year_list)

# NAs introduced
make_filename("some_input")
}
}
