% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename() function returns make a filename by the given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{- year of accidents}
}
\value{
filename - file name as a character string
}
\description{
Make a file name as a character string by the given year
}
\examples{
\dontrun{
make_filename(2016)
}

}
