% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_original.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename for a specific year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year, numeric integer value in the form yyyy.}
}
\value{
This function returns a character string of the form "accident_%d.csv.bz2" where %d specify a year.
}
\description{
This function create a filename of the form "accident_%d.csv.bz2" where %d specify a year.
It used the \code{sprintf} base function that returns a character
string containing a formatted combination of text and variable values.
}
\examples{
file<-make_filename(year=2013)

}
