% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Title A simple formatting function to create a filename which contains a formatted combination of input values}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer from \code{\link{year}}}
}
\value{
A string which is the created filename for a particular year
}
\description{
Title A simple formatting function to create a filename which contains a formatted combination of input values
}
\examples{
filename<- make_filename(2013)
}
