% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print "make_filename"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of data in format YYYY}
}
\value{
An string with the name generated
        Error if year is not numeric
}
\description{
Generates a file name according the naming criteria of FARS
}
\examples{
make_filename(2016)

}

