% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename to read data for a specific year.}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or an object that can be coerced to an integer.}
}
\value{
This function returns a filename defining the file extension of the National
 Highway data available for extraction.
}
\description{
The function takes a single argment, which is a year defining a time period when data
from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
is available for this function.
}
\examples{
\dontrun{make_filename(2013)
make_filename(2014)}
}
