% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{produces a string represetation of a valid fars file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer, or a string value representing a year}
}
\value{
a string that appends the input year to a filename
to provide the full name of a valid fars file
}
\description{
\code{make_filename} produces a string containing a valid filename
}
\details{
This function requires a year as input and returns a full FARS filename.
}
\note{
If the year cannot be coerced to integer, NA is used.
}
\examples{
\dontrun{
make_filename(year = 2016)
}

\dontrun{
make_filename(year = 'two thousand sixteen') #  error
}

#' @export
}
