% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create the File name you want to load.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number of the year (one of 2013-2015) you want to load.}
}
\value{
This function returns a filename for fars_read function to read.
}
\description{
This is a simple function that create the file name of the year you chose.
}
\examples{
\dontrun{filename=make_filename(2013)}
\dontrun{ print(filename)}
\dontrun{df=fars_read(filename)}
\dontrun{head(df)}

}
