% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Create File Name Based on Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year the accident occured}
}
\value{
The function returns a character string
}
\description{
Returns a string which uses year as suffix
}
\examples{
make_filename(2000)
}
