% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Helper function to compose a full filename
according with the \code{fars_data} conventions
by providing only the year number.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A value representing the year. It will
be coerced to integer, therefore it does not need
to be of a specific type, as long as the coercion
is feasible.}
}
\value{
A string with the full filename.
}
\description{
Used by \code{fars_read_years} and \code{fars_map_state},
before \code{fars_read}.
}
\examples{
make_filename(2014)

}
