% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number specifying the year of interest}
}
\value{
this function returns a character string of filename starting with
"accident_" and then the integer year and then ending with ".csv.bz2"
}
\description{
this is a function that compiles a character string of filename
with specified year as input
}
\examples{
#use examples instead of example then no warning of example doesn't exist
\dontrun{make_filename(2013)}

}
