% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer, a numeric or string of the year.}
}
\value{
this function returns a string that is the proper FARS data
   filename for the given year.
}
\description{
This function make csv filenames related to the argument \code{year}.
This function doesn't check the file available or not.
}
\examples{
make_filename(year = '2013')

}
