% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Acquire the filename for a given year data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value of the year you want to check.}
}
\value{
A character that gives correct filename for accident data
}
\description{
Helper function to acquire the filename for a given year data, to be used with other function
}
\examples{
\dontrun{file <- make_filename(2014)
dat <- fars_read(fname)
}
}
\seealso{
\code{\link[=fars_read]{fars_read()}} to read the file
\code{\link[=fars_read_years]{fars_read_years()}} to read several years of data at once
}
