% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer number to be passed to the filename.}
}
\value{
The function returns an object of class "character".
}
\description{
The function creates a standard filename form a year input.
}
\details{
The function attempts to convert its argument to integer.
If unsuccessful, the function throws an error message.
If successful, the function passes the outcome to the sprintf function,
which then prints a standard filename that also contains the year which was specified.
}
\examples{
make_filename(2013)
}
