% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function to create FARS filename with the year parameter.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Valid year as integer or string.}
}
\value{
Filename with the year parameter in the filename as string
}
\description{
This function takes integer or string as year parameter.
It converts year parameter as integer.
It returns output filename as string with the input year.
}
\examples{
make_filename(2012)
make_filename(year=2012)
make_filename("2012")
make_filename(year="2012")

}

