% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or a list of integers containing a year}
}
\value{
a character vector containing a formatted combination of text and variable values
}
\description{
This function combines text with an input value to create a file name for the data from 
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\examples{
\dontrun{
make_filename(list(2013, 2014, 2015))
make_filename(2014:2015)
}
}
\references{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}

