% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate File Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year that is appended to the file name.}
}
\value{
This function returns a file name with the year, as defined by the
user, appended.
}
\description{
This is a simple function that appends a year, as defined by the user, to the
file name "accident_(year).csv.bz2".
}
\examples{

\dontrun{
make_filename(year = 2021)
}

}
