% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the data}
}
\value{
A properly formatted file name, of the format "accident_\code{year}.csv.bz2"
}
\description{
This function takes in year (\code{year}) and returns a properly formatted string pointing to
the file in the local directory.
}
\examples{
make_filename(1991)
make_filename(2001)

\dontrun{
make_filename("ababab")
}

}
