% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Define year for making filename}
}
\value{
Returns a string name
}
\description{
Generate a file name
}
\details{
This function generates a string accident_year.cs.bz2
}
\examples{
\dontrun{make_filename(2013)}
}
\author{
James Bertschy \email{james.bertschy@gmail.ch}
}
