% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{This function just adds the year to the accident's data format}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer}
}
\value{
This function returs a string with the filename with the year especified
}
\description{
This function just adds the year to the accident's data format
}
\examples{
\dontrun{make_filename(2013)}
}
