% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Returns a character string in a format "accident_<year>.csv.bz2" (this can later be used as a file name).}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year to be used for creating the character string.}
}
\value{
A character string.
}
\description{
Returns a character string in a format "accident_<year>.csv.bz2", based on the \code{year} argument.
}
\examples{
make_file(1992)
make_file(year = 1992)

}
