% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{creates a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numerical of integer input indicating a year of a data set}
}
\value{
a character string in a format "accident_<year>.csv.bz2" which can be used as a file name with the given year argument in a form "accident_<year>.csv.bz2"
}
\description{
creates a file name
}
