% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct a file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string referencing a valid year,
four-digit format (i.e. 2013).}
}
\value{
This function returns a file name as a character string.
}
\description{
\code{make_file} takes in a character string referencing a year
and returns a file name as a character string.
}
\examples{
\\code{filename <- make_file('2013')}
\\code{filename}
}
