% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print the filename corresponding to the specified year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or a string corresponding to a specific year for which
the filename should be labeled}
}
\value{
The make_filename function returns the filename as a string.
}
\description{
The make_filename function accepts as input an integer or a string that represents
a specified year.  It then prints the filename corresponding to the year.
}
\details{
The make_filename function results in an error if values other than
years are passed to it or if the data cannot be found.
}
\examples{
\dontrun{
  make_filename(2015)
}

}
