% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer giving the year of a .csv file.}
}
\value{
This function returns a character string.
}
\description{
This function returns a string, which is meant to be passed to the
\code{filename} argument of the \code{fars_read()} function.
}
\details{
This function takes a year as an argument, coerces that year to be an
integer and then uses the \code{sprintf} function to concatenate the \code{year}
argument into "accident_\code{year}.csv.bz2"
}
\examples{
make_filename(2015)

}
