% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer}
}
\value{
character string
}
\description{
Creates a file name based on a year.
}
\examples{
if(interactive()){
   fp <- make_filename(2013)
}

}
