% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Name datasets}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a vector of one or multiple years}
}
\value{
a vector containing one or more fileName values
}
\description{
A function that create the name(s) of the file(s)/datasets that will
be used.
}
\examples{
\dontrun{
make_filename(2019)
make_filename(c(2011, 2012, 2013))
}
}
