% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate FARS Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A 4 digit year, coerced to integer.}
}
\value{
This function returns a FARS filename in character string format.
}
\description{
Generates a standard Fatality Analysis Reporting System (FARS) filename for the year provided.
}
\note{
If the year cannot be coerced to integer, NA is used.
}
\examples{
make_filename('2013')

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}

