% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Create FARS filename from specified year: 2013, 2014 or 2015}
\usage{
make_filename(year = 2013)
}
\arguments{
\item{year}{(integer)}
}
\value{
filename (character)
}
\description{
For a specified year, create the filename to read the (compressed) Fatality Analysis
Reporting System (FARS) data.
This function is not exported.
}
\details{
Also prints the created filename.
}
\examples{
\dontrun{make_filename(2013)}
}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
