% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filenames for Fatality Analysis Reporting System}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year as an integer.}
}
\value{
A filename for the Fatality Analysis Reporting System corresponding
to the specific year in the format "accident_<\code{year}>.csv.bz2".
If type of \code{year} is not correct, an error message is generated and
nothing is returned.
}
\description{
The function \code{make_filename} generates filenames for the US National
Highway Traffic Safety Administration's Fatality Analysis Reporting System
corresponding to a specific year.
}
\examples{
make_filename(2015)

}
