% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year the accident occurs as a character, number or integer}
}
\value{
A character with the filename character.
}
\description{
This function takes year as an input and creates a filename character.
The function only creates the name character and does not create a file.
}
\examples{
\dontrun{make_filename(2000)}
\dontrun{make_filename("2005")}
\dontrun{make_filename(1998L)}

}
