% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Making FARS filenames from given YEAR}
\source{
Coursera, Building R Packages
}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A vector of years, provided as integers (1984)
or strings ("1984", "2001"). Do not provide R 'Date' objects or shortened
suffixes (eg 84) for a year.}
}
\value{
A vector of FARS filenames.
}
\description{
Replacing percent \%d with year (given as input parameter) in string "accident_\%d.csv.bz2"
}
\note{
Year is converted to integer internally.
We don't need to export it, as it's use case is mainly in functions from package
}
\examples{
make_filename(c(2012,2013)) #[1] "accident_2012.csv.bz2" "accident_2013.csv.bz2"
}
\keyword{FARS,}
\keyword{NHTSA,}
\keyword{Road}
\keyword{safety}
\keyword{traffic}
