% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{This is a function that generates datafile names for each year, which can be specified with \code{year} argument. This function is used together with \code{\link{fars_read}} function.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string that can be converted to an integer or an integer.}
}
\value{
A character string that is the file name for the data file for a specific year. The function will halt if the argument is not an integer or a character string that can be converted sucessfully to an integer.
}
\description{
This is a function that generates datafile names for each year, which can be specified with \code{year} argument. This function is used together with \code{\link{fars_read}} function.
}
\examples{
\dontrun{
make_filename(2013)
make_filename(year=2013)
}

}
