% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function \code{make_filename} creates file name based on provided year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Four-digit year, as integer or as a string (i.e. 2001, "2001").}
}
\value{
file name string, such as "accident_YEAR.csv.bz2"
}
\description{
Function \code{make_filename} creates file name based on provided year.
}
\examples{
\dontrun{
make_filename(2013)
make_filename("2013")
}
}
