% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\value{
This function returns a character string with a filename including the year.
}
\description{
This function makes a filename based on a \code{year}.
}
\note{
This functions throws a warning if the input cannot be converted to an integer.
      The year can be a string, but all characters must be digits.
}
\examples{
make_filename(2013)
make_filename("2015")

}
