% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fars_pack.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{-   This is the year - as a 4 digit number. he parameter can be provided as a string, integer or numer since the process
uses /link{as.integer} to coerce the parameter to an integer.}
}
\value{
A character vector containing the filename based on the year provided:
              accident_[Year].csv.bz2
}
\description{
This function creates a filename in a format as created by
the Fatality Analysis Reporting System (FARS) for a specified year.
}
\note{
The use of any non-numeric data will therefore cause the function to fail with a
warning : NAs introduced by coercion
}
\examples{
   make_filename("2019")
   make_filename(2019)
   make_filename(2019.0)

}
\references{
The US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census, provides the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
You can download the FARS data at \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\seealso{
\link{fars_read_years}
 \link{fars_map_state}
}
