% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Returns a string with a file name made from a year like
"accident_1987.csv.bz2"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{input numeric}
}
\value{
character vector
}
\description{
Returns a string with a file name made from a year like
"accident_1987.csv.bz2"
}
\examples{
make_filename(1987)
}
