% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create string with the filename and year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer with the value of the year}
}
\value{
This function returns a string with the filename
}
\description{
This function creates a string composed by the prefix accident_ and year
}
\examples{
\dontrun{
make_filename(2013)
make_filename(2014)
make_filename(2015)
}

}
