% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name for a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric scalar. The year being used for the file name.}
}
\value{
A character string. The created file name.
}
\description{
\code{make_filename} creates a file name in the accident data format for the
specific input year.
}
\examples{
make_filename(2013)
make_filename(2014)
}
