% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year that must be included in the filename}
}
\value{
String of the form "accident_{year}.csv.bz2"
}
\description{
This funciton generates a String of the format "accident_{year}.csv.bz2"
}
\examples{
\dontrun{
fars:::make_filename(2013)
}
}
