% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Name file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number representing the desired year.}
}
\value{
This function returns a character vector with a combination of the
main string and the year inputted.
}
\description{
This function makes a file name.
First, the inputted year will be converted to an integer. The value must be
a simple year without quotes.
Second, the year will be added to the middle of a string that represents the
main pattern of the file name, in the placeholder for an integer.
}
\examples{
\dontrun{
make_filename(2013)
make_filename(2014)
make_filename(2015)
}
}
