% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name for a FARS year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{four digit year either as a number or character string}
}
\value{
a character string matching a FARS filename
}
\description{
.
}
\examples{
make_filename(2013)
make_filename(2014)
}

