% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make the FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year as an integer}
}
\value{
Returns the FARS filename as a string
}
\description{
This function constructs the FARS filename for the datafile of a selected year. The returned string
can be used as an input to \link{fars_read}.
}
\note{
If the FARS system changes their file naming convention this function will need to be updated

Be sure that the datafile resides in the current working directory (use \code{setwd()})
}
\examples{
make_filename(2017)

}
