% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename path}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year (in numbers)}
}
\value{
String
}
\description{
Make a filename path to FARS csv data for a given year.
}
\examples{

make_filename(2013)
}
