% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{From the year, create a valid file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or integer representation of a year.}
}
\value{
A string matching a valid file name for the dataset.

Errors: Anything that can't be coerced by as.integer will
generate a warning, and the same file:
\code{\dontrun{
       Warning message:
       In make_filename("zz") : NAs introduced by coercion
      > zz
      [1] "accident_NA.csv.bz2"
}}
}
\description{
Using \code{\link[dplyr]{tbl_df}} read a file supported by that
function.
}
\examples{
\\code{make_filename(2015)}

}
