% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{FARS make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or integer representation of a year.}
}
\value{
A string matching a valid file name for the dataset.
}
\description{
From the year, create a valid file name to work with FARS dataset.
}
