% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename in a given structure}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or character string that can be converted to integer
 meaning year}
}
\value{
This function returns a character string returns a character vector containing
 a formatted combination of text and variable values in a format:
 accident_%d.csv.bz2".
}
\description{
This is a simple function that creates filename in a appropriate structure:
 "accident_%d.csv.bz2"
}
\examples{
make_filename("2015")
make_filename(2015)
}

