% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Name  accident file year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric string or integer containing the year accident}
}
\value{
This function returns the string accident file name combined with the \code{year}
}
\description{
This is a support function to format the accident file string year
}
\examples{
make_filename(2013)
make_filename("2013")
}
\seealso{
\code{as.integr}
\code{sprintf}
}

