% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year as a number, or a list of numbers,
in form of integer of character}
}
\value{
This function returns a filename based on the given year in .csv format.
}
\description{
This function gets a year and generates filename for that given year
}
\examples{
\dontrun{make_filename("2013")}
\dontrun{make_filename(2013)}
\dontrun{make_filename(c(2013,2014))}

}
