% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Imprime "accident_year.csv.bz2"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{un valor entero (año)}
}
\value{
Esta función regresa un vector de caracteres junto con
  el año ingresado por el usuario de la siquiente manera:"accident_year.csv.bz2"
}
\description{
Esta funcion imprime un vector de caracteres con la variable year ingresada
}
\details{
Es importante ingresar un valor entero o que pueda ser convertido
  a entero mediante coerción, de lo contrario la función regresara un mensaje de advertencia
}
\examples{
make_filename(2013)
make_filename(year= 2013)

}
