% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Produce name of a dataset given an year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer representing an year}
}
\value{
This function returns a string (a file name). An error will be produced
if the input cannot be coerced to an integer.
}
\description{
This function can be used to obtain the name of one of the datasets contained
in this package, providing an year via the \code{year} argument. The according
file name is produced.
}
\examples{
\dontrun{
make_filename(2015)
}

}
