% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Print "make_filename"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A 4-digit year number/string.}
}
\value{
File name for a given year, that satisfies internal naming convention
}
\description{
This function generates file name convention "accident_\%d.csv.bz2"
}
