% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Create filename in R.}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System \url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year that will be included in the string output}
}
\value{
The function will return the following string: "accident_year.csv.bz2".  'year' is the parameter value entered.
}
\description{
This function creates a character string that can be used as the name of a file for other functions in this package.
}
\examples{
make_filename(2013)
}
