% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a FARS Filename from Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric. Will be coerced to an integer with \code{\link[=as.integer]{as.integer()}}}
}
\value{
A string in the format "accident_XXXX.csv.bz2" where XXXX is the year
}
\description{
This function takes as input a year and creates a filename in the format of
data sets from the US National Highway Traffis Safety
Administration's
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Anlaysis Reporting System}
(FARS).
}
\examples{
make_filename(2013)

}
