% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct FARS filename from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The numeric year value for the data for which you need a filename}
}
\value{
This function returns a character string containing the filename of the
   FARS data file for the provided data year.
}
\description{
This function constructs a filename for a FARS CSV file using the data year provided.
FARS is the US National Traffic Safety Commission's
\href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System}.
}
\examples{
make_filename(2014)
make_filename(2013)

}

