% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename for FARS files}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric. A number to format the filename}
}
\value{
A string with a custon filename
}
\description{
Given a year creates a custom filename for FARS files.
}
\examples{
make_filename(2013)
}
