% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The input year that will be converted into a file name.}
}
\value{
This function returns the file name in the format: "accident_YYYY.csv.bz2".
}
\description{
This function takes an input year, converts it to an integer, and then into a
file name in the format: "accident_YYYY.csv.bz2", where YYYY represents the
input year.
}
\note{
Examples: make_filename(1900), make_filename(2000), make_filename(2050)

If a non-year argument is entered as the value for the function parameter
"year", an error will be generated.
try(make_filename(today))
try(make_filename(January))
}
