% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename from year input}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year from which to create the name of the file}
}
\value{
This function returns name of the file containing year.
}
\description{
This function takes year as input and creates a filename using that year.
}
\examples{
make_filename(1980)

}
