% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a Fatality Analysis Reporting System filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{(\emph{character} or (\emph{integer}) ) The year for which the
filename will be created.}
}
\value{
A character containing the filename.
}
\description{
\code{make_filename} creates a filename for a given \code{year} according to
the FARS (Fatality Analysis Reporting System) filenaming scheme.
}
\details{
Both \emph{integer} and \emph{character} are allowed for \code{year}.
\code{year} is coerced to integer, if \code{year} cannot be coerced to integer
a warning is thrown and the returned filename may be invalid. \cr
\code{\link{make_filename}} is a helper function for \code{\link{fars_read}}.
}
\examples{
\dontrun{
flname <- make_filename("2014")
flname <- make_filename(2014)
}

}
\references{
US National Highway Traffic Safety Administration \cr
(\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\seealso{
\code{\link{fars_read}}
}
