% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year as a character string or numeric value.}
}
\value{
Filename as string.
}
\description{
This function makes a filename by adding a year to
default data file name "accident_YYYY.csv.bz2".
}
\examples{
\dontrun{
make_filename(2014)
}

}
