% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate filenames for specific years corresponding to the naming scheme of the FARS}
\usage{
make_filename(year)
}
\arguments{
\item{year}{(numeric) The years for which the respective filenames should be generated}
}
\value{
A character vector of filenames ending in .csv.bz2
}
\description{
This function generates one or more filenames corresponding to the naming scheme of the
US National Highway Traffic Safety Administration's Fatality Analysis
Reporting System.
}
\examples{
make_filename(c(2013, 2014))
}

