% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{FARS create a file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of interest as \code{integer} or \code{character}.}
}
\value{
A \code{character} representing the file name.
}
\description{
Create a file name, using the provided \code{year}, following the FARS pattern
\code{"accident_<year>.csv.bz2"}.
}
\examples{
make_filename(1976)
make_filename("2017")
}
