% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numerical value containing the year that is being searched}
}
\value{
A character string that can be used as a filename. Format: "accident_<year>.csv.bz2"
}
\description{
This function creates a filename for the .csv.bz2 file based on the \code{year} parameter. The parameter must be
an integer or the function will return an error message.
}
\examples{
\dontrun{make_filename(2016)}
}
