% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer}
}
\value{
a string representing a file name
}
\description{
Create a file name for a given year
}
\examples{
make_filename(2014)

}
