% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Represents the year for which the filename needs to be generated;
if year is not a numeric, an NA value will be introduced by coercion thereby
generating an improper filename, "accident_NA.csv.bz2".}
}
\value{
A string is returned in the format "accident_\%d.csv.bz2" but with
the "\%d" replaced with the year parameter supplied by the user.
}
\description{
The \code{make_filename()} function generates a filename string for a specific
year, following the specific format used by the Fatality Analysis Reporting
System (FARS) using base R functions.
}
\examples{
\dontrun{
make_filename(2023)
make_filename(1985)
}

}
