% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print "make_filename"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A int value contained in file name}
}
\value{
This function returns a data frame tbl
}
\description{
This is a simple function that build file names using year
}
\examples{
\dontrun{
make_filename(2020)
}

}
