% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Build an accident data file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Accident data year expressed as a four digit number e.g. 2012.
This parameter may also be any type coercible to a suitable integer,
such as the character string "2012", for example.}
}
\value{
A character string of the form "accident_xxxx.csv.bz2", where
   xxxx is the accident data year expressed as a four digit number.
}
\description{
Given a four digit \code{year}, this function builds an accident data file name
in the "accident_xxxx.csv.bz2" format, where "xxxx" is the supplied \code{year}.
}
\examples{
make_filename(2012)
make_filename("2012")

}
