% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An \code{integer}, or any other type could be coerced to an integer, represents the year of the \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data file.}
}
\value{
A \code{string} for a \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data file for the given year.
}
\description{
This function construct a name for a \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data file give a year.
}
\details{
The function constructs a string for a \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS} data file for the given year. It addes the prefix "accident_" to the given year then adds ".csv.bz2" suffix to the string.
}
\examples{
\dontrun{
# input year as an intger
make_filename(2013)

# input year as a character
make_filename("2014")
}

}
