% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes a filename by concatenating "accident_\%d.csv.bz2" and a user-supplied year, where \%d becomes the year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer representing the year that will be inserted into the file name}
}
\value{
returns the complete file name
}
\description{
Makes a filename by concatenating "accident_\%d.csv.bz2" and a user-supplied year, where \%d becomes the year.
}
\note{
If \code{year} cannot be coerced into an integer, the function will throw an error
}
\examples{
make_filename(2013)
\dontrun{make_filename(not_a_number)}

}
