% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year (integer/numeric) in the format yyyy}
}
\value{
A file name (character) in the format accident_<yyyy>.csv.bz2
where yyyy is the passed value in the argument year.
}
\description{
This function takes in an argument for a year in the format yyyy and
generates a file name in the format accident_<yyyy>.csv.bz2.
}
\examples{
\dontrun{
make_filename(2013)
make_filename(2014)
}

}
