% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An atomic character or numeric vector representing the year to
augment to the file name.}
}
\value{
An atomic character vector representing the file name.

\dontrun{
    make_filename("2/2/2015")
 }
}
\description{
\code{make_filename} completes a file name with an input year.
}
\details{
The function uses the argument given to the function to augment to the
predefined template, and outputs a filename.
}
\section{Required functions to import from other packages}{

None
}

