% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename for the chosen year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer representing the year}
}
\value{
Returns a string with the filename for the given year
}
\description{
This simple function creates a filename based on the given year.Input year is an integer.
}
\examples{
\dontrun{make_filename(2015)}

}
