% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make individual names of each data file for a particular year.}
\usage{
make_filename(year)
}
\arguments{
\item{Input}{class 'integer' value of a 'year' as the one argument.}
}
\value{
This function simply returns the names of the data files
year wise.
}
\description{
This function creates individual file names for each data frame
with .csv format corresponding to a year.
}
\examples{
\dontrun{
make_filename(2013)
make_filename(2014)
}

}
