% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{This function creates the name for the data file to be downloaded. The function takes the input of the year and returns the filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{This is the year of the data to be included in the file name.
This can be inputted as an integer, numeric, or characterised number - "2013".
An error will occur if an invalid year is entered.}
}
\value{
This function returns the full file name of the file "accident" and the for the year specified.
}
\description{
This function creates the name for the data file to be downloaded. The function takes the input of the year and returns the filename.
}
\examples{
\dontrun{
make_filename(2013)
make_filename("2014")
}
}
