% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates the filename of the fars-data for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year as an integer value. If \code{year} is not in this form, it will be coerced to integer.}
}
\value{
The function returns the filename of the data from the FARS system of the corresponding \code{year} .
}
\description{
Creates the filename of the fars-data for a given year
}
\examples{
make_filename(2015)
}

