% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename(): Generates filename per given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An year of interest to be made for the accident filename (2013-2015)}
}
\value{
Filename will be returned
}
\description{
This function accepts a specific year and generates a filename based on that parameter.
This result will be read into fars_read()
}
\examples{
make_filename(2013)
}
