% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer number to denote the year.}
}
\value{
Returns a character string in the form of "accident_<year>.csv.bz2"
}
\description{
This is a simple function that prints "accident_<year>.csv.bz2",
defined by the \code{year} argument.
}
\examples{
\dontrun{
make_filename(2013)
make_filename(2014)
make_filename(2015)
}

}
