% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a accident file name with extensions}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or string giving the year for the filename}
}
\value{
This function returns a character vector with a year preceded by "accident_"
and followed by ".csv.bz2"
}
\description{
This is a function that builds a filename in the standard file format for
"Highway Fatality Report", given a year.
}
\details{
If the provided \code{year} is not a integer, the value is truncated. If the provided \code{year} can't be converted to a numeric, a warning about N/A's is printed.
}
\examples{
make_filename(1984)
make_filename("1984")
}
