% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create standardised file names}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the FARS report in any format that can be coerced to an integer}
}
\value{
This function returns a character string to be used as file name. The string has
        the following structure where YYYY is replaced with what has been passed as the
        \code{year} argument: \code{accident_YYYY.csv.bz2}
}
\description{
The \code{make_filename()} function generates standardised file names for FARS
data sets based on the year you pass to it as argument.
}
\examples{
make_filename(2013)

}
