% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer value indicating a year of a data set.}
}
\value{
string in a format "accident_<year>.csv.bz2"
}
\description{
This function creates a filename for a .csv.bz2 file based on the \code{year}
argument in a form "accident_<year>.csv.bz2". Function will end with an error
if input is not a an integer value.
}
\examples{
\dontrun{
makefilename(2017)
}
}
