% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{character string indicating year}
}
\value{
This function returns a comparessed .csv file
}
\description{
This function create a corresponding filename according to input year.
}
\examples{
make_filename(2016)
}
