% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_W4.R
\name{make_filename}
\alias{make_filename}
\title{Make file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year to be attached to the file name.}
}
\value{
This function returns a file name with the input year.
}
\description{
This function creates an "accident_year.csv.bz2" file name given the year entered by the user.
}
\examples{
fname <- make_filename(2015)

}
