% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Create FARS filename for a year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for the FARS data}
}
\value{
Filename of the FARS dataset for the corresponding year
}
\description{
Create FARS filename for a year
}
\details{

}
\examples{
make_filename(2013)
}
