% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Get the name of the file containing FARS data for a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An atomic numeric value, which can be coerced to four-digit integer
representing a year to process.}
}
\value{
It returns the name of the file containing the FARS data of the year.
}
\description{
The function returns the name of the data file
containing FARS data for a year of interest (defined using \code{year} argument).
For more information about FARS data, see
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}.
}
\examples{
FileName2013 <- make_filename(2013)

}
