% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create complete file name for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or numeric specifing the year.}
}
\value{
This function returns a character string containing the complete
   filename for the given year
}
\description{
A helper (not exported) function used by \code{fars_read_years} and \code{fars_map_state}
functions. It takes a year as argument and returns the complete file name
containing data for the given year.
}
\examples{
\dontrun{
make_filename(2013)
make_filename("2013")
}

}
