% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print filename for a provided year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer}
}
\value{
This function returns a character vector containing a combination of text and provided year number.
}
\description{
This function prints a character vector consisting of a filename format combined with the passed year argument.
}
\examples{
\dontrun{
make_filename(2017)
make_filename(2018)
}

}
