% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.r
\name{make_filename}
\alias{make_filename}
\title{Create standardized filename}
\usage{
make_filename(year, demo = FALSE)
}
\arguments{
\item{year}{a character/numeric value reporting the year of file
that you want to read}

\item{demo}{logical if you want to use the function with demo data that comes with the package.
Default to FALSE}
}
\value{
a character string with the name for zipped csv files
}
\description{
This function create a standard naming for a zipped csv file
}
\examples{
make_filename("2010")
make_filename(2010)

}
