% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{is an integer representing year}
}
\value{
string representing a filename
}
\description{
Make filename
}
\examples{
\dontrun{ make_filename(2013) }
}
