% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create Fatality Analysis Reporting System filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric or an integer or a number string with four digits.
This specifies year the data was recorded.}
}
\value{
This function returns a length 1 character string specifies the name
of the data from the input year. The name is of the form accident_year.csv.bz2
}
\description{
Create Fatality Analysis Reporting System filename
}
\examples{
\dontrun{make_filename(2014)}
}
\seealso{
\code{sprintf}
}
