% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Helper Function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer giving the year for which the file name is to be created.}
}
\value{
This function returns a character string of the file name corresponding to
the data for the given year.
}
\description{
This is a helper function used by other functions to read data.
}
\note{
Not exported to the user.
}
