% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_filename}
\alias{make_filename}
\title{Define 5 functions for export : need to add export before each function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a single year}

\item{state.num}{Number representing a US state.}

\item{years}{a list of years}

\item{filename}{is the location of the data file}
}
\description{
\code{fars_read(filename)} read csv file with name filename
\code{make_filename(year)} make csv file of the year
\code{fars_read_years(years)} make csv file of years
\code{fars_summarize_years(years)} summarise data of years.
\code{fars_map_state(state.num, year)} plot map of the state.num
}
\examples{
\dontrun{
fars_map_state(18, 2015)
}

}
