% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{\code{make_filename} takes in a specific 'year', creates and returns a filename specifying the year of the dataset. The filename includes the
the \code{bz2} extension for importing zipped files. For example, if the year specified is \code{2011}, the filename
will be \code{accident_2011.csv.bz2}.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or integer input specifying the year of the dataset.}
}
\value{
A character string which is the filename.
}
\description{
\code{make_filename} takes in a specific 'year', creates and returns a filename specifying the year of the dataset. The filename includes the
the \code{bz2} extension for importing zipped files. For example, if the year specified is \code{2011}, the filename
will be \code{accident_2011.csv.bz2}.
}
\examples{
\dontrun{
make_filename(2011)
}
}
