% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename associated with given year' data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer specifing year of interest}
}
\value{
This function returns a character string specifying file name containing
data of interest.
}
\description{
This function creates a name of a file containg accident data for a year
of interest specified in the \code{year} argument. \code{year} must be an integer
or be possible to be coerced to an integer, otherwise an error will be generated.
}
\examples{
\dontrun{make_filename(2001)}

}
