% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{create file name by year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year, a numeric value}
}
\value{
A character string contains year information as file name
}
\description{
This is a function to create a file name for accident by using year,
which can be customized by users.
}
