% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make data file name with year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer or string with input year}
}
\value{
A string of file name for given year
}
\description{
From the given year (four digit number), make file name of csv data
}
\examples{
\dontrun{make_filename(2013)}


}
\seealso{
\link{fars_read}
}
