% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate a file name in a specific convension}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer of 4 digits representing a year.}
}
\value{
A string representing the coresponding file name.
}
\description{
This funtion serves as a template to generate a specific file name,
in the format of "accident_yyyy.csv.bz2", based on the input value (yyyy).
}
\examples{
\dontrun{
make_filename(2013)
}
}
