% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename Functions}
\usage{
make_filename(year)
}
\arguments{
\item{A}{character, numeric or integer with the year you are interested in}
}
\value{
A character vector with the name of the file you want to access.
}
\description{
This is adunction that creates the filename from which you will access the data with the help of the fars read function
}
