% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename for FARS annual data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numeric vector that specifies the year for which the filename will be generated}
}
\value{
This function returns a character string
}
\description{
This is a function that, generates specific filename
needed to access the file of anual FARS data specified by \code{year} argument.
Filemane generated by this function could be used by \code{fars_read} function.
}
\details{
If used inside fars_read single number agrument recommended
}
\examples{
\dontrun{
make_filename(2014)
fars_read(make_filename(2014))
}

}
