% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_fars_fun.r
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of data to be read.}
}
\value{
make_filename returns string filename of data set to import.
}
\description{
This function takes year as an argument, converts to integer, and returns a filename
with that can be read as "accident_",year,".csv.bz2")
}
