% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fars}
\alias{plot.fars}
\title{Plot Method for \code{fars} Object}
\usage{
\method{plot}{fars}(x, dates = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fars}.}

\item{dates}{Optional vector of dates (as \code{Date} or \code{zoo::yearqtr}) to use for the x-axis.
If not provided, a simple index is used.}

\item{...}{Additional arguments (ignored).}
}
\value{
No return value. Called for plot generation.
}
\description{
Generates line plots of estimated quantiles from a \code{fars} object.
If a stressed scenario is available, it is plotted in a separate panel.
}
