% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fars_density}
\alias{plot.fars_density}
\title{Plot Method for \code{fars_density} Object}
\usage{
\method{plot}{fars_density}(x, time_index = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fars_density}.}

\item{time_index}{Optional vector for the time axis (default is 1:nrow).}

\item{...}{Additional arguments (ignored).}
}
\description{
Plots the evolution of the estimated density over time as a 3D surface.
}
