% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.mldfm.R
\name{plot_factors.mldfm}
\alias{plot_factors.mldfm}
\title{Plot Extracted Factors from MLDFM}
\usage{
plot_factors.mldfm(x, dates = NULL, ...)
}
\description{
Plot Extracted Factors from MLDFM
}
\keyword{internal}
