% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_factors.mldfm}
\alias{plot_factors.mldfm}
\title{Plot Factors from \code{mldfm} Object}
\usage{
plot_factors.mldfm(x, dates = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{dates}{Optional vector of dates. If NULL, uses 1:n as default.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays time series plots of the estimated factors with 95\% confidence bands.
}
\keyword{internal}
