% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_residuals.mldfm}
\alias{plot_residuals.mldfm}
\title{Plot Residuals from \code{mldfm} Object}
\usage{
plot_residuals.mldfm(x, var_names = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{var_names}{Optional vector of variable names. If NULL, default names are used.}

\item{...}{Additional arguments (ignored).}
}
\description{
Displays a correlation heatmap of the residuals.
}
\keyword{internal}
