% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_mldfm.R
\name{plot_residuals.mldfm}
\alias{plot_residuals.mldfm}
\title{Plot Residuals from MLDFM}
\usage{
plot_residuals.mldfm(x, var_names = NULL, ...)
}
\description{
Plot Residuals from MLDFM
}
\keyword{internal}
