% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fars.R
\name{print.fars}
\alias{print.fars}
\title{Print method for fars object}
\usage{
\method{print}{fars}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fars_quantiles}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Prints a short summary of the fars object
}
