% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fars}
\alias{print.fars}
\title{Print Method for \code{fars} Object}
\usage{
\method{print}{fars}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fars}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{fars} object, returned invisibly.
}
\description{
Prints a short summary of the fars object.
}
