% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fars_density}
\alias{print.fars_density}
\title{Print Method for \code{fars_density} Object}
\usage{
\method{print}{fars_density}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fars_density}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{fars_density} object, invisibly.
}
\description{
Displays a brief summary of the \code{fars_density} object.
}
