% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mldfm.R
\name{print.mldfm}
\alias{print.mldfm}
\title{Print Method for MLDFM Object}
\usage{
\method{print}{mldfm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mldfm}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Prints a short summary of the multilevel dynamic factor model
}
