% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_reg.R
\name{q_reg}
\alias{q_reg}
\title{Quantile Regression with Stressed Scenario Projection}
\usage{
q_reg(dep_variable, factors, Stressed_Factors = NULL, h = 1, QTAU = 0.05)
}
\description{
Estimates quantile regressions of a dependent variable on dynamic factors.
}
\keyword{internal}
