% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fars.R
\name{summary.fars}
\alias{summary.fars}
\title{Summary Method for fars Object}
\usage{
\method{summary}{fars}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fars_quantiles}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object \code{object}, returned invisibly.
}
\description{
Prints a complete summary of the fars object.
}
