% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mldfm}
\alias{summary.mldfm}
\title{Summary Method for \code{mldfm} Object}
\usage{
\method{summary}{mldfm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mldfm}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{mldfm} object, invisibly.
}
\description{
Provides a complete summary of the MLDFM object.
}
