% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mldfm.R
\name{summary.mldfm}
\alias{summary.mldfm}
\title{Summary Method for MLDFM Object}
\usage{
\method{summary}{mldfm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mldfm}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Provides a compact summary of a fitted multilevel dynamic factor model.
}
