% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_factor_list.R
\name{update_factor_list}
\alias{update_factor_list}
\title{Update Factor List from a given Factor Matrix}
\usage{
update_factor_list(Factor_list, FinalFactors, r_list)
}
\description{
Update Factor List from a given Factor Matrix
}
\keyword{internal}
